feat(WMGV2): Improved method of adding genes to heatmap
