--- frontend/src/views/WheresMyGene/components/GeneSearchBar/components/SaveExport/ExportBanner/index.tsx
@@ -1,41 +1,0 @@
-import React from "react";
-
-// Default dimensions
-export const DATA_MESSAGE_BANNER_WIDTH_PX = 433;
-export const DATA_MESSAGE_BANNER_HEIGHT_PX = 64;
-
-export const DATA_MESSAGE_BANNER_ID = "data-message-banner";
-
-interface Props {
-  height?: number;
-  width?: number;
-  centered?: boolean;
-}
-
-export const UnderlyingDataChangeBanner = React.memo(
-  function UnderlyingDataChangeMessage({
-    height = DATA_MESSAGE_BANNER_HEIGHT_PX,
-    width = DATA_MESSAGE_BANNER_WIDTH_PX,
-    centered = true,
-  }: Props): JSX.Element {
-    return (
-      <svg
-        id={DATA_MESSAGE_BANNER_ID}
-        fill="none"
-        width={width}
-        height={height}
-        xmlns="http://www.w3.org/2000/svg"
-        viewBox={`0 0 ${width} ${height}`}
-      >
-        <rect width="100%" height="100%" rx="4" fill="#F8F8F8" />
-        <path fill="#767676" d={centered ? CENTERED_TEXT : LEFT_ALIGNED_TEXT} />
-      </svg>
-    );
-  }
-);
-
-const CENTERED_TEXT =
-  "M15.2903 27L12.7051 17.5455H13.8684L15.8443 25.2457H15.9366L17.9494 17.5455H19.242L21.2548 25.2457H21.3471L23.323 17.5455H24.4863L21.9011 27H20.7193L18.6326 19.4659H18.5588L16.4721 27H15.2903ZM28.0389 27.1477C27.3557 27.1477 26.7663 26.9969 26.2708 26.6953C25.7784 26.3906 25.3983 25.9659 25.1306 25.4212C24.8659 24.8733 24.7335 24.2363 24.7335 23.5099C24.7335 22.7836 24.8659 22.1435 25.1306 21.5895C25.3983 21.0324 25.7707 20.5985 26.2478 20.2876C26.7279 19.9737 27.288 19.8168 27.9281 19.8168C28.2975 19.8168 28.6622 19.8783 29.0223 20.0014C29.3823 20.1245 29.7101 20.3246 30.0056 20.6016C30.301 20.8755 30.5365 21.2386 30.7119 21.6911C30.8873 22.1435 30.975 22.7005 30.975 23.3622V23.8239H25.5091V22.8821H29.8671C29.8671 22.482 29.7871 22.125 29.627 21.8111C29.4701 21.4972 29.2454 21.2494 28.953 21.0678C28.6637 20.8862 28.3221 20.7955 27.9281 20.7955C27.4942 20.7955 27.1187 20.9032 26.8017 21.1186C26.4878 21.331 26.2462 21.608 26.0769 21.9496C25.9077 22.2912 25.823 22.6574 25.823 23.0483V23.6761C25.823 24.2116 25.9154 24.6656 26.1 25.038C26.2878 25.4073 26.5478 25.6889 26.8802 25.8828C27.2126 26.0736 27.5988 26.169 28.0389 26.169C28.3252 26.169 28.5837 26.129 28.8145 26.049C29.0484 25.9659 29.25 25.8428 29.4193 25.6797C29.5885 25.5135 29.7193 25.3073 29.8117 25.0611L30.8642 25.3565C30.7534 25.7135 30.5672 26.0275 30.3056 26.2983C30.044 26.5661 29.7209 26.7753 29.3362 26.9261C28.9515 27.0739 28.5191 27.1477 28.0389 27.1477ZM36.2106 27V19.9091H37.2631V20.9801H37.337C37.4663 20.6293 37.7002 20.3446 38.0387 20.1261C38.3773 19.9076 38.7589 19.7983 39.1836 19.7983C39.2636 19.7983 39.3636 19.7998 39.4837 19.8029C39.6037 19.806 39.6945 19.8106 39.756 19.8168V20.9247C39.7191 20.9155 39.6345 20.9016 39.5021 20.8832C39.3729 20.8616 39.2359 20.8509 39.0913 20.8509C38.7466 20.8509 38.4388 20.9232 38.168 21.0678C37.9002 21.2094 37.6879 21.4064 37.5309 21.6587C37.377 21.908 37.3001 22.1927 37.3001 22.5128V27H36.2106ZM43.7657 27.1477C43.0825 27.1477 42.4931 26.9969 41.9976 26.6953C41.5052 26.3906 41.1251 25.9659 40.8573 25.4212C40.5926 24.8733 40.4603 24.2363 40.4603 23.5099C40.4603 22.7836 40.5926 22.1435 40.8573 21.5895C41.1251 21.0324 41.4975 20.5985 41.9745 20.2876C42.4546 19.9737 43.0147 19.8168 43.6549 19.8168C44.0242 19.8168 44.3889 19.8783 44.749 20.0014C45.1091 20.1245 45.4369 20.3246 45.7323 20.6016C46.0278 20.8755 46.2632 21.2386 46.4386 21.6911C46.6141 22.1435 46.7018 22.7005 46.7018 23.3622V23.8239H41.2359V22.8821H45.5938C45.5938 22.482 45.5138 22.125 45.3538 21.8111C45.1968 21.4972 44.9721 21.2494 44.6798 21.0678C44.3905 20.8862 44.0488 20.7955 43.6549 20.7955C43.2209 20.7955 42.8455 20.9032 42.5285 21.1186C42.2146 21.331 41.973 21.608 41.8037 21.9496C41.6344 22.2912 41.5498 22.6574 41.5498 23.0483V23.6761C41.5498 24.2116 41.6421 24.6656 41.8268 25.038C42.0145 25.4073 42.2746 25.6889 42.607 25.8828C42.9393 26.0736 43.3256 26.169 43.7657 26.169C44.0519 26.169 44.3104 26.129 44.5413 26.049C44.7752 25.9659 44.9767 25.8428 45.146 25.6797C45.3153 25.5135 45.4461 25.3073 45.5384 25.0611L46.591 25.3565C46.4802 25.7135 46.294 26.0275 46.0324 26.2983C45.7708 26.5661 45.4476 26.7753 45.0629 26.9261C44.6782 27.0739 44.2458 27.1477 43.7657 27.1477ZM51.1823 29.8068C50.656 29.8068 50.2036 29.7391 49.8251 29.6037C49.4465 29.4714 49.131 29.2959 48.8787 29.0774C48.6294 28.862 48.4309 28.6312 48.2832 28.3849L49.1511 27.7756C49.2495 27.9048 49.3742 28.0526 49.525 28.2188C49.6758 28.388 49.882 28.5342 50.1436 28.6573C50.4083 28.7835 50.7545 28.8466 51.1823 28.8466C51.7547 28.8466 52.2272 28.7081 52.5996 28.4311C52.972 28.1541 53.1582 27.7202 53.1582 27.1293V25.6889H53.0658C52.9858 25.8182 52.8719 25.9782 52.7242 26.169C52.5796 26.3568 52.3703 26.5245 52.0964 26.6722C51.8255 26.8169 51.4593 26.8892 50.9976 26.8892C50.4252 26.8892 49.9112 26.7538 49.4557 26.483C49.0033 26.2121 48.6448 25.8182 48.3801 25.3011C48.1185 24.7841 47.9877 24.1562 47.9877 23.4176C47.9877 22.6913 48.1154 22.0588 48.3709 21.5202C48.6263 20.9786 48.9818 20.56 49.4373 20.2646C49.8928 19.966 50.419 19.8168 51.0161 19.8168C51.4778 19.8168 51.844 19.8937 52.1148 20.0476C52.3887 20.1984 52.598 20.3707 52.7427 20.5646C52.8904 20.7554 53.0043 20.9124 53.0843 21.0355H53.1951V19.9091H54.2476V27.2031C54.2476 27.8125 54.1092 28.308 53.8322 28.6896C53.5583 29.0743 53.1889 29.3559 52.7242 29.5344C52.2626 29.716 51.7486 29.8068 51.1823 29.8068ZM51.1454 25.9105C51.5824 25.9105 51.9517 25.8105 52.2533 25.6104C52.5549 25.4104 52.7842 25.1226 52.9412 24.7472C53.0981 24.3717 53.1766 23.9223 53.1766 23.3991C53.1766 22.8883 53.0997 22.4374 52.9458 22.0465C52.7919 21.6557 52.5642 21.3494 52.2626 21.1278C51.961 20.9062 51.5886 20.7955 51.1454 20.7955C50.6837 20.7955 50.299 20.9124 49.9913 21.1463C49.6866 21.3802 49.4573 21.6941 49.3034 22.0881C49.1526 22.482 49.0772 22.919 49.0772 23.3991C49.0772 23.8916 49.1541 24.3271 49.308 24.7056C49.465 25.0811 49.6958 25.3765 50.0005 25.592C50.3083 25.8043 50.6899 25.9105 51.1454 25.9105ZM60.6717 24.1009V19.9091H61.7612V27H60.6717V25.7997H60.5978C60.4317 26.1598 60.1731 26.466 59.8223 26.7184C59.4714 26.9677 59.0282 27.0923 58.4927 27.0923C58.0496 27.0923 57.6556 26.9954 57.3109 26.8015C56.9662 26.6045 56.6954 26.3091 56.4984 25.9151C56.3014 25.5181 56.203 25.018 56.203 24.4148V19.9091H57.2925V24.3409C57.2925 24.858 57.4371 25.2704 57.7264 25.5781C58.0188 25.8859 58.3912 26.0398 58.8436 26.0398C59.1144 26.0398 59.3899 25.9705 59.6699 25.832C59.9531 25.6935 60.1901 25.4812 60.3809 25.195C60.5748 24.9087 60.6717 24.544 60.6717 24.1009ZM64.8072 17.5455V27H63.7177V17.5455H64.8072ZM68.8503 27.1662C68.4009 27.1662 67.9932 27.0816 67.6269 26.9123C67.2607 26.7399 66.9698 26.4922 66.7544 26.169C66.539 25.8428 66.4312 25.4489 66.4312 24.9872C66.4312 24.581 66.5113 24.2517 66.6713 23.9993C66.8313 23.7438 67.0452 23.5438 67.313 23.3991C67.5808 23.2545 67.8762 23.1468 68.1994 23.076C68.5256 23.0021 68.8534 22.9437 69.1827 22.9006C69.6135 22.8452 69.9629 22.8036 70.2306 22.7759C70.5014 22.7451 70.6984 22.6944 70.8215 22.6236C70.9477 22.5528 71.0108 22.4297 71.0108 22.2543V22.2173C71.0108 21.7618 70.8861 21.4079 70.6369 21.1555C70.3906 20.9032 70.0167 20.777 69.5151 20.777C68.9949 20.777 68.5871 20.8909 68.2917 21.1186C67.9962 21.3464 67.7885 21.5895 67.6685 21.848L66.6344 21.4787C66.819 21.0478 67.0652 20.7124 67.373 20.4723C67.6839 20.2292 68.0224 20.0599 68.3886 19.9645C68.758 19.866 69.1211 19.8168 69.4781 19.8168C69.7059 19.8168 69.9675 19.8445 70.2629 19.8999C70.5615 19.9522 70.8492 20.0614 71.1262 20.2276C71.4063 20.3938 71.6386 20.6446 71.8233 20.9801C72.008 21.3156 72.1003 21.7649 72.1003 22.3281V27H71.0108V26.0398H70.9554C70.8815 26.1937 70.7584 26.3583 70.5861 26.5337C70.4137 26.7092 70.1844 26.8584 69.8982 26.9815C69.612 27.1046 69.2627 27.1662 68.8503 27.1662ZM69.0165 26.1875C69.4473 26.1875 69.8105 26.1029 70.106 25.9336C70.4045 25.7643 70.6292 25.5458 70.78 25.2781C70.9339 25.0103 71.0108 24.7287 71.0108 24.4332V23.4361C70.9646 23.4915 70.8631 23.5423 70.7061 23.5884C70.5522 23.6315 70.3737 23.67 70.1706 23.7038C69.9705 23.7346 69.7751 23.7623 69.5843 23.7869C69.3966 23.8085 69.2442 23.8269 69.1273 23.8423C68.8441 23.8793 68.5794 23.9393 68.3332 24.0224C68.0901 24.1024 67.8931 24.224 67.7423 24.3871C67.5946 24.5471 67.5207 24.7656 67.5207 25.0426C67.5207 25.4212 67.6608 25.7074 67.9408 25.9013C68.224 26.0921 68.5825 26.1875 69.0165 26.1875ZM74.0498 27V19.9091H75.1024V20.9801H75.1762C75.3055 20.6293 75.5394 20.3446 75.878 20.1261C76.2165 19.9076 76.5981 19.7983 77.0228 19.7983C77.1029 19.7983 77.2029 19.7998 77.3229 19.8029C77.4429 19.806 77.5337 19.8106 77.5953 19.8168V20.9247C77.5584 20.9155 77.4737 20.9016 77.3414 20.8832C77.2121 20.8616 77.0752 20.8509 76.9305 20.8509C76.5858 20.8509 76.278 20.9232 76.0072 21.0678C75.7395 21.2094 75.5271 21.4064 75.3701 21.6587C75.2163 21.908 75.1393 22.1927 75.1393 22.5128V27H74.0498ZM79.9372 17.5455V27H78.8477V17.5455H79.9372ZM82.5954 29.6591C82.4107 29.6591 82.2461 29.6437 82.1014 29.6129C81.9568 29.5852 81.8568 29.5575 81.8014 29.5298L82.0784 28.5696C82.343 28.6373 82.5769 28.6619 82.7801 28.6435C82.9832 28.625 83.1632 28.5342 83.3202 28.3711C83.4802 28.2111 83.6264 27.951 83.7588 27.5909L83.9619 27.0369L81.3397 19.9091H82.5215L84.4789 25.5597H84.5528L86.5102 19.9091H87.692L84.6821 28.0341C84.5466 28.4003 84.3789 28.7035 84.1789 28.9435C83.9788 29.1867 83.7465 29.3667 83.4818 29.4837C83.2202 29.6006 82.9247 29.6591 82.5954 29.6591ZM95.6813 27.1477C94.9981 27.1477 94.4087 26.9969 93.9132 26.6953C93.4208 26.3906 93.0407 25.9659 92.7729 25.4212C92.5082 24.8733 92.3759 24.2363 92.3759 23.5099C92.3759 22.7836 92.5082 22.1435 92.7729 21.5895C93.0407 21.0324 93.4131 20.5985 93.8901 20.2876C94.3702 19.9737 94.9303 19.8168 95.5705 19.8168C95.9398 19.8168 96.3045 19.8783 96.6646 20.0014C97.0247 20.1245 97.3525 20.3246 97.6479 20.6016C97.9434 20.8755 98.1788 21.2386 98.3542 21.6911C98.5297 22.1435 98.6174 22.7005 98.6174 23.3622V23.8239H93.1515V22.8821H97.5094C97.5094 22.482 97.4294 22.125 97.2694 21.8111C97.1124 21.4972 96.8877 21.2494 96.5954 21.0678C96.3061 20.8862 95.9644 20.7955 95.5705 20.7955C95.1365 20.7955 94.7611 20.9032 94.4441 21.1186C94.1302 21.331 93.8886 21.608 93.7193 21.9496C93.55 22.2912 93.4654 22.6574 93.4654 23.0483V23.6761C93.4654 24.2116 93.5577 24.6656 93.7424 25.038C93.9301 25.4073 94.1902 25.6889 94.5226 25.8828C94.8549 26.0736 95.2412 26.169 95.6813 26.169C95.9675 26.169 96.226 26.129 96.4569 26.049C96.6908 25.9659 96.8923 25.8428 97.0616 25.6797C97.2309 25.5135 97.3617 25.3073 97.454 25.0611L98.5066 25.3565C98.3958 25.7135 98.2096 26.0275 97.948 26.2983C97.6864 26.5661 97.3632 26.7753 96.9785 26.9261C96.5938 27.0739 96.1614 27.1477 95.6813 27.1477ZM100.858 19.9091L102.557 22.8082L104.255 19.9091H105.511L103.221 23.4545L105.511 27H104.255L102.557 24.2486L100.858 27H99.6021L101.855 23.4545L99.6021 19.9091H100.858ZM107.027 29.6591V19.9091H108.079V21.0355H108.209C108.289 20.9124 108.399 20.7554 108.541 20.5646C108.686 20.3707 108.892 20.1984 109.16 20.0476C109.43 19.8937 109.797 19.8168 110.258 19.8168C110.855 19.8168 111.382 19.966 111.837 20.2646C112.293 20.5631 112.648 20.9863 112.904 21.5341C113.159 22.0819 113.287 22.7282 113.287 23.473C113.287 24.224 113.159 24.8749 112.904 25.4258C112.648 25.9736 112.294 26.3983 111.842 26.6999C111.389 26.9985 110.868 27.1477 110.277 27.1477C109.821 27.1477 109.457 27.0723 109.183 26.9215C108.909 26.7676 108.698 26.5937 108.55 26.3999C108.402 26.2029 108.289 26.0398 108.209 25.9105H108.116V29.6591H107.027ZM108.098 23.4545C108.098 23.9901 108.176 24.4625 108.333 24.8718C108.49 25.2781 108.719 25.5966 109.021 25.8274C109.323 26.0552 109.692 26.169 110.129 26.169C110.585 26.169 110.965 26.049 111.269 25.8089C111.577 25.5658 111.808 25.2396 111.962 24.8303C112.119 24.4179 112.197 23.9593 112.197 23.4545C112.197 22.956 112.12 22.5066 111.966 22.1065C111.816 21.7034 111.586 21.3848 111.279 21.1509C110.974 20.9139 110.591 20.7955 110.129 20.7955C109.686 20.7955 109.313 20.9078 109.012 21.1325C108.71 21.354 108.482 21.6649 108.329 22.065C108.175 22.462 108.098 22.9252 108.098 23.4545ZM116.996 27.1662C116.547 27.1662 116.139 27.0816 115.773 26.9123C115.407 26.7399 115.116 26.4922 114.9 26.169C114.685 25.8428 114.577 25.4489 114.577 24.9872C114.577 24.581 114.657 24.2517 114.817 23.9993C114.977 23.7438 115.191 23.5438 115.459 23.3991C115.727 23.2545 116.022 23.1468 116.345 23.076C116.672 23.0021 116.999 22.9437 117.329 22.9006C117.76 22.8452 118.109 22.8036 118.377 22.7759C118.647 22.7451 118.844 22.6944 118.968 22.6236C119.094 22.5528 119.157 22.4297 119.157 22.2543V22.2173C119.157 21.7618 119.032 21.4079 118.783 21.1555C118.537 20.9032 118.163 20.777 117.661 20.777C117.141 20.777 116.733 20.8909 116.438 21.1186C116.142 21.3464 115.935 21.5895 115.814 21.848L114.78 21.4787C114.965 21.0478 115.211 20.7124 115.519 20.4723C115.83 20.2292 116.168 20.0599 116.535 19.9645C116.904 19.866 117.267 19.8168 117.624 19.8168C117.852 19.8168 118.113 19.8445 118.409 19.8999C118.707 19.9522 118.995 20.0614 119.272 20.2276C119.552 20.3938 119.785 20.6446 119.969 20.9801C120.154 21.3156 120.246 21.7649 120.246 22.3281V27H119.157V26.0398H119.101C119.028 26.1937 118.904 26.3583 118.732 26.5337C118.56 26.7092 118.33 26.8584 118.044 26.9815C117.758 27.1046 117.409 27.1662 116.996 27.1662ZM117.162 26.1875C117.593 26.1875 117.957 26.1029 118.252 25.9336C118.551 25.7643 118.775 25.5458 118.926 25.2781C119.08 25.0103 119.157 24.7287 119.157 24.4332V23.4361C119.111 23.4915 119.009 23.5423 118.852 23.5884C118.698 23.6315 118.52 23.67 118.317 23.7038C118.117 23.7346 117.921 23.7623 117.73 23.7869C117.543 23.8085 117.39 23.8269 117.273 23.8423C116.99 23.8793 116.725 23.9393 116.479 24.0224C116.236 24.1024 116.039 24.224 115.888 24.3871C115.741 24.5471 115.667 24.7656 115.667 25.0426C115.667 25.4212 115.807 25.7074 116.087 25.9013C116.37 26.0921 116.729 26.1875 117.162 26.1875ZM123.285 22.7344V27H122.196V19.9091H123.248V21.017H123.341C123.507 20.657 123.759 20.3677 124.098 20.1491C124.436 19.9276 124.873 19.8168 125.409 19.8168C125.889 19.8168 126.309 19.9152 126.669 20.1122C127.029 20.3061 127.309 20.6016 127.509 20.9986C127.709 21.3925 127.809 21.8911 127.809 22.4943V27H126.72V22.5682C126.72 22.0111 126.575 21.5772 126.286 21.2663C125.997 20.9524 125.6 20.7955 125.095 20.7955C124.747 20.7955 124.436 20.8709 124.162 21.0217C123.892 21.1725 123.678 21.3925 123.521 21.6818C123.364 21.9711 123.285 22.322 123.285 22.7344ZM132.439 27.1477C131.848 27.1477 131.326 26.9985 130.874 26.6999C130.421 26.3983 130.068 25.9736 129.812 25.4258C129.557 24.8749 129.429 24.224 129.429 23.473C129.429 22.7282 129.557 22.0819 129.812 21.5341C130.068 20.9863 130.423 20.5631 130.879 20.2646C131.334 19.966 131.86 19.8168 132.457 19.8168C132.919 19.8168 133.284 19.8937 133.551 20.0476C133.822 20.1984 134.028 20.3707 134.17 20.5646C134.315 20.7554 134.427 20.9124 134.507 21.0355H134.599V17.5455H135.689V27H134.636V25.9105H134.507C134.427 26.0398 134.313 26.2029 134.165 26.3999C134.018 26.5937 133.807 26.7676 133.533 26.9215C133.259 27.0723 132.894 27.1477 132.439 27.1477ZM132.587 26.169C133.024 26.169 133.393 26.0552 133.695 25.8274C133.996 25.5966 134.225 25.2781 134.382 24.8718C134.539 24.4625 134.618 23.9901 134.618 23.4545C134.618 22.9252 134.541 22.462 134.387 22.065C134.233 21.6649 134.005 21.354 133.704 21.1325C133.402 20.9078 133.03 20.7955 132.587 20.7955C132.125 20.7955 131.74 20.9139 131.432 21.1509C131.128 21.3848 130.899 21.7034 130.745 22.1065C130.594 22.5066 130.518 22.956 130.518 23.4545C130.518 23.9593 130.595 24.4179 130.749 24.8303C130.906 25.2396 131.137 25.5658 131.442 25.8089C131.749 26.049 132.131 26.169 132.587 26.169ZM144.294 27.1477C143.654 27.1477 143.093 26.9954 142.609 26.6907C142.129 26.386 141.754 25.9598 141.483 25.4119C141.215 24.8641 141.081 24.224 141.081 23.4915C141.081 22.7528 141.215 22.1081 141.483 21.5572C141.754 21.0063 142.129 20.5785 142.609 20.2738C143.093 19.9691 143.654 19.8168 144.294 19.8168C144.935 19.8168 145.495 19.9691 145.975 20.2738C146.458 20.5785 146.834 21.0063 147.101 21.5572C147.372 22.1081 147.508 22.7528 147.508 23.4915C147.508 24.224 147.372 24.8641 147.101 25.4119C146.834 25.9598 146.458 26.386 145.975 26.6907C145.495 26.9954 144.935 27.1477 144.294 27.1477ZM144.294 26.169C144.781 26.169 145.181 26.0444 145.495 25.7951C145.809 25.5458 146.041 25.218 146.192 24.8118C146.343 24.4055 146.418 23.9654 146.418 23.4915C146.418 23.0175 146.343 22.5759 146.192 22.1665C146.041 21.7572 145.809 21.4264 145.495 21.174C145.181 20.9216 144.781 20.7955 144.294 20.7955C143.808 20.7955 143.408 20.9216 143.094 21.174C142.78 21.4264 142.548 21.7572 142.397 22.1665C142.246 22.5759 142.171 23.0175 142.171 23.4915C142.171 23.9654 142.246 24.4055 142.397 24.8118C142.548 25.218 142.78 25.5458 143.094 25.7951C143.408 26.0444 143.808 26.169 144.294 26.169ZM153.6 24.1009V19.9091H154.69V27H153.6V25.7997H153.527C153.36 26.1598 153.102 26.466 152.751 26.7184C152.4 26.9677 151.957 27.0923 151.421 27.0923C150.978 27.0923 150.584 26.9954 150.24 26.8015C149.895 26.6045 149.624 26.3091 149.427 25.9151C149.23 25.5181 149.132 25.018 149.132 24.4148V19.9091H150.221V24.3409C150.221 24.858 150.366 25.2704 150.655 25.5781C150.947 25.8859 151.32 26.0398 151.772 26.0398C152.043 26.0398 152.319 25.9705 152.599 25.832C152.882 25.6935 153.119 25.4812 153.31 25.195C153.503 24.9087 153.6 24.544 153.6 24.1009ZM156.646 27V19.9091H157.699V20.9801H157.773C157.902 20.6293 158.136 20.3446 158.474 20.1261C158.813 19.9076 159.195 19.7983 159.619 19.7983C159.699 19.7983 159.799 19.7998 159.919 19.8029C160.039 19.806 160.13 19.8106 160.192 19.8168V20.9247C160.155 20.9155 160.07 20.9016 159.938 20.8832C159.809 20.8616 159.672 20.8509 159.527 20.8509C159.182 20.8509 158.875 20.9232 158.604 21.0678C158.336 21.2094 158.124 21.4064 157.967 21.6587C157.813 21.908 157.736 22.1927 157.736 22.5128V27H156.646ZM170.084 21.4972L169.106 21.7741C169.044 21.611 168.953 21.4525 168.833 21.2987C168.716 21.1417 168.556 21.0124 168.353 20.9109C168.15 20.8093 167.89 20.7585 167.573 20.7585C167.139 20.7585 166.777 20.8585 166.488 21.0586C166.202 21.2556 166.059 21.5064 166.059 21.8111C166.059 22.0819 166.157 22.2958 166.354 22.4528C166.551 22.6097 166.859 22.7405 167.277 22.8452L168.33 23.1037C168.964 23.2576 169.436 23.493 169.747 23.81C170.058 24.1239 170.213 24.5286 170.213 25.0241C170.213 25.4304 170.097 25.7936 169.863 26.1136C169.632 26.4337 169.309 26.6861 168.893 26.8707C168.478 27.0554 167.995 27.1477 167.444 27.1477C166.72 27.1477 166.122 26.9908 165.648 26.6768C165.174 26.3629 164.874 25.9044 164.748 25.3011L165.782 25.0426C165.88 25.4242 166.066 25.7105 166.34 25.9013C166.617 26.0921 166.979 26.1875 167.425 26.1875C167.933 26.1875 168.336 26.0798 168.635 25.8643C168.936 25.6458 169.087 25.3842 169.087 25.0795C169.087 24.8333 169.001 24.6271 168.829 24.4609C168.656 24.2917 168.392 24.1655 168.035 24.0824L166.853 23.8054C166.203 23.6515 165.726 23.413 165.422 23.0898C165.12 22.7636 164.969 22.3558 164.969 21.8665C164.969 21.4664 165.082 21.1125 165.306 20.8047C165.534 20.4969 165.843 20.2553 166.234 20.0799C166.628 19.9045 167.074 19.8168 167.573 19.8168C168.275 19.8168 168.825 19.9706 169.226 20.2784C169.629 20.5862 169.915 20.9924 170.084 21.4972ZM171.814 27V19.9091H172.904V27H171.814ZM172.368 18.7273C172.156 18.7273 171.973 18.6549 171.819 18.5103C171.668 18.3656 171.593 18.1918 171.593 17.9886C171.593 17.7855 171.668 17.6116 171.819 17.467C171.973 17.3223 172.156 17.25 172.368 17.25C172.581 17.25 172.762 17.3223 172.913 17.467C173.067 17.6116 173.144 17.7855 173.144 17.9886C173.144 18.1918 173.067 18.3656 172.913 18.5103C172.762 18.6549 172.581 18.7273 172.368 18.7273ZM175.95 22.7344V27H174.86V19.9091H175.913V21.017H176.005C176.172 20.657 176.424 20.3677 176.762 20.1491C177.101 19.9276 177.538 19.8168 178.074 19.8168C178.554 19.8168 178.974 19.9152 179.334 20.1122C179.694 20.3061 179.974 20.6016 180.174 20.9986C180.374 21.3925 180.474 21.8911 180.474 22.4943V27H179.385V22.5682C179.385 22.0111 179.24 21.5772 178.951 21.2663C178.661 20.9524 178.264 20.7955 177.76 20.7955C177.412 20.7955 177.101 20.8709 176.827 21.0217C176.556 21.1725 176.342 21.3925 176.185 21.6818C176.028 21.9711 175.95 22.322 175.95 22.7344ZM185.288 29.8068C184.762 29.8068 184.309 29.7391 183.931 29.6037C183.552 29.4714 183.237 29.2959 182.985 29.0774C182.735 28.862 182.537 28.6312 182.389 28.3849L183.257 27.7756C183.355 27.9048 183.48 28.0526 183.631 28.2188C183.782 28.388 183.988 28.5342 184.249 28.6573C184.514 28.7835 184.86 28.8466 185.288 28.8466C185.861 28.8466 186.333 28.7081 186.705 28.4311C187.078 28.1541 187.264 27.7202 187.264 27.1293V25.6889H187.172C187.092 25.8182 186.978 25.9782 186.83 26.169C186.685 26.3568 186.476 26.5245 186.202 26.6722C185.931 26.8169 185.565 26.8892 185.104 26.8892C184.531 26.8892 184.017 26.7538 183.562 26.483C183.109 26.2121 182.751 25.8182 182.486 25.3011C182.224 24.7841 182.094 24.1562 182.094 23.4176C182.094 22.6913 182.221 22.0588 182.477 21.5202C182.732 20.9786 183.088 20.56 183.543 20.2646C183.999 19.966 184.525 19.8168 185.122 19.8168C185.584 19.8168 185.95 19.8937 186.221 20.0476C186.495 20.1984 186.704 20.3707 186.849 20.5646C186.996 20.7554 187.11 20.9124 187.19 21.0355H187.301V19.9091H188.354V27.2031C188.354 27.8125 188.215 28.308 187.938 28.6896C187.664 29.0743 187.295 29.3559 186.83 29.5344C186.368 29.716 185.854 29.8068 185.288 29.8068ZM185.251 25.9105C185.688 25.9105 186.058 25.8105 186.359 25.6104C186.661 25.4104 186.89 25.1226 187.047 24.7472C187.204 24.3717 187.282 23.9223 187.282 23.3991C187.282 22.8883 187.206 22.4374 187.052 22.0465C186.898 21.6557 186.67 21.3494 186.368 21.1278C186.067 20.9062 185.694 20.7955 185.251 20.7955C184.79 20.7955 184.405 20.9124 184.097 21.1463C183.792 21.3802 183.563 21.6941 183.409 22.0881C183.258 22.482 183.183 22.919 183.183 23.3991C183.183 23.8916 183.26 24.3271 183.414 24.7056C183.571 25.0811 183.802 25.3765 184.106 25.592C184.414 25.8043 184.796 25.9105 185.251 25.9105ZM191.398 17.5455V27H190.309V17.5455H191.398ZM196.328 27.1477C195.645 27.1477 195.055 26.9969 194.56 26.6953C194.067 26.3906 193.687 25.9659 193.419 25.4212C193.155 24.8733 193.022 24.2363 193.022 23.5099C193.022 22.7836 193.155 22.1435 193.419 21.5895C193.687 21.0324 194.06 20.5985 194.537 20.2876C195.017 19.9737 195.577 19.8168 196.217 19.8168C196.586 19.8168 196.951 19.8783 197.311 20.0014C197.671 20.1245 197.999 20.3246 198.294 20.6016C198.59 20.8755 198.825 21.2386 199.001 21.6911C199.176 22.1435 199.264 22.7005 199.264 23.3622V23.8239H193.798V22.8821H198.156C198.156 22.482 198.076 22.125 197.916 21.8111C197.759 21.4972 197.534 21.2494 197.242 21.0678C196.953 20.8862 196.611 20.7955 196.217 20.7955C195.783 20.7955 195.408 20.9032 195.091 21.1186C194.777 21.331 194.535 21.608 194.366 21.9496C194.197 22.2912 194.112 22.6574 194.112 23.0483V23.6761C194.112 24.2116 194.204 24.6656 194.389 25.038C194.577 25.4073 194.837 25.6889 195.169 25.8828C195.501 26.0736 195.888 26.169 196.328 26.169C196.614 26.169 196.873 26.129 197.103 26.049C197.337 25.9659 197.539 25.8428 197.708 25.6797C197.877 25.5135 198.008 25.3073 198.101 25.0611L199.153 25.3565C199.042 25.7135 198.856 26.0275 198.594 26.2983C198.333 26.5661 198.01 26.7753 197.625 26.9261C197.24 27.0739 196.808 27.1477 196.328 27.1477ZM207.38 27.1477C206.715 27.1477 206.143 26.9908 205.663 26.6768C205.183 26.3629 204.813 25.9305 204.555 25.3796C204.296 24.8287 204.167 24.1993 204.167 23.4915C204.167 22.7713 204.299 22.1358 204.564 21.5849C204.832 21.0309 205.204 20.5985 205.681 20.2876C206.161 19.9737 206.722 19.8168 207.362 19.8168C207.86 19.8168 208.31 19.9091 208.71 20.0938C209.11 20.2784 209.438 20.5369 209.693 20.8693C209.948 21.2017 210.107 21.5895 210.168 22.0327H209.079C208.996 21.7095 208.811 21.4233 208.525 21.174C208.242 20.9216 207.86 20.7955 207.38 20.7955C206.955 20.7955 206.583 20.9062 206.263 21.1278C205.946 21.3464 205.698 21.6557 205.52 22.0558C205.344 22.4528 205.257 22.919 205.257 23.4545C205.257 24.0024 205.343 24.4794 205.515 24.8857C205.69 25.2919 205.937 25.6074 206.254 25.832C206.574 26.0567 206.949 26.169 207.38 26.169C207.663 26.169 207.92 26.1198 208.151 26.0213C208.382 25.9228 208.577 25.7812 208.737 25.5966C208.897 25.4119 209.011 25.1903 209.079 24.9318H210.168C210.107 25.3504 209.955 25.7274 209.711 26.0629C209.471 26.3952 209.153 26.6599 208.756 26.8569C208.362 27.0508 207.903 27.1477 207.38 27.1477ZM214.695 27.1477C214.012 27.1477 213.423 26.9969 212.927 26.6953C212.435 26.3906 212.055 25.9659 211.787 25.4212C211.522 24.8733 211.39 24.2363 211.39 23.5099C211.39 22.7836 211.522 22.1435 211.787 21.5895C212.055 21.0324 212.427 20.5985 212.904 20.2876C213.384 19.9737 213.944 19.8168 214.584 19.8168C214.954 19.8168 215.318 19.8783 215.678 20.0014C216.039 20.1245 216.366 20.3246 216.662 20.6016C216.957 20.8755 217.193 21.2386 217.368 21.6911C217.544 22.1435 217.631 22.7005 217.631 23.3622V23.8239H212.165V22.8821H216.523C216.523 22.482 216.443 22.125 216.283 21.8111C216.126 21.4972 215.902 21.2494 215.609 21.0678C215.32 20.8862 214.978 20.7955 214.584 20.7955C214.15 20.7955 213.775 20.9032 213.458 21.1186C213.144 21.331 212.902 21.608 212.733 21.9496C212.564 22.2912 212.479 22.6574 212.479 23.0483V23.6761C212.479 24.2116 212.572 24.6656 212.756 25.038C212.944 25.4073 213.204 25.6889 213.536 25.8828C213.869 26.0736 214.255 26.169 214.695 26.169C214.981 26.169 215.24 26.129 215.471 26.049C215.705 25.9659 215.906 25.8428 216.076 25.6797C216.245 25.5135 216.376 25.3073 216.468 25.0611L217.52 25.3565C217.41 25.7135 217.223 26.0275 216.962 26.2983C216.7 26.5661 216.377 26.7753 215.992 26.9261C215.608 27.0739 215.175 27.1477 214.695 27.1477ZM220.339 17.5455V27H219.25V17.5455H220.339ZM223.385 17.5455V27H222.296V17.5455H223.385ZM231.636 27.1477C231.045 27.1477 230.524 26.9985 230.071 26.6999C229.619 26.3983 229.265 25.9736 229.01 25.4258C228.754 24.8749 228.626 24.224 228.626 23.473C228.626 22.7282 228.754 22.0819 229.01 21.5341C229.265 20.9863 229.62 20.5631 230.076 20.2646C230.531 19.966 231.058 19.8168 231.655 19.8168C232.116 19.8168 232.481 19.8937 232.749 20.0476C233.02 20.1984 233.226 20.3707 233.367 20.5646C233.512 20.7554 233.624 20.9124 233.704 21.0355H233.797V17.5455H234.886V27H233.834V25.9105H233.704C233.624 26.0398 233.511 26.2029 233.363 26.3999C233.215 26.5937 233.004 26.7676 232.73 26.9215C232.456 27.0723 232.092 27.1477 231.636 27.1477ZM231.784 26.169C232.221 26.169 232.59 26.0552 232.892 25.8274C233.194 25.5966 233.423 25.2781 233.58 24.8718C233.737 24.4625 233.815 23.9901 233.815 23.4545C233.815 22.9252 233.738 22.462 233.584 22.065C233.431 21.6649 233.203 21.354 232.901 21.1325C232.6 20.9078 232.227 20.7955 231.784 20.7955C231.322 20.7955 230.938 20.9139 230.63 21.1509C230.325 21.3848 230.096 21.7034 229.942 22.1065C229.791 22.5066 229.716 22.956 229.716 23.4545C229.716 23.9593 229.793 24.4179 229.947 24.8303C230.104 25.2396 230.334 25.5658 230.639 25.8089C230.947 26.049 231.329 26.169 231.784 26.169ZM239.081 27.1662C238.631 27.1662 238.223 27.0816 237.857 26.9123C237.491 26.7399 237.2 26.4922 236.985 26.169C236.769 25.8428 236.662 25.4489 236.662 24.9872C236.662 24.581 236.742 24.2517 236.902 23.9993C237.062 23.7438 237.276 23.5438 237.543 23.3991C237.811 23.2545 238.107 23.1468 238.43 23.076C238.756 23.0021 239.084 22.9437 239.413 22.9006C239.844 22.8452 240.193 22.8036 240.461 22.7759C240.732 22.7451 240.929 22.6944 241.052 22.6236C241.178 22.5528 241.241 22.4297 241.241 22.2543V22.2173C241.241 21.7618 241.116 21.4079 240.867 21.1555C240.621 20.9032 240.247 20.777 239.745 20.777C239.225 20.777 238.817 20.8909 238.522 21.1186C238.227 21.3464 238.019 21.5895 237.899 21.848L236.865 21.4787C237.049 21.0478 237.296 20.7124 237.603 20.4723C237.914 20.2292 238.253 20.0599 238.619 19.9645C238.988 19.866 239.351 19.8168 239.708 19.8168C239.936 19.8168 240.198 19.8445 240.493 19.8999C240.792 19.9522 241.08 20.0614 241.357 20.2276C241.637 20.3938 241.869 20.6446 242.054 20.9801C242.238 21.3156 242.331 21.7649 242.331 22.3281V27H241.241V26.0398H241.186C241.112 26.1937 240.989 26.3583 240.816 26.5337C240.644 26.7092 240.415 26.8584 240.129 26.9815C239.842 27.1046 239.493 27.1662 239.081 27.1662ZM239.247 26.1875C239.678 26.1875 240.041 26.1029 240.336 25.9336C240.635 25.7643 240.86 25.5458 241.01 25.2781C241.164 25.0103 241.241 24.7287 241.241 24.4332V23.4361C241.195 23.4915 241.093 23.5423 240.936 23.5884C240.783 23.6315 240.604 23.67 240.401 23.7038C240.201 23.7346 240.005 23.7623 239.815 23.7869C239.627 23.8085 239.475 23.8269 239.358 23.8423C239.074 23.8793 238.81 23.9393 238.564 24.0224C238.32 24.1024 238.123 24.224 237.973 24.3871C237.825 24.5471 237.751 24.7656 237.751 25.0426C237.751 25.4212 237.891 25.7074 238.171 25.9013C238.454 26.0921 238.813 26.1875 239.247 26.1875ZM247.364 19.9091V20.8324H243.689V19.9091H247.364ZM244.76 18.2102H245.85V24.9688C245.85 25.2765 245.894 25.5073 245.984 25.6612C246.076 25.812 246.193 25.9136 246.334 25.9659C246.479 26.0152 246.631 26.0398 246.792 26.0398C246.912 26.0398 247.01 26.0336 247.087 26.0213C247.164 26.0059 247.225 25.9936 247.272 25.9844L247.493 26.9631C247.419 26.9908 247.316 27.0185 247.184 27.0462C247.052 27.0769 246.884 27.0923 246.681 27.0923C246.373 27.0923 246.071 27.0262 245.776 26.8938C245.484 26.7615 245.24 26.5599 245.046 26.2891C244.856 26.0182 244.76 25.6766 244.76 25.2642V18.2102ZM251.05 27.1662C250.601 27.1662 250.193 27.0816 249.827 26.9123C249.461 26.7399 249.17 26.4922 248.955 26.169C248.739 25.8428 248.631 25.4489 248.631 24.9872C248.631 24.581 248.711 24.2517 248.871 23.9993C249.032 23.7438 249.245 23.5438 249.513 23.3991C249.781 23.2545 250.076 23.1468 250.4 23.076C250.726 23.0021 251.054 22.9437 251.383 22.9006C251.814 22.8452 252.163 22.8036 252.431 22.7759C252.702 22.7451 252.899 22.6944 253.022 22.6236C253.148 22.5528 253.211 22.4297 253.211 22.2543V22.2173C253.211 21.7618 253.086 21.4079 252.837 21.1555C252.591 20.9032 252.217 20.777 251.715 20.777C251.195 20.777 250.787 20.8909 250.492 21.1186C250.196 21.3464 249.989 21.5895 249.869 21.848L248.835 21.4787C249.019 21.0478 249.265 20.7124 249.573 20.4723C249.884 20.2292 250.223 20.0599 250.589 19.9645C250.958 19.866 251.321 19.8168 251.678 19.8168C251.906 19.8168 252.168 19.8445 252.463 19.8999C252.762 19.9522 253.049 20.0614 253.326 20.2276C253.606 20.3938 253.839 20.6446 254.023 20.9801C254.208 21.3156 254.3 21.7649 254.3 22.3281V27H253.211V26.0398H253.156C253.082 26.1937 252.959 26.3583 252.786 26.5337C252.614 26.7092 252.385 26.8584 252.098 26.9815C251.812 27.1046 251.463 27.1662 251.05 27.1662ZM251.217 26.1875C251.648 26.1875 252.011 26.1029 252.306 25.9336C252.605 25.7643 252.829 25.5458 252.98 25.2781C253.134 25.0103 253.211 24.7287 253.211 24.4332V23.4361C253.165 23.4915 253.063 23.5423 252.906 23.5884C252.752 23.6315 252.574 23.67 252.371 23.7038C252.171 23.7346 251.975 23.7623 251.784 23.7869C251.597 23.8085 251.444 23.8269 251.327 23.8423C251.044 23.8793 250.78 23.9393 250.533 24.0224C250.29 24.1024 250.093 24.224 249.943 24.3871C249.795 24.5471 249.721 24.7656 249.721 25.0426C249.721 25.4212 249.861 25.7074 250.141 25.9013C250.424 26.0921 250.783 26.1875 251.217 26.1875ZM262.748 27.1477C262.083 27.1477 261.511 26.9908 261.031 26.6768C260.55 26.3629 260.181 25.9305 259.923 25.3796C259.664 24.8287 259.535 24.1993 259.535 23.4915C259.535 22.7713 259.667 22.1358 259.932 21.5849C260.2 21.0309 260.572 20.5985 261.049 20.2876C261.529 19.9737 262.089 19.8168 262.729 19.8168C263.228 19.8168 263.677 19.9091 264.077 20.0938C264.478 20.2784 264.805 20.5369 265.061 20.8693C265.316 21.2017 265.475 21.5895 265.536 22.0327H264.447C264.364 21.7095 264.179 21.4233 263.893 21.174C263.61 20.9216 263.228 20.7955 262.748 20.7955C262.323 20.7955 261.951 20.9062 261.631 21.1278C261.314 21.3464 261.066 21.6557 260.888 22.0558C260.712 22.4528 260.624 22.919 260.624 23.4545C260.624 24.0024 260.711 24.4794 260.883 24.8857C261.058 25.2919 261.305 25.6074 261.622 25.832C261.942 26.0567 262.317 26.169 262.748 26.169C263.031 26.169 263.288 26.1198 263.519 26.0213C263.75 25.9228 263.945 25.7812 264.105 25.5966C264.265 25.4119 264.379 25.1903 264.447 24.9318H265.536C265.475 25.3504 265.322 25.7274 265.079 26.0629C264.839 26.3952 264.521 26.6599 264.124 26.8569C263.73 27.0508 263.271 27.1477 262.748 27.1477ZM269.971 27.1477C269.331 27.1477 268.769 26.9954 268.286 26.6907C267.806 26.386 267.43 25.9598 267.159 25.4119C266.891 24.8641 266.758 24.224 266.758 23.4915C266.758 22.7528 266.891 22.1081 267.159 21.5572C267.43 21.0063 267.806 20.5785 268.286 20.2738C268.769 19.9691 269.331 19.8168 269.971 19.8168C270.611 19.8168 271.171 19.9691 271.651 20.2738C272.134 20.5785 272.51 21.0063 272.777 21.5572C273.048 22.1081 273.184 22.7528 273.184 23.4915C273.184 24.224 273.048 24.8641 272.777 25.4119C272.51 25.9598 272.134 26.386 271.651 26.6907C271.171 26.9954 270.611 27.1477 269.971 27.1477ZM269.971 26.169C270.457 26.169 270.857 26.0444 271.171 25.7951C271.485 25.5458 271.717 25.218 271.868 24.8118C272.019 24.4055 272.094 23.9654 272.094 23.4915C272.094 23.0175 272.019 22.5759 271.868 22.1665C271.717 21.7572 271.485 21.4264 271.171 21.174C270.857 20.9216 270.457 20.7955 269.971 20.7955C269.484 20.7955 269.084 20.9216 268.77 21.174C268.456 21.4264 268.224 21.7572 268.073 22.1665C267.922 22.5759 267.847 23.0175 267.847 23.4915C267.847 23.9654 267.922 24.4055 268.073 24.8118C268.224 25.218 268.456 25.5458 268.77 25.7951C269.084 26.0444 269.484 26.169 269.971 26.169ZM274.808 27V19.9091H275.86V20.9801H275.934C276.064 20.6293 276.297 20.3446 276.636 20.1261C276.974 19.9076 277.356 19.7983 277.781 19.7983C277.861 19.7983 277.961 19.7998 278.081 19.8029C278.201 19.806 278.292 19.8106 278.353 19.8168V20.9247C278.316 20.9155 278.232 20.9016 278.099 20.8832C277.97 20.8616 277.833 20.8509 277.689 20.8509C277.344 20.8509 277.036 20.9232 276.765 21.0678C276.497 21.2094 276.285 21.4064 276.128 21.6587C275.974 21.908 275.897 22.1927 275.897 22.5128V27H274.808ZM279.606 29.6591V19.9091H280.658V21.0355H280.788C280.868 20.9124 280.978 20.7554 281.12 20.5646C281.265 20.3707 281.471 20.1984 281.739 20.0476C282.009 19.8937 282.376 19.8168 282.837 19.8168C283.434 19.8168 283.961 19.966 284.416 20.2646C284.872 20.5631 285.227 20.9863 285.483 21.5341C285.738 22.0819 285.866 22.7282 285.866 23.473C285.866 24.224 285.738 24.8749 285.483 25.4258C285.227 25.9736 284.873 26.3983 284.421 26.6999C283.968 26.9985 283.447 27.1477 282.856 27.1477C282.4 27.1477 282.036 27.0723 281.762 26.9215C281.488 26.7676 281.277 26.5937 281.129 26.3999C280.981 26.2029 280.868 26.0398 280.788 25.9105H280.695V29.6591H279.606ZM280.677 23.4545C280.677 23.9901 280.755 24.4625 280.912 24.8718C281.069 25.2781 281.298 25.5966 281.6 25.8274C281.902 26.0552 282.271 26.169 282.708 26.169C283.164 26.169 283.544 26.049 283.848 25.8089C284.156 25.5658 284.387 25.2396 284.541 24.8303C284.698 24.4179 284.776 23.9593 284.776 23.4545C284.776 22.956 284.699 22.5066 284.545 22.1065C284.395 21.7034 284.165 21.3848 283.858 21.1509C283.553 20.9139 283.17 20.7955 282.708 20.7955C282.265 20.7955 281.892 20.9078 281.591 21.1325C281.289 21.354 281.061 21.6649 280.908 22.065C280.754 22.462 280.677 22.9252 280.677 23.4545ZM291.957 24.1009V19.9091H293.047V27H291.957V25.7997H291.884C291.717 26.1598 291.459 26.466 291.108 26.7184C290.757 26.9677 290.314 27.0923 289.778 27.0923C289.335 27.0923 288.941 26.9954 288.597 26.8015C288.252 26.6045 287.981 26.3091 287.784 25.9151C287.587 25.5181 287.489 25.018 287.489 24.4148V19.9091H288.578V24.3409C288.578 24.858 288.723 25.2704 289.012 25.5781C289.304 25.8859 289.677 26.0398 290.129 26.0398C290.4 26.0398 290.676 25.9705 290.956 25.832C291.239 25.6935 291.476 25.4812 291.667 25.195C291.86 24.9087 291.957 24.544 291.957 24.1009ZM300.026 21.4972L299.047 21.7741C298.986 21.611 298.895 21.4525 298.775 21.2987C298.658 21.1417 298.498 21.0124 298.295 20.9109C298.092 20.8093 297.832 20.7585 297.515 20.7585C297.081 20.7585 296.719 20.8585 296.43 21.0586C296.144 21.2556 296 21.5064 296 21.8111C296 22.0819 296.099 22.2958 296.296 22.4528C296.493 22.6097 296.801 22.7405 297.219 22.8452L298.272 23.1037C298.906 23.2576 299.378 23.493 299.689 23.81C300 24.1239 300.155 24.5286 300.155 25.0241C300.155 25.4304 300.038 25.7936 299.804 26.1136C299.574 26.4337 299.25 26.6861 298.835 26.8707C298.42 27.0554 297.936 27.1477 297.385 27.1477C296.662 27.1477 296.064 26.9908 295.59 26.6768C295.116 26.3629 294.816 25.9044 294.689 25.3011L295.723 25.0426C295.822 25.4242 296.008 25.7105 296.282 25.9013C296.559 26.0921 296.921 26.1875 297.367 26.1875C297.875 26.1875 298.278 26.0798 298.576 25.8643C298.878 25.6458 299.029 25.3842 299.029 25.0795C299.029 24.8333 298.943 24.6271 298.77 24.4609C298.598 24.2917 298.333 24.1655 297.976 24.0824L296.795 23.8054C296.145 23.6515 295.668 23.413 295.363 23.0898C295.062 22.7636 294.911 22.3558 294.911 21.8665C294.911 21.4664 295.023 21.1125 295.248 20.8047C295.476 20.4969 295.785 20.2553 296.176 20.0799C296.57 19.9045 297.016 19.8168 297.515 19.8168C298.216 19.8168 298.767 19.9706 299.167 20.2784C299.571 20.5862 299.857 20.9924 300.026 21.4972ZM308.457 19.9091V20.8324H304.783V19.9091H308.457ZM305.854 18.2102H306.943V24.9688C306.943 25.2765 306.988 25.5073 307.077 25.6612C307.169 25.812 307.286 25.9136 307.428 25.9659C307.573 26.0152 307.725 26.0398 307.885 26.0398C308.005 26.0398 308.103 26.0336 308.18 26.0213C308.257 26.0059 308.319 25.9936 308.365 25.9844L308.587 26.9631C308.513 26.9908 308.41 27.0185 308.277 27.0462C308.145 27.0769 307.977 27.0923 307.774 27.0923C307.466 27.0923 307.165 27.0262 306.869 26.8938C306.577 26.7615 306.334 26.5599 306.14 26.2891C305.949 26.0182 305.854 25.6766 305.854 25.2642V18.2102ZM312.862 27.1477C312.222 27.1477 311.66 26.9954 311.177 26.6907C310.697 26.386 310.321 25.9598 310.05 25.4119C309.783 24.8641 309.649 24.224 309.649 23.4915C309.649 22.7528 309.783 22.1081 310.05 21.5572C310.321 21.0063 310.697 20.5785 311.177 20.2738C311.66 19.9691 312.222 19.8168 312.862 19.8168C313.502 19.8168 314.062 19.9691 314.542 20.2738C315.025 20.5785 315.401 21.0063 315.669 21.5572C315.939 22.1081 316.075 22.7528 316.075 23.4915C316.075 24.224 315.939 24.8641 315.669 25.4119C315.401 25.9598 315.025 26.386 314.542 26.6907C314.062 26.9954 313.502 27.1477 312.862 27.1477ZM312.862 26.169C313.348 26.169 313.748 26.0444 314.062 25.7951C314.376 25.5458 314.608 25.218 314.759 24.8118C314.91 24.4055 314.985 23.9654 314.985 23.4915C314.985 23.0175 314.91 22.5759 314.759 22.1665C314.608 21.7572 314.376 21.4264 314.062 21.174C313.748 20.9216 313.348 20.7955 312.862 20.7955C312.375 20.7955 311.975 20.9216 311.661 21.174C311.348 21.4264 311.115 21.7572 310.964 22.1665C310.814 22.5759 310.738 23.0175 310.738 23.4915C310.738 23.9654 310.814 24.4055 310.964 24.8118C311.115 25.218 311.348 25.5458 311.661 25.7951C311.975 26.0444 312.375 26.169 312.862 26.169ZM321.316 27V19.9091H322.406V27H321.316ZM321.87 18.7273C321.658 18.7273 321.475 18.6549 321.321 18.5103C321.17 18.3656 321.095 18.1918 321.095 17.9886C321.095 17.7855 321.17 17.6116 321.321 17.467C321.475 17.3223 321.658 17.25 321.87 17.25C322.082 17.25 322.264 17.3223 322.415 17.467C322.569 17.6116 322.646 17.7855 322.646 17.9886C322.646 18.1918 322.569 18.3656 322.415 18.5103C322.264 18.6549 322.082 18.7273 321.87 18.7273ZM324.362 27V19.9091H325.415V21.017H325.507C325.655 20.6385 325.893 20.3446 326.223 20.1353C326.552 19.9229 326.947 19.8168 327.409 19.8168C327.877 19.8168 328.266 19.9229 328.577 20.1353C328.891 20.3446 329.136 20.6385 329.311 21.017H329.385C329.566 20.6508 329.839 20.36 330.202 20.1445C330.565 19.926 331.001 19.8168 331.508 19.8168C332.142 19.8168 332.661 20.0153 333.064 20.4123C333.467 20.8062 333.669 21.4202 333.669 22.2543V27H332.579V22.2543C332.579 21.7311 332.436 21.3571 332.15 21.1325C331.864 20.9078 331.527 20.7955 331.139 20.7955C330.64 20.7955 330.254 20.9463 329.98 21.2479C329.706 21.5464 329.569 21.925 329.569 22.3835V27H328.462V22.1435C328.462 21.7403 328.331 21.4156 328.069 21.1694C327.808 20.9201 327.471 20.7955 327.058 20.7955C326.775 20.7955 326.51 20.8709 326.264 21.0217C326.021 21.1725 325.824 21.3817 325.673 21.6495C325.525 21.9142 325.452 22.2204 325.452 22.5682V27H324.362ZM335.622 29.6591V19.9091H336.674V21.0355H336.804C336.884 20.9124 336.995 20.7554 337.136 20.5646C337.281 20.3707 337.487 20.1984 337.755 20.0476C338.026 19.8937 338.392 19.8168 338.853 19.8168C339.451 19.8168 339.977 19.966 340.432 20.2646C340.888 20.5631 341.243 20.9863 341.499 21.5341C341.754 22.0819 341.882 22.7282 341.882 23.473C341.882 24.224 341.754 24.8749 341.499 25.4258C341.243 25.9736 340.889 26.3983 340.437 26.6999C339.984 26.9985 339.463 27.1477 338.872 27.1477C338.416 27.1477 338.052 27.0723 337.778 26.9215C337.504 26.7676 337.293 26.5937 337.145 26.3999C336.998 26.2029 336.884 26.0398 336.804 25.9105H336.711V29.6591H335.622ZM336.693 23.4545C336.693 23.9901 336.771 24.4625 336.928 24.8718C337.085 25.2781 337.315 25.5966 337.616 25.8274C337.918 26.0552 338.287 26.169 338.724 26.169C339.18 26.169 339.56 26.049 339.864 25.8089C340.172 25.5658 340.403 25.2396 340.557 24.8303C340.714 24.4179 340.792 23.9593 340.792 23.4545C340.792 22.956 340.715 22.5066 340.562 22.1065C340.411 21.7034 340.181 21.3848 339.874 21.1509C339.569 20.9139 339.186 20.7955 338.724 20.7955C338.281 20.7955 337.909 20.9078 337.607 21.1325C337.305 21.354 337.078 21.6649 336.924 22.065C336.77 22.462 336.693 22.9252 336.693 23.4545ZM343.505 27V19.9091H344.557V20.9801H344.631C344.76 20.6293 344.994 20.3446 345.333 20.1261C345.671 19.9076 346.053 19.7983 346.478 19.7983C346.558 19.7983 346.658 19.7998 346.778 19.8029C346.898 19.806 346.989 19.8106 347.05 19.8168V20.9247C347.013 20.9155 346.929 20.9016 346.796 20.8832C346.667 20.8616 346.53 20.8509 346.385 20.8509C346.041 20.8509 345.733 20.9232 345.462 21.0678C345.194 21.2094 344.982 21.4064 344.825 21.6587C344.671 21.908 344.594 22.1927 344.594 22.5128V27H343.505ZM350.968 27.1477C350.327 27.1477 349.766 26.9954 349.283 26.6907C348.802 26.386 348.427 25.9598 348.156 25.4119C347.888 24.8641 347.754 24.224 347.754 23.4915C347.754 22.7528 347.888 22.1081 348.156 21.5572C348.427 21.0063 348.802 20.5785 349.283 20.2738C349.766 19.9691 350.327 19.8168 350.968 19.8168C351.608 19.8168 352.168 19.9691 352.648 20.2738C353.131 20.5785 353.507 21.0063 353.774 21.5572C354.045 22.1081 354.181 22.7528 354.181 23.4915C354.181 24.224 354.045 24.8641 353.774 25.4119C353.507 25.9598 353.131 26.386 352.648 26.6907C352.168 26.9954 351.608 27.1477 350.968 27.1477ZM350.968 26.169C351.454 26.169 351.854 26.0444 352.168 25.7951C352.482 25.5458 352.714 25.218 352.865 24.8118C353.016 24.4055 353.091 23.9654 353.091 23.4915C353.091 23.0175 353.016 22.5759 352.865 22.1665C352.714 21.7572 352.482 21.4264 352.168 21.174C351.854 20.9216 351.454 20.7955 350.968 20.7955C350.481 20.7955 350.081 20.9216 349.767 21.174C349.453 21.4264 349.221 21.7572 349.07 22.1665C348.919 22.5759 348.844 23.0175 348.844 23.4915C348.844 23.9654 348.919 24.4055 349.07 24.8118C349.221 25.218 349.453 25.5458 349.767 25.7951C350.081 26.0444 350.481 26.169 350.968 26.169ZM361.349 19.9091L358.727 27H357.619L354.997 19.9091H356.179L358.136 25.5597H358.21L360.167 19.9091H361.349ZM365.467 27.1477C364.784 27.1477 364.195 26.9969 363.699 26.6953C363.207 26.3906 362.827 25.9659 362.559 25.4212C362.294 24.8733 362.162 24.2363 362.162 23.5099C362.162 22.7836 362.294 22.1435 362.559 21.5895C362.827 21.0324 363.199 20.5985 363.676 20.2876C364.156 19.9737 364.716 19.8168 365.356 19.8168C365.726 19.8168 366.09 19.8783 366.451 20.0014C366.811 20.1245 367.138 20.3246 367.434 20.6016C367.729 20.8755 367.965 21.2386 368.14 21.6911C368.316 22.1435 368.403 22.7005 368.403 23.3622V23.8239H362.937V22.8821H367.295C367.295 22.482 367.215 22.125 367.055 21.8111C366.898 21.4972 366.674 21.2494 366.381 21.0678C366.092 20.8862 365.75 20.7955 365.356 20.7955C364.922 20.7955 364.547 20.9032 364.23 21.1186C363.916 21.331 363.675 21.608 363.505 21.9496C363.336 22.2912 363.251 22.6574 363.251 23.0483V23.6761C363.251 24.2116 363.344 24.6656 363.528 25.038C363.716 25.4073 363.976 25.6889 364.309 25.8828C364.641 26.0736 365.027 26.169 365.467 26.169C365.753 26.169 366.012 26.129 366.243 26.049C366.477 25.9659 366.678 25.8428 366.848 25.6797C367.017 25.5135 367.148 25.3073 367.24 25.0611L368.293 25.3565C368.182 25.7135 367.996 26.0275 367.734 26.2983C367.472 26.5661 367.149 26.7753 366.764 26.9261C366.38 27.0739 365.947 27.1477 365.467 27.1477ZM373.639 27V19.9091H374.691V20.9801H374.765C374.895 20.6293 375.128 20.3446 375.467 20.1261C375.806 19.9076 376.187 19.7983 376.612 19.7983C376.692 19.7983 376.792 19.7998 376.912 19.8029C377.032 19.806 377.123 19.8106 377.184 19.8168V20.9247C377.147 20.9155 377.063 20.9016 376.93 20.8832C376.801 20.8616 376.664 20.8509 376.52 20.8509C376.175 20.8509 375.867 20.9232 375.596 21.0678C375.329 21.2094 375.116 21.4064 374.959 21.6587C374.805 21.908 374.728 22.1927 374.728 22.5128V27H373.639ZM381.194 27.1477C380.511 27.1477 379.921 26.9969 379.426 26.6953C378.933 26.3906 378.553 25.9659 378.286 25.4212C378.021 24.8733 377.889 24.2363 377.889 23.5099C377.889 22.7836 378.021 22.1435 378.286 21.5895C378.553 21.0324 378.926 20.5985 379.403 20.2876C379.883 19.9737 380.443 19.8168 381.083 19.8168C381.453 19.8168 381.817 19.8783 382.177 20.0014C382.537 20.1245 382.865 20.3246 383.161 20.6016C383.456 20.8755 383.692 21.2386 383.867 21.6911C384.042 22.1435 384.13 22.7005 384.13 23.3622V23.8239H378.664V22.8821H383.022C383.022 22.482 382.942 22.125 382.782 21.8111C382.625 21.4972 382.4 21.2494 382.108 21.0678C381.819 20.8862 381.477 20.7955 381.083 20.7955C380.649 20.7955 380.274 20.9032 379.957 21.1186C379.643 21.331 379.401 21.608 379.232 21.9496C379.063 22.2912 378.978 22.6574 378.978 23.0483V23.6761C378.978 24.2116 379.07 24.6656 379.255 25.038C379.443 25.4073 379.703 25.6889 380.035 25.8828C380.368 26.0736 380.754 26.169 381.194 26.169C381.48 26.169 381.739 26.129 381.97 26.049C382.203 25.9659 382.405 25.8428 382.574 25.6797C382.744 25.5135 382.874 25.3073 382.967 25.0611L384.019 25.3565C383.908 25.7135 383.722 26.0275 383.461 26.2983C383.199 26.5661 382.876 26.7753 382.491 26.9261C382.107 27.0739 381.674 27.1477 381.194 27.1477ZM390.771 21.4972L389.792 21.7741C389.731 21.611 389.64 21.4525 389.52 21.2987C389.403 21.1417 389.243 21.0124 389.04 20.9109C388.837 20.8093 388.577 20.7585 388.26 20.7585C387.826 20.7585 387.464 20.8585 387.175 21.0586C386.889 21.2556 386.746 21.5064 386.746 21.8111C386.746 22.0819 386.844 22.2958 387.041 22.4528C387.238 22.6097 387.546 22.7405 387.964 22.8452L389.017 23.1037C389.651 23.2576 390.123 23.493 390.434 23.81C390.745 24.1239 390.9 24.5286 390.9 25.0241C390.9 25.4304 390.783 25.7936 390.55 26.1136C390.319 26.4337 389.996 26.6861 389.58 26.8707C389.165 27.0554 388.681 27.1477 388.13 27.1477C387.407 27.1477 386.809 26.9908 386.335 26.6768C385.861 26.3629 385.561 25.9044 385.434 25.3011L386.469 25.0426C386.567 25.4242 386.753 25.7105 387.027 25.9013C387.304 26.0921 387.666 26.1875 388.112 26.1875C388.62 26.1875 389.023 26.0798 389.322 25.8643C389.623 25.6458 389.774 25.3842 389.774 25.0795C389.774 24.8333 389.688 24.6271 389.515 24.4609C389.343 24.2917 389.078 24.1655 388.721 24.0824L387.54 23.8054C386.89 23.6515 386.413 23.413 386.108 23.0898C385.807 22.7636 385.656 22.3558 385.656 21.8665C385.656 21.4664 385.768 21.1125 385.993 20.8047C386.221 20.4969 386.53 20.2553 386.921 20.0799C387.315 19.9045 387.761 19.8168 388.26 19.8168C388.961 19.8168 389.512 19.9706 389.912 20.2784C390.316 20.5862 390.602 20.9924 390.771 21.4972ZM396.97 24.1009V19.9091H398.06V27H396.97V25.7997H396.896C396.73 26.1598 396.472 26.466 396.121 26.7184C395.77 26.9677 395.327 27.0923 394.791 27.0923C394.348 27.0923 393.954 26.9954 393.609 26.8015C393.265 26.6045 392.994 26.3091 392.797 25.9151C392.6 25.5181 392.501 25.018 392.501 24.4148V19.9091H393.591V24.3409C393.591 24.858 393.736 25.2704 394.025 25.5781C394.317 25.8859 394.69 26.0398 395.142 26.0398C395.413 26.0398 395.688 25.9705 395.968 25.832C396.251 25.6935 396.488 25.4812 396.679 25.195C396.873 24.9087 396.97 24.544 396.97 24.1009ZM401.106 17.5455V27H400.016V17.5455H401.106ZM406.146 19.9091V20.8324H402.471V19.9091H406.146ZM403.542 18.2102H404.632V24.9688C404.632 25.2765 404.676 25.5073 404.766 25.6612C404.858 25.812 404.975 25.9136 405.116 25.9659C405.261 26.0152 405.413 26.0398 405.573 26.0398C405.693 26.0398 405.792 26.0336 405.869 26.0213C405.946 26.0059 406.007 25.9936 406.054 25.9844L406.275 26.9631C406.201 26.9908 406.098 27.0185 405.966 27.0462C405.833 27.0769 405.666 27.0923 405.463 27.0923C405.155 27.0923 404.853 27.0262 404.558 26.8938C404.265 26.7615 404.022 26.5599 403.828 26.2891C403.638 26.0182 403.542 25.6766 403.542 25.2642V18.2102ZM412.641 21.4972L411.663 21.7741C411.601 21.611 411.51 21.4525 411.39 21.2987C411.273 21.1417 411.113 21.0124 410.91 20.9109C410.707 20.8093 410.447 20.7585 410.13 20.7585C409.696 20.7585 409.335 20.8585 409.045 21.0586C408.759 21.2556 408.616 21.5064 408.616 21.8111C408.616 22.0819 408.714 22.2958 408.911 22.4528C409.108 22.6097 409.416 22.7405 409.835 22.8452L410.887 23.1037C411.521 23.2576 411.994 23.493 412.304 23.81C412.615 24.1239 412.771 24.5286 412.771 25.0241C412.771 25.4304 412.654 25.7936 412.42 26.1136C412.189 26.4337 411.866 26.6861 411.45 26.8707C411.035 27.0554 410.552 27.1477 410.001 27.1477C409.278 27.1477 408.679 26.9908 408.205 26.6768C407.731 26.3629 407.431 25.9044 407.305 25.3011L408.339 25.0426C408.437 25.4242 408.624 25.7105 408.898 25.9013C409.174 26.0921 409.536 26.1875 409.982 26.1875C410.49 26.1875 410.893 26.0798 411.192 25.8643C411.494 25.6458 411.644 25.3842 411.644 25.0795C411.644 24.8333 411.558 24.6271 411.386 24.4609C411.213 24.2917 410.949 24.1655 410.592 24.0824L409.41 23.8054C408.761 23.6515 408.284 23.413 407.979 23.0898C407.677 22.7636 407.526 22.3558 407.526 21.8665C407.526 21.4664 407.639 21.1125 407.863 20.8047C408.091 20.4969 408.4 20.2553 408.791 20.0799C409.185 19.9045 409.632 19.8168 410.13 19.8168C410.832 19.8168 411.383 19.9706 411.783 20.2784C412.186 20.5862 412.472 20.9924 412.641 21.4972ZM415.166 27.0739C414.938 27.0739 414.743 26.9923 414.579 26.8292C414.416 26.6661 414.335 26.4706 414.335 26.2429C414.335 26.0152 414.416 25.8197 414.579 25.6566C414.743 25.4935 414.938 25.4119 415.166 25.4119C415.394 25.4119 415.589 25.4935 415.752 25.6566C415.915 25.8197 415.997 26.0152 415.997 26.2429C415.997 26.3937 415.958 26.5322 415.881 26.6584C415.807 26.7846 415.707 26.8861 415.581 26.9631C415.458 27.0369 415.32 27.0739 415.166 27.0739ZM57.4896 47H54.572V37.5455H57.6189C58.536 37.5455 59.3208 37.7347 59.9733 38.1133C60.6257 38.4888 61.1258 39.0289 61.4736 39.7337C61.8214 40.4354 61.9953 41.2756 61.9953 42.2543C61.9953 43.2391 61.8198 44.087 61.469 44.7979C61.1181 45.5058 60.6073 46.0505 59.9363 46.4322C59.2654 46.8107 58.4498 47 57.4896 47ZM55.7169 45.9844H57.4157C58.1975 45.9844 58.8453 45.8336 59.3593 45.532C59.8732 45.2304 60.2564 44.801 60.5088 44.244C60.7611 43.6869 60.8873 43.0237 60.8873 42.2543C60.8873 41.491 60.7627 40.8339 60.5134 40.283C60.2641 39.729 59.8917 39.3043 59.3962 39.0089C58.9007 38.7103 58.2836 38.5611 57.545 38.5611H55.7169V45.9844ZM66.6097 47.1477C65.9695 47.1477 65.4079 46.9954 64.9247 46.6907C64.4446 46.386 64.0691 45.9598 63.7982 45.4119C63.5305 44.8641 63.3966 44.224 63.3966 43.4915C63.3966 42.7528 63.5305 42.1081 63.7982 41.5572C64.0691 41.0063 64.4446 40.5785 64.9247 40.2738C65.4079 39.9691 65.9695 39.8168 66.6097 39.8168C67.2498 39.8168 67.81 39.9691 68.2901 40.2738C68.7733 40.5785 69.1487 41.0063 69.4165 41.5572C69.6873 42.1081 69.8228 42.7528 69.8228 43.4915C69.8228 44.224 69.6873 44.8641 69.4165 45.4119C69.1487 45.9598 68.7733 46.386 68.2901 46.6907C67.81 46.9954 67.2498 47.1477 66.6097 47.1477ZM66.6097 46.169C67.096 46.169 67.496 46.0444 67.81 45.7951C68.1239 45.5458 68.3563 45.218 68.5071 44.8118C68.6579 44.4055 68.7333 43.9654 68.7333 43.4915C68.7333 43.0175 68.6579 42.5759 68.5071 42.1665C68.3563 41.7572 68.1239 41.4264 67.81 41.174C67.496 40.9216 67.096 40.7955 66.6097 40.7955C66.1234 40.7955 65.7233 40.9216 65.4094 41.174C65.0955 41.4264 64.8631 41.7572 64.7123 42.1665C64.5615 42.5759 64.4861 43.0175 64.4861 43.4915C64.4861 43.9654 64.5615 44.4055 64.7123 44.8118C64.8631 45.218 65.0955 45.5458 65.4094 45.7951C65.7233 46.0444 66.1234 46.169 66.6097 46.169ZM72.8122 47L70.6517 39.9091H71.7965L73.3292 45.3381H73.4031L74.9173 39.9091H76.0806L77.5764 45.3196H77.6502L79.1829 39.9091H80.3278L78.1673 47H77.0963L75.5451 41.5526H75.4343L73.8832 47H72.8122ZM82.8186 42.7344V47H81.7291V39.9091H82.7817V41.017H82.874C83.0402 40.657 83.2926 40.3677 83.6311 40.1491C83.9697 39.9276 84.4067 39.8168 84.9422 39.8168C85.4223 39.8168 85.8424 39.9152 86.2025 40.1122C86.5626 40.3061 86.8426 40.6016 87.0427 40.9986C87.2427 41.3925 87.3428 41.8911 87.3428 42.4943V47H86.2533V42.5682C86.2533 42.0111 86.1086 41.5772 85.8193 41.2663C85.53 40.9524 85.133 40.7955 84.6283 40.7955C84.2805 40.7955 83.9697 40.8709 83.6957 41.0217C83.4249 41.1725 83.211 41.3925 83.0541 41.6818C82.8971 41.9711 82.8186 42.322 82.8186 42.7344ZM90.3841 37.5455V47H89.2946V37.5455H90.3841ZM95.2213 47.1477C94.5811 47.1477 94.0194 46.9954 93.5362 46.6907C93.0561 46.386 92.6807 45.9598 92.4098 45.4119C92.1421 44.8641 92.0082 44.224 92.0082 43.4915C92.0082 42.7528 92.1421 42.1081 92.4098 41.5572C92.6807 41.0063 93.0561 40.5785 93.5362 40.2738C94.0194 39.9691 94.5811 39.8168 95.2213 39.8168C95.8614 39.8168 96.4215 39.9691 96.9017 40.2738C97.3849 40.5785 97.7603 41.0063 98.0281 41.5572C98.2989 42.1081 98.4343 42.7528 98.4343 43.4915C98.4343 44.224 98.2989 44.8641 98.0281 45.4119C97.7603 45.9598 97.3849 46.386 96.9017 46.6907C96.4215 46.9954 95.8614 47.1477 95.2213 47.1477ZM95.2213 46.169C95.7075 46.169 96.1076 46.0444 96.4215 45.7951C96.7355 45.5458 96.9678 45.218 97.1186 44.8118C97.2694 44.4055 97.3448 43.9654 97.3448 43.4915C97.3448 43.0175 97.2694 42.5759 97.1186 42.1665C96.9678 41.7572 96.7355 41.4264 96.4215 41.174C96.1076 40.9216 95.7075 40.7955 95.2213 40.7955C94.735 40.7955 94.3349 40.9216 94.021 41.174C93.7071 41.4264 93.4747 41.7572 93.3239 42.1665C93.1731 42.5759 93.0977 43.0175 93.0977 43.4915C93.0977 43.9654 93.1731 44.4055 93.3239 44.8118C93.4747 45.218 93.7071 45.5458 94.021 45.7951C94.3349 46.0444 94.735 46.169 95.2213 46.169ZM102.145 47.1662C101.696 47.1662 101.288 47.0816 100.922 46.9123C100.555 46.7399 100.265 46.4922 100.049 46.169C99.8337 45.8428 99.726 45.4489 99.726 44.9872C99.726 44.581 99.806 44.2517 99.9661 43.9993C100.126 43.7438 100.34 43.5438 100.608 43.3991C100.876 43.2545 101.171 43.1468 101.494 43.076C101.82 43.0021 102.148 42.9437 102.477 42.9006C102.908 42.8452 103.258 42.8036 103.525 42.7759C103.796 42.7451 103.993 42.6944 104.116 42.6236C104.242 42.5528 104.306 42.4297 104.306 42.2543V42.2173C104.306 41.7618 104.181 41.4079 103.932 41.1555C103.685 40.9032 103.311 40.777 102.81 40.777C102.29 40.777 101.882 40.8909 101.586 41.1186C101.291 41.3464 101.083 41.5895 100.963 41.848L99.9292 41.4787C100.114 41.0478 100.36 40.7124 100.668 40.4723C100.979 40.2292 101.317 40.0599 101.683 39.9645C102.053 39.866 102.416 39.8168 102.773 39.8168C103.001 39.8168 103.262 39.8445 103.558 39.8999C103.856 39.9522 104.144 40.0614 104.421 40.2276C104.701 40.3938 104.933 40.6446 105.118 40.9801C105.303 41.3156 105.395 41.7649 105.395 42.3281V47H104.306V46.0398H104.25C104.176 46.1937 104.053 46.3583 103.881 46.5337C103.709 46.7092 103.479 46.8584 103.193 46.9815C102.907 47.1046 102.557 47.1662 102.145 47.1662ZM102.311 46.1875C102.742 46.1875 103.105 46.1029 103.401 45.9336C103.699 45.7643 103.924 45.5458 104.075 45.2781C104.229 45.0103 104.306 44.7287 104.306 44.4332V43.4361C104.259 43.4915 104.158 43.5423 104.001 43.5884C103.847 43.6315 103.669 43.67 103.465 43.7038C103.265 43.7346 103.07 43.7623 102.879 43.7869C102.691 43.8085 102.539 43.8269 102.422 43.8423C102.139 43.8793 101.874 43.9393 101.628 44.0224C101.385 44.1024 101.188 44.224 101.037 44.3871C100.889 44.5471 100.816 44.7656 100.816 45.0426C100.816 45.4212 100.956 45.7074 101.236 45.9013C101.519 46.0921 101.877 46.1875 102.311 46.1875ZM110.022 47.1477C109.431 47.1477 108.91 46.9985 108.457 46.6999C108.005 46.3983 107.651 45.9736 107.395 45.4258C107.14 44.8749 107.012 44.224 107.012 43.473C107.012 42.7282 107.14 42.0819 107.395 41.5341C107.651 40.9863 108.006 40.5631 108.462 40.2646C108.917 39.966 109.444 39.8168 110.041 39.8168C110.502 39.8168 110.867 39.8937 111.135 40.0476C111.406 40.1984 111.612 40.3707 111.753 40.5646C111.898 40.7554 112.01 40.9124 112.09 41.0355H112.183V37.5455H113.272V47H112.22V45.9105H112.09C112.01 46.0398 111.896 46.2029 111.749 46.3999C111.601 46.5937 111.39 46.7676 111.116 46.9215C110.842 47.0723 110.478 47.1477 110.022 47.1477ZM110.17 46.169C110.607 46.169 110.976 46.0552 111.278 45.8274C111.579 45.5966 111.809 45.2781 111.966 44.8718C112.123 44.4625 112.201 43.9901 112.201 43.4545C112.201 42.9252 112.124 42.462 111.97 42.065C111.816 41.6649 111.589 41.354 111.287 41.1325C110.985 40.9078 110.613 40.7955 110.17 40.7955C109.708 40.7955 109.324 40.9139 109.016 41.1509C108.711 41.3848 108.482 41.7034 108.328 42.1065C108.177 42.5066 108.102 42.956 108.102 43.4545C108.102 43.9593 108.179 44.4179 108.333 44.8303C108.49 45.2396 108.72 45.5658 109.025 45.8089C109.333 46.049 109.714 46.169 110.17 46.169ZM118.353 47.1477C117.67 47.1477 117.08 46.9969 116.585 46.6953C116.092 46.3906 115.712 45.9659 115.444 45.4212C115.18 44.8733 115.047 44.2363 115.047 43.5099C115.047 42.7836 115.18 42.1435 115.444 41.5895C115.712 41.0324 116.085 40.5985 116.562 40.2876C117.042 39.9737 117.602 39.8168 118.242 39.8168C118.611 39.8168 118.976 39.8783 119.336 40.0014C119.696 40.1245 120.024 40.3246 120.319 40.6016C120.615 40.8755 120.85 41.2386 121.026 41.6911C121.201 42.1435 121.289 42.7005 121.289 43.3622V43.8239H115.823V42.8821H120.181C120.181 42.482 120.101 42.125 119.941 41.8111C119.784 41.4972 119.559 41.2494 119.267 41.0678C118.978 40.8862 118.636 40.7955 118.242 40.7955C117.808 40.7955 117.433 40.9032 117.116 41.1186C116.802 41.331 116.56 41.608 116.391 41.9496C116.222 42.2912 116.137 42.6574 116.137 43.0483V43.6761C116.137 44.2116 116.229 44.6656 116.414 45.038C116.602 45.4073 116.862 45.6889 117.194 45.8828C117.526 46.0736 117.913 46.169 118.353 46.169C118.639 46.169 118.898 46.129 119.128 46.049C119.362 45.9659 119.564 45.8428 119.733 45.6797C119.902 45.5135 120.033 45.3073 120.126 45.0611L121.178 45.3565C121.067 45.7135 120.881 46.0275 120.62 46.2983C120.358 46.5661 120.035 46.7753 119.65 46.9261C119.265 47.0739 118.833 47.1477 118.353 47.1477ZM125.585 47.1477C124.994 47.1477 124.472 46.9985 124.02 46.6999C123.567 46.3983 123.213 45.9736 122.958 45.4258C122.703 44.8749 122.575 44.224 122.575 43.473C122.575 42.7282 122.703 42.0819 122.958 41.5341C123.213 40.9863 123.569 40.5631 124.024 40.2646C124.48 39.966 125.006 39.8168 125.603 39.8168C126.065 39.8168 126.43 39.8937 126.697 40.0476C126.968 40.1984 127.174 40.3707 127.316 40.5646C127.461 40.7554 127.573 40.9124 127.653 41.0355H127.745V37.5455H128.835V47H127.782V45.9105H127.653C127.573 46.0398 127.459 46.2029 127.311 46.3999C127.164 46.5937 126.953 46.7676 126.679 46.9215C126.405 47.0723 126.04 47.1477 125.585 47.1477ZM125.733 46.169C126.17 46.169 126.539 46.0552 126.84 45.8274C127.142 45.5966 127.371 45.2781 127.528 44.8718C127.685 44.4625 127.764 43.9901 127.764 43.4545C127.764 42.9252 127.687 42.462 127.533 42.065C127.379 41.6649 127.151 41.354 126.85 41.1325C126.548 40.9078 126.176 40.7955 125.733 40.7955C125.271 40.7955 124.886 40.9139 124.578 41.1509C124.274 41.3848 124.044 41.7034 123.891 42.1065C123.74 42.5066 123.664 42.956 123.664 43.4545C123.664 43.9593 123.741 44.4179 123.895 44.8303C124.052 45.2396 124.283 45.5658 124.588 45.8089C124.895 46.049 125.277 46.169 125.733 46.169ZM137.237 47.1477C136.646 47.1477 136.125 46.9985 135.672 46.6999C135.22 46.3983 134.866 45.9736 134.61 45.4258C134.355 44.8749 134.227 44.224 134.227 43.473C134.227 42.7282 134.355 42.0819 134.61 41.5341C134.866 40.9863 135.221 40.5631 135.677 40.2646C136.132 39.966 136.659 39.8168 137.256 39.8168C137.717 39.8168 138.082 39.8937 138.35 40.0476C138.621 40.1984 138.827 40.3707 138.968 40.5646C139.113 40.7554 139.225 40.9124 139.305 41.0355H139.398V37.5455H140.487V47H139.435V45.9105H139.305C139.225 46.0398 139.112 46.2029 138.964 46.3999C138.816 46.5937 138.605 46.7676 138.331 46.9215C138.057 47.0723 137.693 47.1477 137.237 47.1477ZM137.385 46.169C137.822 46.169 138.191 46.0552 138.493 45.8274C138.795 45.5966 139.024 45.2781 139.181 44.8718C139.338 44.4625 139.416 43.9901 139.416 43.4545C139.416 42.9252 139.339 42.462 139.185 42.065C139.032 41.6649 138.804 41.354 138.502 41.1325C138.201 40.9078 137.828 40.7955 137.385 40.7955C136.923 40.7955 136.539 40.9139 136.231 41.1509C135.926 41.3848 135.697 41.7034 135.543 42.1065C135.392 42.5066 135.317 42.956 135.317 43.4545C135.317 43.9593 135.394 44.4179 135.548 44.8303C135.705 45.2396 135.935 45.5658 136.24 45.8089C136.548 46.049 136.929 46.169 137.385 46.169ZM144.682 47.1662C144.232 47.1662 143.824 47.0816 143.458 46.9123C143.092 46.7399 142.801 46.4922 142.586 46.169C142.37 45.8428 142.263 45.4489 142.263 44.9872C142.263 44.581 142.343 44.2517 142.503 43.9993C142.663 43.7438 142.877 43.5438 143.144 43.3991C143.412 43.2545 143.707 43.1468 144.031 43.076C144.357 43.0021 144.685 42.9437 145.014 42.9006C145.445 42.8452 145.794 42.8036 146.062 42.7759C146.333 42.7451 146.53 42.6944 146.653 42.6236C146.779 42.5528 146.842 42.4297 146.842 42.2543V42.2173C146.842 41.7618 146.717 41.4079 146.468 41.1555C146.222 40.9032 145.848 40.777 145.346 40.777C144.826 40.777 144.418 40.8909 144.123 41.1186C143.828 41.3464 143.62 41.5895 143.5 41.848L142.466 41.4787C142.65 41.0478 142.897 40.7124 143.204 40.4723C143.515 40.2292 143.854 40.0599 144.22 39.9645C144.589 39.866 144.952 39.8168 145.309 39.8168C145.537 39.8168 145.799 39.8445 146.094 39.8999C146.393 39.9522 146.681 40.0614 146.957 40.2276C147.238 40.3938 147.47 40.6446 147.655 40.9801C147.839 41.3156 147.932 41.7649 147.932 42.3281V47H146.842V46.0398H146.787C146.713 46.1937 146.59 46.3583 146.417 46.5337C146.245 46.7092 146.016 46.8584 145.73 46.9815C145.443 47.1046 145.094 47.1662 144.682 47.1662ZM144.848 46.1875C145.279 46.1875 145.642 46.1029 145.937 45.9336C146.236 45.7643 146.46 45.5458 146.611 45.2781C146.765 45.0103 146.842 44.7287 146.842 44.4332V43.4361C146.796 43.4915 146.694 43.5423 146.537 43.5884C146.384 43.6315 146.205 43.67 146.002 43.7038C145.802 43.7346 145.606 43.7623 145.416 43.7869C145.228 43.8085 145.076 43.8269 144.959 43.8423C144.675 43.8793 144.411 43.9393 144.165 44.0224C143.921 44.1024 143.724 44.224 143.574 44.3871C143.426 44.5471 143.352 44.7656 143.352 45.0426C143.352 45.4212 143.492 45.7074 143.772 45.9013C144.055 46.0921 144.414 46.1875 144.848 46.1875ZM152.965 39.9091V40.8324H149.29V39.9091H152.965ZM150.361 38.2102H151.451V44.9688C151.451 45.2765 151.495 45.5073 151.585 45.6612C151.677 45.812 151.794 45.9136 151.935 45.9659C152.08 46.0152 152.232 46.0398 152.392 46.0398C152.513 46.0398 152.611 46.0336 152.688 46.0213C152.765 46.0059 152.826 45.9936 152.873 45.9844L153.094 46.9631C153.02 46.9908 152.917 47.0185 152.785 47.0462C152.653 47.0769 152.485 47.0923 152.282 47.0923C151.974 47.0923 151.672 47.0262 151.377 46.8938C151.084 46.7615 150.841 46.5599 150.647 46.2891C150.457 46.0182 150.361 45.6766 150.361 45.2642V38.2102ZM156.651 47.1662C156.202 47.1662 155.794 47.0816 155.428 46.9123C155.062 46.7399 154.771 46.4922 154.556 46.169C154.34 45.8428 154.232 45.4489 154.232 44.9872C154.232 44.581 154.312 44.2517 154.472 43.9993C154.632 43.7438 154.846 43.5438 155.114 43.3991C155.382 43.2545 155.677 43.1468 156.001 43.076C156.327 43.0021 156.655 42.9437 156.984 42.9006C157.415 42.8452 157.764 42.8036 158.032 42.7759C158.303 42.7451 158.5 42.6944 158.623 42.6236C158.749 42.5528 158.812 42.4297 158.812 42.2543V42.2173C158.812 41.7618 158.687 41.4079 158.438 41.1555C158.192 40.9032 157.818 40.777 157.316 40.777C156.796 40.777 156.388 40.8909 156.093 41.1186C155.797 41.3464 155.59 41.5895 155.47 41.848L154.436 41.4787C154.62 41.0478 154.866 40.7124 155.174 40.4723C155.485 40.2292 155.824 40.0599 156.19 39.9645C156.559 39.866 156.922 39.8168 157.279 39.8168C157.507 39.8168 157.769 39.8445 158.064 39.8999C158.363 39.9522 158.65 40.0614 158.927 40.2276C159.207 40.3938 159.44 40.6446 159.624 40.9801C159.809 41.3156 159.901 41.7649 159.901 42.3281V47H158.812V46.0398H158.757C158.683 46.1937 158.56 46.3583 158.387 46.5337C158.215 46.7092 157.986 46.8584 157.699 46.9815C157.413 47.1046 157.064 47.1662 156.651 47.1662ZM156.818 46.1875C157.248 46.1875 157.612 46.1029 157.907 45.9336C158.206 45.7643 158.43 45.5458 158.581 45.2781C158.735 45.0103 158.812 44.7287 158.812 44.4332V43.4361C158.766 43.4915 158.664 43.5423 158.507 43.5884C158.353 43.6315 158.175 43.67 157.972 43.7038C157.772 43.7346 157.576 43.7623 157.385 43.7869C157.198 43.8085 157.045 43.8269 156.928 43.8423C156.645 43.8793 156.381 43.9393 156.134 44.0224C155.891 44.1024 155.694 44.224 155.543 44.3871C155.396 44.5471 155.322 44.7656 155.322 45.0426C155.322 45.4212 155.462 45.7074 155.742 45.9013C156.025 46.0921 156.384 46.1875 156.818 46.1875ZM167.555 47.1662C167.106 47.1662 166.698 47.0816 166.332 46.9123C165.965 46.7399 165.674 46.4922 165.459 46.169C165.244 45.8428 165.136 45.4489 165.136 44.9872C165.136 44.581 165.216 44.2517 165.376 43.9993C165.536 43.7438 165.75 43.5438 166.018 43.3991C166.285 43.2545 166.581 43.1468 166.904 43.076C167.23 43.0021 167.558 42.9437 167.887 42.9006C168.318 42.8452 168.667 42.8036 168.935 42.7759C169.206 42.7451 169.403 42.6944 169.526 42.6236C169.652 42.5528 169.715 42.4297 169.715 42.2543V42.2173C169.715 41.7618 169.591 41.4079 169.341 41.1555C169.095 40.9032 168.721 40.777 168.22 40.777C167.7 40.777 167.292 40.8909 166.996 41.1186C166.701 41.3464 166.493 41.5895 166.373 41.848L165.339 41.4787C165.524 41.0478 165.77 40.7124 166.078 40.4723C166.388 40.2292 166.727 40.0599 167.093 39.9645C167.463 39.866 167.826 39.8168 168.183 39.8168C168.41 39.8168 168.672 39.8445 168.968 39.8999C169.266 39.9522 169.554 40.0614 169.831 40.2276C170.111 40.3938 170.343 40.6446 170.528 40.9801C170.713 41.3156 170.805 41.7649 170.805 42.3281V47H169.715V46.0398H169.66C169.586 46.1937 169.463 46.3583 169.291 46.5337C169.118 46.7092 168.889 46.8584 168.603 46.9815C168.317 47.1046 167.967 47.1662 167.555 47.1662ZM167.721 46.1875C168.152 46.1875 168.515 46.1029 168.811 45.9336C169.109 45.7643 169.334 45.5458 169.485 45.2781C169.638 45.0103 169.715 44.7287 169.715 44.4332V43.4361C169.669 43.4915 169.568 43.5423 169.411 43.5884C169.257 43.6315 169.078 43.67 168.875 43.7038C168.675 43.7346 168.48 43.7623 168.289 43.7869C168.101 43.8085 167.949 43.8269 167.832 43.8423C167.549 43.8793 167.284 43.9393 167.038 44.0224C166.795 44.1024 166.598 44.224 166.447 44.3871C166.299 44.5471 166.225 44.7656 166.225 45.0426C166.225 45.4212 166.365 45.7074 166.645 45.9013C166.929 46.0921 167.287 46.1875 167.721 46.1875ZM173.844 42.7344V47H172.754V39.9091H173.807V41.017H173.899C174.066 40.657 174.318 40.3677 174.656 40.1491C174.995 39.9276 175.432 39.8168 175.967 39.8168C176.448 39.8168 176.868 39.9152 177.228 40.1122C177.588 40.3061 177.868 40.6016 178.068 40.9986C178.268 41.3925 178.368 41.8911 178.368 42.4943V47H177.279V42.5682C177.279 42.0111 177.134 41.5772 176.845 41.2663C176.555 40.9524 176.158 40.7955 175.654 40.7955C175.306 40.7955 174.995 40.8709 174.721 41.0217C174.45 41.1725 174.236 41.3925 174.079 41.6818C173.922 41.9711 173.844 42.322 173.844 42.7344ZM182.997 47.1477C182.407 47.1477 181.885 46.9985 181.432 46.6999C180.98 46.3983 180.626 45.9736 180.371 45.4258C180.115 44.8749 179.988 44.224 179.988 43.473C179.988 42.7282 180.115 42.0819 180.371 41.5341C180.626 40.9863 180.982 40.5631 181.437 40.2646C181.893 39.966 182.419 39.8168 183.016 39.8168C183.478 39.8168 183.842 39.8937 184.11 40.0476C184.381 40.1984 184.587 40.3707 184.729 40.5646C184.873 40.7554 184.986 40.9124 185.066 41.0355H185.158V37.5455H186.247V47H185.195V45.9105H185.066C184.986 46.0398 184.872 46.2029 184.724 46.3999C184.576 46.5937 184.365 46.7676 184.092 46.9215C183.818 47.0723 183.453 47.1477 182.997 47.1477ZM183.145 46.169C183.582 46.169 183.952 46.0552 184.253 45.8274C184.555 45.5966 184.784 45.2781 184.941 44.8718C185.098 44.4625 185.176 43.9901 185.176 43.4545C185.176 42.9252 185.1 42.462 184.946 42.065C184.792 41.6649 184.564 41.354 184.262 41.1325C183.961 40.9078 183.588 40.7955 183.145 40.7955C182.684 40.7955 182.299 40.9139 181.991 41.1509C181.686 41.3848 181.457 41.7034 181.303 42.1065C181.152 42.5066 181.077 42.956 181.077 43.4545C181.077 43.9593 181.154 44.4179 181.308 44.8303C181.465 45.2396 181.696 45.5658 182 45.8089C182.308 46.049 182.69 46.169 183.145 46.169ZM195.13 39.9091V40.8324H191.308V39.9091H195.13ZM192.452 47V38.9304C192.452 38.5241 192.548 38.1856 192.739 37.9148C192.93 37.6439 193.177 37.4408 193.482 37.3054C193.787 37.17 194.108 37.1023 194.447 37.1023C194.715 37.1023 194.933 37.1238 195.102 37.1669C195.272 37.21 195.398 37.25 195.481 37.2869L195.167 38.2287C195.112 38.2102 195.035 38.1871 194.936 38.1594C194.841 38.1317 194.715 38.1179 194.558 38.1179C194.198 38.1179 193.937 38.2087 193.777 38.3903C193.62 38.5719 193.542 38.8381 193.542 39.1889V47H192.452ZM196.618 47V39.9091H197.707V47H196.618ZM197.172 38.7273C196.96 38.7273 196.776 38.6549 196.623 38.5103C196.472 38.3656 196.396 38.1918 196.396 37.9886C196.396 37.7855 196.472 37.6116 196.623 37.467C196.776 37.3223 196.96 37.25 197.172 37.25C197.384 37.25 197.566 37.3223 197.717 37.467C197.871 37.6116 197.947 37.7855 197.947 37.9886C197.947 38.1918 197.871 38.3656 197.717 38.5103C197.566 38.6549 197.384 38.7273 197.172 38.7273ZM202.526 49.8068C202 49.8068 201.547 49.7391 201.169 49.6037C200.79 49.4714 200.475 49.2959 200.223 49.0774C199.973 48.862 199.775 48.6312 199.627 48.3849L200.495 47.7756C200.593 47.9048 200.718 48.0526 200.869 48.2188C201.02 48.388 201.226 48.5342 201.487 48.6573C201.752 48.7835 202.098 48.8466 202.526 48.8466C203.099 48.8466 203.571 48.7081 203.943 48.4311C204.316 48.1541 204.502 47.7202 204.502 47.1293V45.6889H204.41C204.33 45.8182 204.216 45.9782 204.068 46.169C203.923 46.3568 203.714 46.5245 203.44 46.6722C203.169 46.8169 202.803 46.8892 202.341 46.8892C201.769 46.8892 201.255 46.7538 200.8 46.483C200.347 46.2121 199.989 45.8182 199.724 45.3011C199.462 44.7841 199.332 44.1562 199.332 43.4176C199.332 42.6913 199.459 42.0588 199.715 41.5202C199.97 40.9786 200.326 40.56 200.781 40.2646C201.237 39.966 201.763 39.8168 202.36 39.8168C202.822 39.8168 203.188 39.8937 203.459 40.0476C203.733 40.1984 203.942 40.3707 204.086 40.5646C204.234 40.7554 204.348 40.9124 204.428 41.0355H204.539V39.9091H205.591V47.2031C205.591 47.8125 205.453 48.308 205.176 48.6896C204.902 49.0743 204.533 49.3559 204.068 49.5344C203.606 49.716 203.092 49.8068 202.526 49.8068ZM202.489 45.9105C202.926 45.9105 203.296 45.8105 203.597 45.6104C203.899 45.4104 204.128 45.1226 204.285 44.7472C204.442 44.3717 204.52 43.9223 204.52 43.3991C204.52 42.8883 204.444 42.4374 204.29 42.0465C204.136 41.6557 203.908 41.3494 203.606 41.1278C203.305 40.9062 202.932 40.7955 202.489 40.7955C202.028 40.7955 201.643 40.9124 201.335 41.1463C201.03 41.3802 200.801 41.6941 200.647 42.0881C200.496 42.482 200.421 42.919 200.421 43.3991C200.421 43.8916 200.498 44.3271 200.652 44.7056C200.809 45.0811 201.04 45.3765 201.344 45.592C201.652 45.8043 202.034 45.9105 202.489 45.9105ZM212.016 44.1009V39.9091H213.105V47H212.016V45.7997H211.942C211.775 46.1598 211.517 46.466 211.166 46.7184C210.815 46.9677 210.372 47.0923 209.837 47.0923C209.393 47.0923 208.999 46.9954 208.655 46.8015C208.31 46.6045 208.039 46.3091 207.842 45.9151C207.645 45.5181 207.547 45.018 207.547 44.4148V39.9091H208.636V44.3409C208.636 44.858 208.781 45.2704 209.07 45.5781C209.363 45.8859 209.735 46.0398 210.187 46.0398C210.458 46.0398 210.734 45.9705 211.014 45.832C211.297 45.6935 211.534 45.4812 211.725 45.195C211.919 44.9087 212.016 44.544 212.016 44.1009ZM215.061 47V39.9091H216.114V40.9801H216.188C216.317 40.6293 216.551 40.3446 216.89 40.1261C217.228 39.9076 217.61 39.7983 218.035 39.7983C218.115 39.7983 218.215 39.7998 218.335 39.8029C218.455 39.806 218.545 39.8106 218.607 39.8168V40.9247C218.57 40.9155 218.485 40.9016 218.353 40.8832C218.224 40.8616 218.087 40.8509 217.942 40.8509C217.597 40.8509 217.29 40.9232 217.019 41.0678C216.751 41.2094 216.539 41.4064 216.382 41.6587C216.228 41.908 216.151 42.1927 216.151 42.5128V47H215.061ZM222.617 47.1477C221.933 47.1477 221.344 46.9969 220.848 46.6953C220.356 46.3906 219.976 45.9659 219.708 45.4212C219.444 44.8733 219.311 44.2363 219.311 43.5099C219.311 42.7836 219.444 42.1435 219.708 41.5895C219.976 41.0324 220.348 40.5985 220.825 40.2876C221.306 39.9737 221.866 39.8168 222.506 39.8168C222.875 39.8168 223.24 39.8783 223.6 40.0014C223.96 40.1245 224.288 40.3246 224.583 40.6016C224.879 40.8755 225.114 41.2386 225.29 41.6911C225.465 42.1435 225.553 42.7005 225.553 43.3622V43.8239H220.087V42.8821H224.445C224.445 42.482 224.365 42.125 224.205 41.8111C224.048 41.4972 223.823 41.2494 223.531 41.0678C223.241 40.8862 222.9 40.7955 222.506 40.7955C222.072 40.7955 221.696 40.9032 221.379 41.1186C221.065 41.331 220.824 41.608 220.655 41.9496C220.485 42.2912 220.401 42.6574 220.401 43.0483V43.6761C220.401 44.2116 220.493 44.6656 220.678 45.038C220.865 45.4073 221.125 45.6889 221.458 45.8828C221.79 46.0736 222.176 46.169 222.617 46.169C222.903 46.169 223.161 46.129 223.392 46.049C223.626 45.9659 223.828 45.8428 223.997 45.6797C224.166 45.5135 224.297 45.3073 224.389 45.0611L225.442 45.3565C225.331 45.7135 225.145 46.0275 224.883 46.2983C224.622 46.5661 224.299 46.7753 223.914 46.9261C223.529 47.0739 223.097 47.1477 222.617 47.1477ZM232.194 41.4972L231.215 41.7741C231.153 41.611 231.063 41.4525 230.943 41.2987C230.826 41.1417 230.666 41.0124 230.463 40.9109C230.259 40.8093 229.999 40.7585 229.682 40.7585C229.248 40.7585 228.887 40.8585 228.597 41.0586C228.311 41.2556 228.168 41.5064 228.168 41.8111C228.168 42.0819 228.267 42.2958 228.464 42.4528C228.661 42.6097 228.968 42.7405 229.387 42.8452L230.439 43.1037C231.073 43.2576 231.546 43.493 231.857 43.81C232.168 44.1239 232.323 44.5286 232.323 45.0241C232.323 45.4304 232.206 45.7936 231.972 46.1136C231.741 46.4337 231.418 46.6861 231.003 46.8707C230.587 47.0554 230.104 47.1477 229.553 47.1477C228.83 47.1477 228.231 46.9908 227.757 46.6768C227.283 46.3629 226.983 45.9044 226.857 45.3011L227.891 45.0426C227.99 45.4242 228.176 45.7105 228.45 45.9013C228.727 46.0921 229.088 46.1875 229.535 46.1875C230.042 46.1875 230.446 46.0798 230.744 45.8643C231.046 45.6458 231.197 45.3842 231.197 45.0795C231.197 44.8333 231.11 44.6271 230.938 44.4609C230.766 44.2917 230.501 44.1655 230.144 44.0824L228.962 43.8054C228.313 43.6515 227.836 43.413 227.531 43.0898C227.229 42.7636 227.079 42.3558 227.079 41.8665C227.079 41.4664 227.191 41.1125 227.416 40.8047C227.643 40.4969 227.953 40.2553 228.344 40.0799C228.738 39.9045 229.184 39.8168 229.682 39.8168C230.384 39.8168 230.935 39.9706 231.335 40.2784C231.738 40.5862 232.024 40.9924 232.194 41.4972ZM237.541 47V39.9091H238.594V41.017H238.686C238.834 40.6385 239.072 40.3446 239.402 40.1353C239.731 39.9229 240.126 39.8168 240.588 39.8168C241.056 39.8168 241.445 39.9229 241.756 40.1353C242.07 40.3446 242.315 40.6385 242.49 41.017H242.564C242.746 40.6508 243.018 40.36 243.381 40.1445C243.744 39.926 244.18 39.8168 244.688 39.8168C245.322 39.8168 245.84 40.0153 246.243 40.4123C246.646 40.8062 246.848 41.4202 246.848 42.2543V47H245.759V42.2543C245.759 41.7311 245.615 41.3571 245.329 41.1325C245.043 40.9078 244.706 40.7955 244.318 40.7955C243.82 40.7955 243.433 40.9463 243.159 41.2479C242.886 41.5464 242.749 41.925 242.749 42.3835V47H241.641V42.1435C241.641 41.7403 241.51 41.4156 241.248 41.1694C240.987 40.9201 240.65 40.7955 240.237 40.7955C239.954 40.7955 239.689 40.8709 239.443 41.0217C239.2 41.1725 239.003 41.3817 238.852 41.6495C238.705 41.9142 238.631 42.2204 238.631 42.5682V47H237.541ZM250.888 47.1662C250.438 47.1662 250.031 47.0816 249.664 46.9123C249.298 46.7399 249.007 46.4922 248.792 46.169C248.576 45.8428 248.469 45.4489 248.469 44.9872C248.469 44.581 248.549 44.2517 248.709 43.9993C248.869 43.7438 249.083 43.5438 249.35 43.3991C249.618 43.2545 249.914 43.1468 250.237 43.076C250.563 43.0021 250.891 42.9437 251.22 42.9006C251.651 42.8452 252 42.8036 252.268 42.7759C252.539 42.7451 252.736 42.6944 252.859 42.6236C252.985 42.5528 253.048 42.4297 253.048 42.2543V42.2173C253.048 41.7618 252.924 41.4079 252.674 41.1555C252.428 40.9032 252.054 40.777 251.552 40.777C251.032 40.777 250.625 40.8909 250.329 41.1186C250.034 41.3464 249.826 41.5895 249.706 41.848L248.672 41.4787C248.856 41.0478 249.103 40.7124 249.41 40.4723C249.721 40.2292 250.06 40.0599 250.426 39.9645C250.795 39.866 251.159 39.8168 251.516 39.8168C251.743 39.8168 252.005 39.8445 252.3 39.8999C252.599 39.9522 252.887 40.0614 253.164 40.2276C253.444 40.3938 253.676 40.6446 253.861 40.9801C254.045 41.3156 254.138 41.7649 254.138 42.3281V47H253.048V46.0398H252.993C252.919 46.1937 252.796 46.3583 252.624 46.5337C252.451 46.7092 252.222 46.8584 251.936 46.9815C251.649 47.1046 251.3 47.1662 250.888 47.1662ZM251.054 46.1875C251.485 46.1875 251.848 46.1029 252.143 45.9336C252.442 45.7643 252.667 45.5458 252.817 45.2781C252.971 45.0103 253.048 44.7287 253.048 44.4332V43.4361C253.002 43.4915 252.901 43.5423 252.744 43.5884C252.59 43.6315 252.411 43.67 252.208 43.7038C252.008 43.7346 251.813 43.7623 251.622 43.7869C251.434 43.8085 251.282 43.8269 251.165 43.8423C250.882 43.8793 250.617 43.9393 250.371 44.0224C250.128 44.1024 249.931 44.224 249.78 44.3871C249.632 44.5471 249.558 44.7656 249.558 45.0426C249.558 45.4212 249.698 45.7074 249.978 45.9013C250.261 46.0921 250.62 46.1875 251.054 46.1875ZM256.535 49.6591C256.35 49.6591 256.186 49.6437 256.041 49.6129C255.896 49.5852 255.796 49.5575 255.741 49.5298L256.018 48.5696C256.283 48.6373 256.517 48.6619 256.72 48.6435C256.923 48.625 257.103 48.5342 257.26 48.3711C257.42 48.2111 257.566 47.951 257.698 47.5909L257.902 47.0369L255.279 39.9091H256.461L258.419 45.5597H258.492L260.45 39.9091H261.632L258.622 48.0341C258.486 48.4003 258.319 48.7035 258.119 48.9435C257.918 49.1867 257.686 49.3667 257.421 49.4837C257.16 49.6006 256.864 49.6591 256.535 49.6591ZM269.325 47.1477C268.735 47.1477 268.213 46.9985 267.761 46.6999C267.308 46.3983 266.954 45.9736 266.699 45.4258C266.443 44.8749 266.316 44.224 266.316 43.473C266.316 42.7282 266.443 42.0819 266.699 41.5341C266.954 40.9863 267.31 40.5631 267.765 40.2646C268.221 39.966 268.747 39.8168 269.344 39.8168C269.806 39.8168 270.17 39.8937 270.438 40.0476C270.709 40.1984 270.915 40.3707 271.057 40.5646C271.201 40.7554 271.314 40.9124 271.394 41.0355H271.486V37.5455H272.575V47H271.523V45.9105H271.394C271.314 46.0398 271.2 46.2029 271.052 46.3999C270.904 46.5937 270.694 46.7676 270.42 46.9215C270.146 47.0723 269.781 47.1477 269.325 47.1477ZM269.473 46.169C269.91 46.169 270.28 46.0552 270.581 45.8274C270.883 45.5966 271.112 45.2781 271.269 44.8718C271.426 44.4625 271.504 43.9901 271.504 43.4545C271.504 42.9252 271.428 42.462 271.274 42.065C271.12 41.6649 270.892 41.354 270.59 41.1325C270.289 40.9078 269.916 40.7955 269.473 40.7955C269.012 40.7955 268.627 40.9139 268.319 41.1509C268.014 41.3848 267.785 41.7034 267.631 42.1065C267.48 42.5066 267.405 42.956 267.405 43.4545C267.405 43.9593 267.482 44.4179 267.636 44.8303C267.793 45.2396 268.024 45.5658 268.328 45.8089C268.636 46.049 269.018 46.169 269.473 46.169ZM274.683 47V39.9091H275.773V47H274.683ZM275.237 38.7273C275.025 38.7273 274.842 38.6549 274.688 38.5103C274.537 38.3656 274.462 38.1918 274.462 37.9886C274.462 37.7855 274.537 37.6116 274.688 37.467C274.842 37.3223 275.025 37.25 275.237 37.25C275.449 37.25 275.631 37.3223 275.782 37.467C275.936 37.6116 276.013 37.7855 276.013 37.9886C276.013 38.1918 275.936 38.3656 275.782 38.5103C275.631 38.6549 275.449 38.7273 275.237 38.7273ZM280.887 39.9091V40.8324H277.064V39.9091H280.887ZM278.209 47V38.9304C278.209 38.5241 278.305 38.1856 278.495 37.9148C278.686 37.6439 278.934 37.4408 279.239 37.3054C279.543 37.17 279.865 37.1023 280.204 37.1023C280.471 37.1023 280.69 37.1238 280.859 37.1669C281.028 37.21 281.155 37.25 281.238 37.2869L280.924 38.2287C280.868 38.2102 280.791 38.1871 280.693 38.1594C280.597 38.1317 280.471 38.1179 280.314 38.1179C279.954 38.1179 279.694 38.2087 279.534 38.3903C279.377 38.5719 279.299 38.8381 279.299 39.1889V47H278.209ZM285.532 39.9091V40.8324H281.71V39.9091H285.532ZM282.855 47V38.9304C282.855 38.5241 282.95 38.1856 283.141 37.9148C283.332 37.6439 283.58 37.4408 283.884 37.3054C284.189 37.17 284.511 37.1023 284.849 37.1023C285.117 37.1023 285.335 37.1238 285.505 37.1669C285.674 37.21 285.8 37.25 285.883 37.2869L285.569 38.2287C285.514 38.2102 285.437 38.1871 285.338 38.1594C285.243 38.1317 285.117 38.1179 284.96 38.1179C284.6 38.1179 284.34 38.2087 284.18 38.3903C284.023 38.5719 283.944 38.8381 283.944 39.1889V47H282.855ZM289.701 47.1477C289.018 47.1477 288.429 46.9969 287.933 46.6953C287.441 46.3906 287.061 45.9659 286.793 45.4212C286.528 44.8733 286.396 44.2363 286.396 43.5099C286.396 42.7836 286.528 42.1435 286.793 41.5895C287.061 41.0324 287.433 40.5985 287.91 40.2876C288.39 39.9737 288.95 39.8168 289.59 39.8168C289.96 39.8168 290.324 39.8783 290.685 40.0014C291.045 40.1245 291.372 40.3246 291.668 40.6016C291.963 40.8755 292.199 41.2386 292.374 41.6911C292.55 42.1435 292.637 42.7005 292.637 43.3622V43.8239H287.171V42.8821H291.529C291.529 42.482 291.449 42.125 291.289 41.8111C291.132 41.4972 290.908 41.2494 290.615 41.0678C290.326 40.8862 289.984 40.7955 289.59 40.7955C289.157 40.7955 288.781 40.9032 288.464 41.1186C288.15 41.331 287.909 41.608 287.739 41.9496C287.57 42.2912 287.485 42.6574 287.485 43.0483V43.6761C287.485 44.2116 287.578 44.6656 287.762 45.038C287.95 45.4073 288.21 45.6889 288.543 45.8828C288.875 46.0736 289.261 46.169 289.701 46.169C289.987 46.169 290.246 46.129 290.477 46.049C290.711 45.9659 290.912 45.8428 291.082 45.6797C291.251 45.5135 291.382 45.3073 291.474 45.0611L292.527 45.3565C292.416 45.7135 292.23 46.0275 291.968 46.2983C291.706 46.5661 291.383 46.7753 290.998 46.9261C290.614 47.0739 290.181 47.1477 289.701 47.1477ZM294.256 47V39.9091H295.308V40.9801H295.382C295.511 40.6293 295.745 40.3446 296.084 40.1261C296.422 39.9076 296.804 39.7983 297.229 39.7983C297.309 39.7983 297.409 39.7998 297.529 39.8029C297.649 39.806 297.74 39.8106 297.801 39.8168V40.9247C297.764 40.9155 297.68 40.9016 297.547 40.8832C297.418 40.8616 297.281 40.8509 297.136 40.8509C296.792 40.8509 296.484 40.9232 296.213 41.0678C295.945 41.2094 295.733 41.4064 295.576 41.6587C295.422 41.908 295.345 42.1927 295.345 42.5128V47H294.256ZM302.671 47V39.9091H303.76V47H302.671ZM303.225 38.7273C303.012 38.7273 302.829 38.6549 302.675 38.5103C302.525 38.3656 302.449 38.1918 302.449 37.9886C302.449 37.7855 302.525 37.6116 302.675 37.467C302.829 37.3223 303.012 37.25 303.225 37.25C303.437 37.25 303.619 37.3223 303.77 37.467C303.923 37.6116 304 37.7855 304 37.9886C304 38.1918 303.923 38.3656 303.77 38.5103C303.619 38.6549 303.437 38.7273 303.225 38.7273ZM306.806 42.7344V47H305.717V39.9091H306.769V41.017H306.862C307.028 40.657 307.28 40.3677 307.619 40.1491C307.957 39.9276 308.394 39.8168 308.93 39.8168C309.41 39.8168 309.83 39.9152 310.19 40.1122C310.55 40.3061 310.83 40.6016 311.03 40.9986C311.23 41.3925 311.33 41.8911 311.33 42.4943V47H310.241V42.5682C310.241 42.0111 310.096 41.5772 309.807 41.2663C309.518 40.9524 309.121 40.7955 308.616 40.7955C308.268 40.7955 307.957 40.8709 307.683 41.0217C307.413 41.1725 307.199 41.3925 307.042 41.6818C306.885 41.9711 306.806 42.322 306.806 42.7344ZM319.983 39.9091V40.8324H316.309V39.9091H319.983ZM317.38 38.2102H318.469V44.9688C318.469 45.2765 318.514 45.5073 318.603 45.6612C318.695 45.812 318.812 45.9136 318.954 45.9659C319.099 46.0152 319.251 46.0398 319.411 46.0398C319.531 46.0398 319.629 46.0336 319.706 46.0213C319.783 46.0059 319.845 45.9936 319.891 45.9844L320.113 46.9631C320.039 46.9908 319.936 47.0185 319.803 47.0462C319.671 47.0769 319.503 47.0923 319.3 47.0923C318.992 47.0923 318.691 47.0262 318.395 46.8938C318.103 46.7615 317.86 46.5599 317.666 46.2891C317.475 46.0182 317.38 45.6766 317.38 45.2642V38.2102ZM322.825 42.7344V47H321.736V37.5455H322.825V41.017H322.917C323.084 40.6508 323.333 40.36 323.665 40.1445C324.001 39.926 324.447 39.8168 325.004 39.8168C325.487 39.8168 325.91 39.9137 326.274 40.1076C326.637 40.2984 326.918 40.5923 327.118 40.9893C327.321 41.3833 327.423 41.8849 327.423 42.4943V47H326.334V42.5682C326.334 42.005 326.187 41.5695 325.895 41.2617C325.606 40.9509 325.204 40.7955 324.69 40.7955C324.333 40.7955 324.013 40.8709 323.73 41.0217C323.45 41.1725 323.228 41.3925 323.065 41.6818C322.905 41.9711 322.825 42.322 322.825 42.7344ZM332.35 47.1477C331.667 47.1477 331.078 46.9969 330.582 46.6953C330.09 46.3906 329.71 45.9659 329.442 45.4212C329.177 44.8733 329.045 44.2363 329.045 43.5099C329.045 42.7836 329.177 42.1435 329.442 41.5895C329.71 41.0324 330.082 40.5985 330.559 40.2876C331.039 39.9737 331.599 39.8168 332.239 39.8168C332.609 39.8168 332.973 39.8783 333.334 40.0014C333.694 40.1245 334.021 40.3246 334.317 40.6016C334.612 40.8755 334.848 41.2386 335.023 41.6911C335.199 42.1435 335.286 42.7005 335.286 43.3622V43.8239H329.82V42.8821H334.178C334.178 42.482 334.098 42.125 333.938 41.8111C333.781 41.4972 333.557 41.2494 333.264 41.0678C332.975 40.8862 332.633 40.7955 332.239 40.7955C331.805 40.7955 331.43 40.9032 331.113 41.1186C330.799 41.331 330.557 41.608 330.388 41.9496C330.219 42.2912 330.134 42.6574 330.134 43.0483V43.6761C330.134 44.2116 330.227 44.6656 330.411 45.038C330.599 45.4073 330.859 45.6889 331.191 45.8828C331.524 46.0736 331.91 46.169 332.35 46.169C332.636 46.169 332.895 46.129 333.126 46.049C333.36 45.9659 333.561 45.8428 333.731 45.6797C333.9 45.5135 334.031 45.3073 334.123 45.0611L335.175 45.3565C335.065 45.7135 334.878 46.0275 334.617 46.2983C334.355 46.5661 334.032 46.7753 333.647 46.9261C333.263 47.0739 332.83 47.1477 332.35 47.1477ZM343.68 39.9091V40.8324H339.857V39.9091H343.68ZM341.002 47V38.9304C341.002 38.5241 341.097 38.1856 341.288 37.9148C341.479 37.6439 341.727 37.4408 342.031 37.3054C342.336 37.17 342.658 37.1023 342.996 37.1023C343.264 37.1023 343.483 37.1238 343.652 37.1669C343.821 37.21 343.947 37.25 344.03 37.2869L343.716 38.2287C343.661 38.2102 343.584 38.1871 343.486 38.1594C343.39 38.1317 343.264 38.1179 343.107 38.1179C342.747 38.1179 342.487 38.2087 342.327 38.3903C342.17 38.5719 342.091 38.8381 342.091 39.1889V47H341.002ZM349.636 44.1009V39.9091H350.726V47H349.636V45.7997H349.562C349.396 46.1598 349.138 46.466 348.787 46.7184C348.436 46.9677 347.993 47.0923 347.457 47.0923C347.014 47.0923 346.62 46.9954 346.275 46.8015C345.931 46.6045 345.66 46.3091 345.463 45.9151C345.266 45.5181 345.167 45.018 345.167 44.4148V39.9091H346.257V44.3409C346.257 44.858 346.402 45.2704 346.691 45.5781C346.983 45.8859 347.356 46.0398 347.808 46.0398C348.079 46.0398 348.354 45.9705 348.634 45.832C348.918 45.6935 349.155 45.4812 349.345 45.195C349.539 44.9087 349.636 44.544 349.636 44.1009ZM355.766 39.9091V40.8324H352.091V39.9091H355.766ZM353.162 38.2102H354.252V44.9688C354.252 45.2765 354.296 45.5073 354.386 45.6612C354.478 45.812 354.595 45.9136 354.736 45.9659C354.881 46.0152 355.033 46.0398 355.193 46.0398C355.314 46.0398 355.412 46.0336 355.489 46.0213C355.566 46.0059 355.627 45.9936 355.674 45.9844L355.895 46.9631C355.821 46.9908 355.718 47.0185 355.586 47.0462C355.454 47.0769 355.286 47.0923 355.083 47.0923C354.775 47.0923 354.473 47.0262 354.178 46.8938C353.885 46.7615 353.642 46.5599 353.448 46.2891C353.258 46.0182 353.162 45.6766 353.162 45.2642V38.2102ZM361.835 44.1009V39.9091H362.924V47H361.835V45.7997H361.761C361.594 46.1598 361.336 46.466 360.985 46.7184C360.634 46.9677 360.191 47.0923 359.656 47.0923C359.212 47.0923 358.818 46.9954 358.474 46.8015C358.129 46.6045 357.858 46.3091 357.661 45.9151C357.464 45.5181 357.366 45.018 357.366 44.4148V39.9091H358.455V44.3409C358.455 44.858 358.6 45.2704 358.889 45.5781C359.182 45.8859 359.554 46.0398 360.006 46.0398C360.277 46.0398 360.553 45.9705 360.833 45.832C361.116 45.6935 361.353 45.4812 361.544 45.195C361.738 44.9087 361.835 44.544 361.835 44.1009ZM364.88 47V39.9091H365.933V40.9801H366.007C366.136 40.6293 366.37 40.3446 366.709 40.1261C367.047 39.9076 367.429 39.7983 367.854 39.7983C367.934 39.7983 368.034 39.7998 368.154 39.8029C368.274 39.806 368.364 39.8106 368.426 39.8168V40.9247C368.389 40.9155 368.304 40.9016 368.172 40.8832C368.043 40.8616 367.906 40.8509 367.761 40.8509C367.416 40.8509 367.109 40.9232 366.838 41.0678C366.57 41.2094 366.358 41.4064 366.201 41.6587C366.047 41.908 365.97 42.1927 365.97 42.5128V47H364.88ZM372.436 47.1477C371.752 47.1477 371.163 46.9969 370.667 46.6953C370.175 46.3906 369.795 45.9659 369.527 45.4212C369.263 44.8733 369.13 44.2363 369.13 43.5099C369.13 42.7836 369.263 42.1435 369.527 41.5895C369.795 41.0324 370.167 40.5985 370.644 40.2876C371.125 39.9737 371.685 39.8168 372.325 39.8168C372.694 39.8168 373.059 39.8783 373.419 40.0014C373.779 40.1245 374.107 40.3246 374.402 40.6016C374.698 40.8755 374.933 41.2386 375.109 41.6911C375.284 42.1435 375.372 42.7005 375.372 43.3622V43.8239H369.906V42.8821H374.264C374.264 42.482 374.184 42.125 374.024 41.8111C373.867 41.4972 373.642 41.2494 373.35 41.0678C373.06 40.8862 372.719 40.7955 372.325 40.7955C371.891 40.7955 371.515 40.9032 371.198 41.1186C370.884 41.331 370.643 41.608 370.474 41.9496C370.304 42.2912 370.22 42.6574 370.22 43.0483V43.6761C370.22 44.2116 370.312 44.6656 370.497 45.038C370.684 45.4073 370.944 45.6889 371.277 45.8828C371.609 46.0736 371.995 46.169 372.436 46.169C372.722 46.169 372.98 46.129 373.211 46.049C373.445 45.9659 373.647 45.8428 373.816 45.6797C373.985 45.5135 374.116 45.3073 374.208 45.0611L375.261 45.3565C375.15 45.7135 374.964 46.0275 374.702 46.2983C374.441 46.5661 374.118 46.7753 373.733 46.9261C373.348 47.0739 372.916 47.1477 372.436 47.1477ZM377.784 47.0739C377.556 47.0739 377.361 46.9923 377.198 46.8292C377.035 46.6661 376.953 46.4706 376.953 46.2429C376.953 46.0152 377.035 45.8197 377.198 45.6566C377.361 45.4935 377.556 45.4119 377.784 45.4119C378.012 45.4119 378.207 45.4935 378.37 45.6566C378.533 45.8197 378.615 46.0152 378.615 46.2429C378.615 46.3937 378.577 46.5322 378.5 46.6584C378.426 46.7846 378.326 46.8861 378.2 46.9631C378.076 47.0369 377.938 47.0739 377.784 47.0739Z";
-
-const LEFT_ALIGNED_TEXT =
-  "M14.8622 27L12.277 17.5455H13.4403L15.4162 25.2457H15.5085L17.5213 17.5455H18.8139L20.8267 25.2457H20.919L22.8949 17.5455H24.0582L21.473 27H20.2912L18.2045 19.4659H18.1307L16.044 27H14.8622ZM27.6109 27.1477C26.9276 27.1477 26.3382 26.9969 25.8427 26.6953C25.3503 26.3906 24.9702 25.9659 24.7025 25.4212C24.4378 24.8733 24.3055 24.2363 24.3055 23.5099C24.3055 22.7836 24.4378 22.1435 24.7025 21.5895C24.9702 21.0324 25.3426 20.5985 25.8197 20.2876C26.2998 19.9737 26.8599 19.8168 27.5001 19.8168C27.8694 19.8168 28.2341 19.8783 28.5942 20.0014C28.9543 20.1245 29.282 20.3246 29.5775 20.6016C29.8729 20.8755 30.1084 21.2386 30.2838 21.6911C30.4592 22.1435 30.5469 22.7005 30.5469 23.3622V23.8239H25.081V22.8821H29.439C29.439 22.482 29.359 22.125 29.1989 21.8111C29.042 21.4972 28.8173 21.2494 28.5249 21.0678C28.2356 20.8862 27.894 20.7955 27.5001 20.7955C27.0661 20.7955 26.6906 20.9032 26.3736 21.1186C26.0597 21.331 25.8181 21.608 25.6489 21.9496C25.4796 22.2912 25.3949 22.6574 25.3949 23.0483V23.6761C25.3949 24.2116 25.4873 24.6656 25.6719 25.038C25.8597 25.4073 26.1197 25.6889 26.4521 25.8828C26.7845 26.0736 27.1708 26.169 27.6109 26.169C27.8971 26.169 28.1556 26.129 28.3864 26.049C28.6203 25.9659 28.8219 25.8428 28.9912 25.6797C29.1605 25.5135 29.2913 25.3073 29.3836 25.0611L30.4361 25.3565C30.3253 25.7135 30.1391 26.0275 29.8775 26.2983C29.6159 26.5661 29.2928 26.7753 28.9081 26.9261C28.5234 27.0739 28.091 27.1477 27.6109 27.1477ZM35.7825 27V19.9091H36.8351V20.9801H36.9089C37.0382 20.6293 37.2721 20.3446 37.6106 20.1261C37.9492 19.9076 38.3308 19.7983 38.7555 19.7983C38.8355 19.7983 38.9356 19.7998 39.0556 19.8029C39.1756 19.806 39.2664 19.8106 39.328 19.8168V20.9247C39.291 20.9155 39.2064 20.9016 39.0741 20.8832C38.9448 20.8616 38.8078 20.8509 38.6632 20.8509C38.3185 20.8509 38.0107 20.9232 37.7399 21.0678C37.4721 21.2094 37.2598 21.4064 37.1028 21.6587C36.9489 21.908 36.872 22.1927 36.872 22.5128V27H35.7825ZM43.3376 27.1477C42.6544 27.1477 42.065 26.9969 41.5695 26.6953C41.0771 26.3906 40.697 25.9659 40.4292 25.4212C40.1645 24.8733 40.0322 24.2363 40.0322 23.5099C40.0322 22.7836 40.1645 22.1435 40.4292 21.5895C40.697 21.0324 41.0694 20.5985 41.5464 20.2876C42.0265 19.9737 42.5867 19.8168 43.2268 19.8168C43.5961 19.8168 43.9608 19.8783 44.3209 20.0014C44.681 20.1245 45.0088 20.3246 45.3042 20.6016C45.5997 20.8755 45.8351 21.2386 46.0105 21.6911C46.186 22.1435 46.2737 22.7005 46.2737 23.3622V23.8239H40.8078V22.8821H45.1657C45.1657 22.482 45.0857 22.125 44.9257 21.8111C44.7687 21.4972 44.544 21.2494 44.2517 21.0678C43.9624 20.8862 43.6208 20.7955 43.2268 20.7955C42.7929 20.7955 42.4174 20.9032 42.1004 21.1186C41.7865 21.331 41.5449 21.608 41.3756 21.9496C41.2063 22.2912 41.1217 22.6574 41.1217 23.0483V23.6761C41.1217 24.2116 41.214 24.6656 41.3987 25.038C41.5864 25.4073 41.8465 25.6889 42.1789 25.8828C42.5113 26.0736 42.8975 26.169 43.3376 26.169C43.6238 26.169 43.8824 26.129 44.1132 26.049C44.3471 25.9659 44.5487 25.8428 44.7179 25.6797C44.8872 25.5135 45.018 25.3073 45.1103 25.0611L46.1629 25.3565C46.0521 25.7135 45.8659 26.0275 45.6043 26.2983C45.3427 26.5661 45.0195 26.7753 44.6348 26.9261C44.2501 27.0739 43.8177 27.1477 43.3376 27.1477ZM50.7542 29.8068C50.2279 29.8068 49.7755 29.7391 49.397 29.6037C49.0184 29.4714 48.703 29.2959 48.4506 29.0774C48.2013 28.862 48.0028 28.6312 47.8551 28.3849L48.723 27.7756C48.8215 27.9048 48.9461 28.0526 49.0969 28.2188C49.2477 28.388 49.4539 28.5342 49.7155 28.6573C49.9802 28.7835 50.3264 28.8466 50.7542 28.8466C51.3267 28.8466 51.7991 28.7081 52.1715 28.4311C52.5439 28.1541 52.7301 27.7202 52.7301 27.1293V25.6889H52.6377C52.5577 25.8182 52.4439 25.9782 52.2961 26.169C52.1515 26.3568 51.9422 26.5245 51.6683 26.6722C51.3974 26.8169 51.0312 26.8892 50.5696 26.8892C49.9971 26.8892 49.4831 26.7538 49.0277 26.483C48.5752 26.2121 48.2167 25.8182 47.952 25.3011C47.6904 24.7841 47.5596 24.1562 47.5596 23.4176C47.5596 22.6913 47.6873 22.0588 47.9428 21.5202C48.1982 20.9786 48.5537 20.56 49.0092 20.2646C49.4647 19.966 49.991 19.8168 50.588 19.8168C51.0497 19.8168 51.4159 19.8937 51.6867 20.0476C51.9607 20.1984 52.1699 20.3707 52.3146 20.5646C52.4623 20.7554 52.5762 20.9124 52.6562 21.0355H52.767V19.9091H53.8196V27.2031C53.8196 27.8125 53.6811 28.308 53.4041 28.6896C53.1302 29.0743 52.7608 29.3559 52.2961 29.5344C51.8345 29.716 51.3205 29.8068 50.7542 29.8068ZM50.7173 25.9105C51.1543 25.9105 51.5236 25.8105 51.8252 25.6104C52.1269 25.4104 52.3561 25.1226 52.5131 24.7472C52.6701 24.3717 52.7485 23.9223 52.7485 23.3991C52.7485 22.8883 52.6716 22.4374 52.5177 22.0465C52.3638 21.6557 52.1361 21.3494 51.8345 21.1278C51.5329 20.9062 51.1605 20.7955 50.7173 20.7955C50.2556 20.7955 49.8709 20.9124 49.5632 21.1463C49.2585 21.3802 49.0292 21.6941 48.8753 22.0881C48.7245 22.482 48.6491 22.919 48.6491 23.3991C48.6491 23.8916 48.726 24.3271 48.8799 24.7056C49.0369 25.0811 49.2677 25.3765 49.5724 25.592C49.8802 25.8043 50.2618 25.9105 50.7173 25.9105ZM60.2436 24.1009V19.9091H61.3331V27H60.2436V25.7997H60.1698C60.0036 26.1598 59.745 26.466 59.3942 26.7184C59.0433 26.9677 58.6002 27.0923 58.0646 27.0923C57.6215 27.0923 57.2275 26.9954 56.8828 26.8015C56.5381 26.6045 56.2673 26.3091 56.0703 25.9151C55.8734 25.5181 55.7749 25.018 55.7749 24.4148V19.9091H56.8644V24.3409C56.8644 24.858 57.009 25.2704 57.2983 25.5781C57.5907 25.8859 57.9631 26.0398 58.4155 26.0398C58.6863 26.0398 58.9618 25.9705 59.2419 25.832C59.525 25.6935 59.762 25.4812 59.9528 25.195C60.1467 24.9087 60.2436 24.544 60.2436 24.1009ZM64.3791 17.5455V27H63.2896V17.5455H64.3791ZM68.4222 27.1662C67.9729 27.1662 67.5651 27.0816 67.1988 26.9123C66.8326 26.7399 66.5418 26.4922 66.3263 26.169C66.1109 25.8428 66.0032 25.4489 66.0032 24.9872C66.0032 24.581 66.0832 24.2517 66.2432 23.9993C66.4033 23.7438 66.6172 23.5438 66.8849 23.3991C67.1527 23.2545 67.4481 23.1468 67.7713 23.076C68.0975 23.0021 68.4253 22.9437 68.7546 22.9006C69.1855 22.8452 69.5348 22.8036 69.8025 22.7759C70.0734 22.7451 70.2703 22.6944 70.3934 22.6236C70.5196 22.5528 70.5827 22.4297 70.5827 22.2543V22.2173C70.5827 21.7618 70.4581 21.4079 70.2088 21.1555C69.9626 20.9032 69.5886 20.777 69.087 20.777C68.5668 20.777 68.1591 20.8909 67.8636 21.1186C67.5681 21.3464 67.3604 21.5895 67.2404 21.848L66.2063 21.4787C66.3909 21.0478 66.6372 20.7124 66.9449 20.4723C67.2558 20.2292 67.5943 20.0599 67.9605 19.9645C68.3299 19.866 68.693 19.8168 69.05 19.8168C69.2778 19.8168 69.5394 19.8445 69.8348 19.8999C70.1334 19.9522 70.4211 20.0614 70.6981 20.2276C70.9782 20.3938 71.2105 20.6446 71.3952 20.9801C71.5799 21.3156 71.6722 21.7649 71.6722 22.3281V27H70.5827V26.0398H70.5273C70.4534 26.1937 70.3303 26.3583 70.158 26.5337C69.9856 26.7092 69.7564 26.8584 69.4701 26.9815C69.1839 27.1046 68.8346 27.1662 68.4222 27.1662ZM68.5884 26.1875C69.0193 26.1875 69.3824 26.1029 69.6779 25.9336C69.9764 25.7643 70.2011 25.5458 70.3519 25.2781C70.5058 25.0103 70.5827 24.7287 70.5827 24.4332V23.4361C70.5365 23.4915 70.435 23.5423 70.278 23.5884C70.1241 23.6315 69.9456 23.67 69.7425 23.7038C69.5425 23.7346 69.347 23.7623 69.1562 23.7869C68.9685 23.8085 68.8161 23.8269 68.6992 23.8423C68.416 23.8793 68.1514 23.9393 67.9052 24.0224C67.662 24.1024 67.465 24.224 67.3142 24.3871C67.1665 24.5471 67.0927 24.7656 67.0927 25.0426C67.0927 25.4212 67.2327 25.7074 67.5128 25.9013C67.7959 26.0921 68.1544 26.1875 68.5884 26.1875ZM73.6217 27V19.9091H74.6743V20.9801H74.7482C74.8774 20.6293 75.1113 20.3446 75.4499 20.1261C75.7884 19.9076 76.17 19.7983 76.5948 19.7983C76.6748 19.7983 76.7748 19.7998 76.8948 19.8029C77.0149 19.806 77.1056 19.8106 77.1672 19.8168V20.9247C77.1303 20.9155 77.0456 20.9016 76.9133 20.8832C76.784 20.8616 76.6471 20.8509 76.5024 20.8509C76.1577 20.8509 75.85 20.9232 75.5791 21.0678C75.3114 21.2094 75.099 21.4064 74.9421 21.6587C74.7882 21.908 74.7112 22.1927 74.7112 22.5128V27H73.6217ZM79.5091 17.5455V27H78.4197V17.5455H79.5091ZM82.1673 29.6591C81.9827 29.6591 81.818 29.6437 81.6734 29.6129C81.5287 29.5852 81.4287 29.5575 81.3733 29.5298L81.6503 28.5696C81.915 28.6373 82.1489 28.6619 82.352 28.6435C82.5551 28.625 82.7352 28.5342 82.8921 28.3711C83.0521 28.2111 83.1983 27.951 83.3307 27.5909L83.5338 27.0369L80.9116 19.9091H82.0935L84.0508 25.5597H84.1247L86.0821 19.9091H87.2639L84.254 28.0341C84.1186 28.4003 83.9508 28.7035 83.7508 28.9435C83.5507 29.1867 83.3184 29.3667 83.0537 29.4837C82.7921 29.6006 82.4966 29.6591 82.1673 29.6591ZM95.2532 27.1477C94.57 27.1477 93.9806 26.9969 93.4851 26.6953C92.9927 26.3906 92.6126 25.9659 92.3448 25.4212C92.0801 24.8733 91.9478 24.2363 91.9478 23.5099C91.9478 22.7836 92.0801 22.1435 92.3448 21.5895C92.6126 21.0324 92.985 20.5985 93.462 20.2876C93.9421 19.9737 94.5023 19.8168 95.1424 19.8168C95.5117 19.8168 95.8764 19.8783 96.2365 20.0014C96.5966 20.1245 96.9244 20.3246 97.2198 20.6016C97.5153 20.8755 97.7507 21.2386 97.9261 21.6911C98.1016 22.1435 98.1893 22.7005 98.1893 23.3622V23.8239H92.7234V22.8821H97.0813C97.0813 22.482 97.0013 22.125 96.8413 21.8111C96.6843 21.4972 96.4596 21.2494 96.1673 21.0678C95.878 20.8862 95.5363 20.7955 95.1424 20.7955C94.7085 20.7955 94.333 20.9032 94.016 21.1186C93.7021 21.331 93.4605 21.608 93.2912 21.9496C93.1219 22.2912 93.0373 22.6574 93.0373 23.0483V23.6761C93.0373 24.2116 93.1296 24.6656 93.3143 25.038C93.502 25.4073 93.7621 25.6889 94.0945 25.8828C94.4269 26.0736 94.8131 26.169 95.2532 26.169C95.5394 26.169 95.798 26.129 96.0288 26.049C96.2627 25.9659 96.4643 25.8428 96.6335 25.6797C96.8028 25.5135 96.9336 25.3073 97.0259 25.0611L98.0785 25.3565C97.9677 25.7135 97.7815 26.0275 97.5199 26.2983C97.2583 26.5661 96.9351 26.7753 96.5504 26.9261C96.1657 27.0739 95.7333 27.1477 95.2532 27.1477ZM100.43 19.9091L102.129 22.8082L103.827 19.9091H105.083L102.793 23.4545L105.083 27H103.827L102.129 24.2486L100.43 27H99.174L101.427 23.4545L99.174 19.9091H100.43ZM106.599 29.6591V19.9091H107.651V21.0355H107.781C107.861 20.9124 107.971 20.7554 108.113 20.5646C108.258 20.3707 108.464 20.1984 108.731 20.0476C109.002 19.8937 109.369 19.8168 109.83 19.8168C110.427 19.8168 110.954 19.966 111.409 20.2646C111.865 20.5631 112.22 20.9863 112.475 21.5341C112.731 22.0819 112.859 22.7282 112.859 23.473C112.859 24.224 112.731 24.8749 112.475 25.4258C112.22 25.9736 111.866 26.3983 111.414 26.6999C110.961 26.9985 110.44 27.1477 109.849 27.1477C109.393 27.1477 109.028 27.0723 108.755 26.9215C108.481 26.7676 108.27 26.5937 108.122 26.3999C107.974 26.2029 107.861 26.0398 107.781 25.9105H107.688V29.6591H106.599ZM107.67 23.4545C107.67 23.9901 107.748 24.4625 107.905 24.8718C108.062 25.2781 108.291 25.5966 108.593 25.8274C108.895 26.0552 109.264 26.169 109.701 26.169C110.156 26.169 110.537 26.049 110.841 25.8089C111.149 25.5658 111.38 25.2396 111.534 24.8303C111.691 24.4179 111.769 23.9593 111.769 23.4545C111.769 22.956 111.692 22.5066 111.538 22.1065C111.388 21.7034 111.158 21.3848 110.85 21.1509C110.546 20.9139 110.163 20.7955 109.701 20.7955C109.258 20.7955 108.885 20.9078 108.584 21.1325C108.282 21.354 108.054 21.6649 107.901 22.065C107.747 22.462 107.67 22.9252 107.67 23.4545ZM116.568 27.1662C116.119 27.1662 115.711 27.0816 115.345 26.9123C114.979 26.7399 114.688 26.4922 114.472 26.169C114.257 25.8428 114.149 25.4489 114.149 24.9872C114.149 24.581 114.229 24.2517 114.389 23.9993C114.549 23.7438 114.763 23.5438 115.031 23.3991C115.299 23.2545 115.594 23.1468 115.917 23.076C116.244 23.0021 116.571 22.9437 116.901 22.9006C117.331 22.8452 117.681 22.8036 117.949 22.7759C118.219 22.7451 118.416 22.6944 118.539 22.6236C118.666 22.5528 118.729 22.4297 118.729 22.2543V22.2173C118.729 21.7618 118.604 21.4079 118.355 21.1555C118.109 20.9032 117.735 20.777 117.233 20.777C116.713 20.777 116.305 20.8909 116.01 21.1186C115.714 21.3464 115.506 21.5895 115.386 21.848L114.352 21.4787C114.537 21.0478 114.783 20.7124 115.091 20.4723C115.402 20.2292 115.74 20.0599 116.107 19.9645C116.476 19.866 116.839 19.8168 117.196 19.8168C117.424 19.8168 117.685 19.8445 117.981 19.8999C118.279 19.9522 118.567 20.0614 118.844 20.2276C119.124 20.3938 119.357 20.6446 119.541 20.9801C119.726 21.3156 119.818 21.7649 119.818 22.3281V27H118.729V26.0398H118.673C118.599 26.1937 118.476 26.3583 118.304 26.5337C118.132 26.7092 117.902 26.8584 117.616 26.9815C117.33 27.1046 116.981 27.1662 116.568 27.1662ZM116.734 26.1875C117.165 26.1875 117.528 26.1029 117.824 25.9336C118.122 25.7643 118.347 25.5458 118.498 25.2781C118.652 25.0103 118.729 24.7287 118.729 24.4332V23.4361C118.683 23.4915 118.581 23.5423 118.424 23.5884C118.27 23.6315 118.092 23.67 117.889 23.7038C117.688 23.7346 117.493 23.7623 117.302 23.7869C117.114 23.8085 116.962 23.8269 116.845 23.8423C116.562 23.8793 116.297 23.9393 116.051 24.0224C115.808 24.1024 115.611 24.224 115.46 24.3871C115.313 24.5471 115.239 24.7656 115.239 25.0426C115.239 25.4212 115.379 25.7074 115.659 25.9013C115.942 26.0921 116.3 26.1875 116.734 26.1875ZM122.857 22.7344V27H121.768V19.9091H122.82V21.017H122.913C123.079 20.657 123.331 20.3677 123.67 20.1491C124.008 19.9276 124.445 19.8168 124.981 19.8168C125.461 19.8168 125.881 19.9152 126.241 20.1122C126.601 20.3061 126.881 20.6016 127.081 20.9986C127.281 21.3925 127.381 21.8911 127.381 22.4943V27H126.292V22.5682C126.292 22.0111 126.147 21.5772 125.858 21.2663C125.569 20.9524 125.172 20.7955 124.667 20.7955C124.319 20.7955 124.008 20.8709 123.734 21.0217C123.464 21.1725 123.25 21.3925 123.093 21.6818C122.936 21.9711 122.857 22.322 122.857 22.7344ZM132.011 27.1477C131.42 27.1477 130.898 26.9985 130.446 26.6999C129.993 26.3983 129.639 25.9736 129.384 25.4258C129.129 24.8749 129.001 24.224 129.001 23.473C129.001 22.7282 129.129 22.0819 129.384 21.5341C129.639 20.9863 129.995 20.5631 130.45 20.2646C130.906 19.966 131.432 19.8168 132.029 19.8168C132.491 19.8168 132.856 19.8937 133.123 20.0476C133.394 20.1984 133.6 20.3707 133.742 20.5646C133.887 20.7554 133.999 20.9124 134.079 21.0355H134.171V17.5455H135.261V27H134.208V25.9105H134.079C133.999 26.0398 133.885 26.2029 133.737 26.3999C133.59 26.5937 133.379 26.7676 133.105 26.9215C132.831 27.0723 132.466 27.1477 132.011 27.1477ZM132.159 26.169C132.596 26.169 132.965 26.0552 133.266 25.8274C133.568 25.5966 133.797 25.2781 133.954 24.8718C134.111 24.4625 134.19 23.9901 134.19 23.4545C134.19 22.9252 134.113 22.462 133.959 22.065C133.805 21.6649 133.577 21.354 133.276 21.1325C132.974 20.9078 132.602 20.7955 132.159 20.7955C131.697 20.7955 131.312 20.9139 131.004 21.1509C130.7 21.3848 130.47 21.7034 130.317 22.1065C130.166 22.5066 130.09 22.956 130.09 23.4545C130.09 23.9593 130.167 24.4179 130.321 24.8303C130.478 25.2396 130.709 25.5658 131.014 25.8089C131.321 26.049 131.703 26.169 132.159 26.169ZM143.866 27.1477C143.226 27.1477 142.665 26.9954 142.181 26.6907C141.701 26.386 141.326 25.9598 141.055 25.4119C140.787 24.8641 140.653 24.224 140.653 23.4915C140.653 22.7528 140.787 22.1081 141.055 21.5572C141.326 21.0063 141.701 20.5785 142.181 20.2738C142.665 19.9691 143.226 19.8168 143.866 19.8168C144.507 19.8168 145.067 19.9691 145.547 20.2738C146.03 20.5785 146.405 21.0063 146.673 21.5572C146.944 22.1081 147.079 22.7528 147.079 23.4915C147.079 24.224 146.944 24.8641 146.673 25.4119C146.405 25.9598 146.03 26.386 145.547 26.6907C145.067 26.9954 144.507 27.1477 143.866 27.1477ZM143.866 26.169C144.353 26.169 144.753 26.0444 145.067 25.7951C145.381 25.5458 145.613 25.218 145.764 24.8118C145.915 24.4055 145.99 23.9654 145.99 23.4915C145.99 23.0175 145.915 22.5759 145.764 22.1665C145.613 21.7572 145.381 21.4264 145.067 21.174C144.753 20.9216 144.353 20.7955 143.866 20.7955C143.38 20.7955 142.98 20.9216 142.666 21.174C142.352 21.4264 142.12 21.7572 141.969 22.1665C141.818 22.5759 141.743 23.0175 141.743 23.4915C141.743 23.9654 141.818 24.4055 141.969 24.8118C142.12 25.218 142.352 25.5458 142.666 25.7951C142.98 26.0444 143.38 26.169 143.866 26.169ZM153.172 24.1009V19.9091H154.262V27H153.172V25.7997H153.098C152.932 26.1598 152.674 26.466 152.323 26.7184C151.972 26.9677 151.529 27.0923 150.993 27.0923C150.55 27.0923 150.156 26.9954 149.812 26.8015C149.467 26.6045 149.196 26.3091 148.999 25.9151C148.802 25.5181 148.704 25.018 148.704 24.4148V19.9091H149.793V24.3409C149.793 24.858 149.938 25.2704 150.227 25.5781C150.519 25.8859 150.892 26.0398 151.344 26.0398C151.615 26.0398 151.89 25.9705 152.171 25.832C152.454 25.6935 152.691 25.4812 152.881 25.195C153.075 24.9087 153.172 24.544 153.172 24.1009ZM156.218 27V19.9091H157.271V20.9801H157.345C157.474 20.6293 157.708 20.3446 158.046 20.1261C158.385 19.9076 158.767 19.7983 159.191 19.7983C159.271 19.7983 159.371 19.7998 159.491 19.8029C159.611 19.806 159.702 19.8106 159.764 19.8168V20.9247C159.727 20.9155 159.642 20.9016 159.51 20.8832C159.381 20.8616 159.244 20.8509 159.099 20.8509C158.754 20.8509 158.446 20.9232 158.176 21.0678C157.908 21.2094 157.696 21.4064 157.539 21.6587C157.385 21.908 157.308 22.1927 157.308 22.5128V27H156.218ZM169.656 21.4972L168.677 21.7741C168.616 21.611 168.525 21.4525 168.405 21.2987C168.288 21.1417 168.128 21.0124 167.925 20.9109C167.722 20.8093 167.462 20.7585 167.145 20.7585C166.711 20.7585 166.349 20.8585 166.06 21.0586C165.774 21.2556 165.631 21.5064 165.631 21.8111C165.631 22.0819 165.729 22.2958 165.926 22.4528C166.123 22.6097 166.431 22.7405 166.849 22.8452L167.902 23.1037C168.536 23.2576 169.008 23.493 169.319 23.81C169.63 24.1239 169.785 24.5286 169.785 25.0241C169.785 25.4304 169.668 25.7936 169.435 26.1136C169.204 26.4337 168.881 26.6861 168.465 26.8707C168.05 27.0554 167.566 27.1477 167.016 27.1477C166.292 27.1477 165.694 26.9908 165.22 26.6768C164.746 26.3629 164.446 25.9044 164.319 25.3011L165.354 25.0426C165.452 25.4242 165.638 25.7105 165.912 25.9013C166.189 26.0921 166.551 26.1875 166.997 26.1875C167.505 26.1875 167.908 26.0798 168.207 25.8643C168.508 25.6458 168.659 25.3842 168.659 25.0795C168.659 24.8333 168.573 24.6271 168.4 24.4609C168.228 24.2917 167.963 24.1655 167.606 24.0824L166.425 23.8054C165.775 23.6515 165.298 23.413 164.994 23.0898C164.692 22.7636 164.541 22.3558 164.541 21.8665C164.541 21.4664 164.653 21.1125 164.878 20.8047C165.106 20.4969 165.415 20.2553 165.806 20.0799C166.2 19.9045 166.646 19.8168 167.145 19.8168C167.846 19.8168 168.397 19.9706 168.797 20.2784C169.201 20.5862 169.487 20.9924 169.656 21.4972ZM171.386 27V19.9091H172.476V27H171.386ZM171.94 18.7273C171.728 18.7273 171.545 18.6549 171.391 18.5103C171.24 18.3656 171.165 18.1918 171.165 17.9886C171.165 17.7855 171.24 17.6116 171.391 17.467C171.545 17.3223 171.728 17.25 171.94 17.25C172.153 17.25 172.334 17.3223 172.485 17.467C172.639 17.6116 172.716 17.7855 172.716 17.9886C172.716 18.1918 172.639 18.3656 172.485 18.5103C172.334 18.6549 172.153 18.7273 171.94 18.7273ZM175.522 22.7344V27H174.432V19.9091H175.485V21.017H175.577C175.743 20.657 175.996 20.3677 176.334 20.1491C176.673 19.9276 177.11 19.8168 177.645 19.8168C178.126 19.8168 178.546 19.9152 178.906 20.1122C179.266 20.3061 179.546 20.6016 179.746 20.9986C179.946 21.3925 180.046 21.8911 180.046 22.4943V27H178.957V22.5682C178.957 22.0111 178.812 21.5772 178.523 21.2663C178.233 20.9524 177.836 20.7955 177.332 20.7955C176.984 20.7955 176.673 20.8709 176.399 21.0217C176.128 21.1725 175.914 21.3925 175.757 21.6818C175.6 21.9711 175.522 22.322 175.522 22.7344ZM184.86 29.8068C184.334 29.8068 183.881 29.7391 183.503 29.6037C183.124 29.4714 182.809 29.2959 182.556 29.0774C182.307 28.862 182.109 28.6312 181.961 28.3849L182.829 27.7756C182.927 27.9048 183.052 28.0526 183.203 28.2188C183.354 28.388 183.56 28.5342 183.821 28.6573C184.086 28.7835 184.432 28.8466 184.86 28.8466C185.433 28.8466 185.905 28.7081 186.277 28.4311C186.65 28.1541 186.836 27.7202 186.836 27.1293V25.6889H186.744C186.664 25.8182 186.55 25.9782 186.402 26.169C186.257 26.3568 186.048 26.5245 185.774 26.6722C185.503 26.8169 185.137 26.8892 184.675 26.8892C184.103 26.8892 183.589 26.7538 183.134 26.483C182.681 26.2121 182.323 25.8182 182.058 25.3011C181.796 24.7841 181.665 24.1562 181.665 23.4176C181.665 22.6913 181.793 22.0588 182.049 21.5202C182.304 20.9786 182.66 20.56 183.115 20.2646C183.571 19.966 184.097 19.8168 184.694 19.8168C185.156 19.8168 185.522 19.8937 185.793 20.0476C186.067 20.1984 186.276 20.3707 186.42 20.5646C186.568 20.7554 186.682 20.9124 186.762 21.0355H186.873V19.9091H187.925V27.2031C187.925 27.8125 187.787 28.308 187.51 28.6896C187.236 29.0743 186.867 29.3559 186.402 29.5344C185.94 29.716 185.426 29.8068 184.86 29.8068ZM184.823 25.9105C185.26 25.9105 185.63 25.8105 185.931 25.6104C186.233 25.4104 186.462 25.1226 186.619 24.7472C186.776 24.3717 186.854 23.9223 186.854 23.3991C186.854 22.8883 186.777 22.4374 186.624 22.0465C186.47 21.6557 186.242 21.3494 185.94 21.1278C185.639 20.9062 185.266 20.7955 184.823 20.7955C184.362 20.7955 183.977 20.9124 183.669 21.1463C183.364 21.3802 183.135 21.6941 182.981 22.0881C182.83 22.482 182.755 22.919 182.755 23.3991C182.755 23.8916 182.832 24.3271 182.986 24.7056C183.143 25.0811 183.374 25.3765 183.678 25.592C183.986 25.8043 184.368 25.9105 184.823 25.9105ZM190.97 17.5455V27H189.881V17.5455H190.97ZM195.9 27.1477C195.216 27.1477 194.627 26.9969 194.132 26.6953C193.639 26.3906 193.259 25.9659 192.991 25.4212C192.727 24.8733 192.594 24.2363 192.594 23.5099C192.594 22.7836 192.727 22.1435 192.991 21.5895C193.259 21.0324 193.631 20.5985 194.109 20.2876C194.589 19.9737 195.149 19.8168 195.789 19.8168C196.158 19.8168 196.523 19.8783 196.883 20.0014C197.243 20.1245 197.571 20.3246 197.866 20.6016C198.162 20.8755 198.397 21.2386 198.573 21.6911C198.748 22.1435 198.836 22.7005 198.836 23.3622V23.8239H193.37V22.8821H197.728C197.728 22.482 197.648 22.125 197.488 21.8111C197.331 21.4972 197.106 21.2494 196.814 21.0678C196.524 20.8862 196.183 20.7955 195.789 20.7955C195.355 20.7955 194.98 20.9032 194.663 21.1186C194.349 21.331 194.107 21.608 193.938 21.9496C193.768 22.2912 193.684 22.6574 193.684 23.0483V23.6761C193.684 24.2116 193.776 24.6656 193.961 25.038C194.149 25.4073 194.409 25.6889 194.741 25.8828C195.073 26.0736 195.46 26.169 195.9 26.169C196.186 26.169 196.444 26.129 196.675 26.049C196.909 25.9659 197.111 25.8428 197.28 25.6797C197.449 25.5135 197.58 25.3073 197.672 25.0611L198.725 25.3565C198.614 25.7135 198.428 26.0275 198.166 26.2983C197.905 26.5661 197.582 26.7753 197.197 26.9261C196.812 27.0739 196.38 27.1477 195.9 27.1477ZM206.952 27.1477C206.287 27.1477 205.715 26.9908 205.235 26.6768C204.755 26.3629 204.385 25.9305 204.127 25.3796C203.868 24.8287 203.739 24.1993 203.739 23.4915C203.739 22.7713 203.871 22.1358 204.136 21.5849C204.404 21.0309 204.776 20.5985 205.253 20.2876C205.733 19.9737 206.293 19.8168 206.934 19.8168C207.432 19.8168 207.881 19.9091 208.282 20.0938C208.682 20.2784 209.009 20.5369 209.265 20.8693C209.52 21.2017 209.679 21.5895 209.74 22.0327H208.651C208.568 21.7095 208.383 21.4233 208.097 21.174C207.814 20.9216 207.432 20.7955 206.952 20.7955C206.527 20.7955 206.155 20.9062 205.835 21.1278C205.518 21.3464 205.27 21.6557 205.092 22.0558C204.916 22.4528 204.828 22.919 204.828 23.4545C204.828 24.0024 204.915 24.4794 205.087 24.8857C205.262 25.2919 205.509 25.6074 205.826 25.832C206.146 26.0567 206.521 26.169 206.952 26.169C207.235 26.169 207.492 26.1198 207.723 26.0213C207.954 25.9228 208.149 25.7812 208.309 25.5966C208.469 25.4119 208.583 25.1903 208.651 24.9318H209.74C209.679 25.3504 209.526 25.7274 209.283 26.0629C209.043 26.3952 208.725 26.6599 208.328 26.8569C207.934 27.0508 207.475 27.1477 206.952 27.1477ZM214.267 27.1477C213.584 27.1477 212.994 26.9969 212.499 26.6953C212.007 26.3906 211.626 25.9659 211.359 25.4212C211.094 24.8733 210.962 24.2363 210.962 23.5099C210.962 22.7836 211.094 22.1435 211.359 21.5895C211.626 21.0324 211.999 20.5985 212.476 20.2876C212.956 19.9737 213.516 19.8168 214.156 19.8168C214.526 19.8168 214.89 19.8783 215.25 20.0014C215.61 20.1245 215.938 20.3246 216.234 20.6016C216.529 20.8755 216.765 21.2386 216.94 21.6911C217.115 22.1435 217.203 22.7005 217.203 23.3622V23.8239H211.737V22.8821H216.095C216.095 22.482 216.015 22.125 215.855 21.8111C215.698 21.4972 215.474 21.2494 215.181 21.0678C214.892 20.8862 214.55 20.7955 214.156 20.7955C213.722 20.7955 213.347 20.9032 213.03 21.1186C212.716 21.331 212.474 21.608 212.305 21.9496C212.136 22.2912 212.051 22.6574 212.051 23.0483V23.6761C212.051 24.2116 212.144 24.6656 212.328 25.038C212.516 25.4073 212.776 25.6889 213.108 25.8828C213.441 26.0736 213.827 26.169 214.267 26.169C214.553 26.169 214.812 26.129 215.043 26.049C215.277 25.9659 215.478 25.8428 215.647 25.6797C215.817 25.5135 215.947 25.3073 216.04 25.0611L217.092 25.3565C216.982 25.7135 216.795 26.0275 216.534 26.2983C216.272 26.5661 215.949 26.7753 215.564 26.9261C215.18 27.0739 214.747 27.1477 214.267 27.1477ZM219.911 17.5455V27H218.821V17.5455H219.911ZM222.957 17.5455V27H221.867V17.5455H222.957ZM231.208 27.1477C230.617 27.1477 230.096 26.9985 229.643 26.6999C229.191 26.3983 228.837 25.9736 228.581 25.4258C228.326 24.8749 228.198 24.224 228.198 23.473C228.198 22.7282 228.326 22.0819 228.581 21.5341C228.837 20.9863 229.192 20.5631 229.648 20.2646C230.103 19.966 230.63 19.8168 231.227 19.8168C231.688 19.8168 232.053 19.8937 232.321 20.0476C232.592 20.1984 232.798 20.3707 232.939 20.5646C233.084 20.7554 233.196 20.9124 233.276 21.0355H233.369V17.5455H234.458V27H233.406V25.9105H233.276C233.196 26.0398 233.082 26.2029 232.935 26.3999C232.787 26.5937 232.576 26.7676 232.302 26.9215C232.028 27.0723 231.664 27.1477 231.208 27.1477ZM231.356 26.169C231.793 26.169 232.162 26.0552 232.464 25.8274C232.766 25.5966 232.995 25.2781 233.152 24.8718C233.309 24.4625 233.387 23.9901 233.387 23.4545C233.387 22.9252 233.31 22.462 233.156 22.065C233.002 21.6649 232.775 21.354 232.473 21.1325C232.172 20.9078 231.799 20.7955 231.356 20.7955C230.894 20.7955 230.51 20.9139 230.202 21.1509C229.897 21.3848 229.668 21.7034 229.514 22.1065C229.363 22.5066 229.288 22.956 229.288 23.4545C229.288 23.9593 229.365 24.4179 229.519 24.8303C229.676 25.2396 229.906 25.5658 230.211 25.8089C230.519 26.049 230.9 26.169 231.356 26.169ZM238.653 27.1662C238.203 27.1662 237.795 27.0816 237.429 26.9123C237.063 26.7399 236.772 26.4922 236.557 26.169C236.341 25.8428 236.233 25.4489 236.233 24.9872C236.233 24.581 236.314 24.2517 236.474 23.9993C236.634 23.7438 236.847 23.5438 237.115 23.3991C237.383 23.2545 237.678 23.1468 238.002 23.076C238.328 23.0021 238.656 22.9437 238.985 22.9006C239.416 22.8452 239.765 22.8036 240.033 22.7759C240.304 22.7451 240.501 22.6944 240.624 22.6236C240.75 22.5528 240.813 22.4297 240.813 22.2543V22.2173C240.813 21.7618 240.688 21.4079 240.439 21.1555C240.193 20.9032 239.819 20.777 239.317 20.777C238.797 20.777 238.389 20.8909 238.094 21.1186C237.798 21.3464 237.591 21.5895 237.471 21.848L236.437 21.4787C236.621 21.0478 236.867 20.7124 237.175 20.4723C237.486 20.2292 237.825 20.0599 238.191 19.9645C238.56 19.866 238.923 19.8168 239.28 19.8168C239.508 19.8168 239.77 19.8445 240.065 19.8999C240.364 19.9522 240.651 20.0614 240.928 20.2276C241.209 20.3938 241.441 20.6446 241.626 20.9801C241.81 21.3156 241.903 21.7649 241.903 22.3281V27H240.813V26.0398H240.758C240.684 26.1937 240.561 26.3583 240.388 26.5337C240.216 26.7092 239.987 26.8584 239.7 26.9815C239.414 27.1046 239.065 27.1662 238.653 27.1662ZM238.819 26.1875C239.25 26.1875 239.613 26.1029 239.908 25.9336C240.207 25.7643 240.431 25.5458 240.582 25.2781C240.736 25.0103 240.813 24.7287 240.813 24.4332V23.4361C240.767 23.4915 240.665 23.5423 240.508 23.5884C240.354 23.6315 240.176 23.67 239.973 23.7038C239.773 23.7346 239.577 23.7623 239.387 23.7869C239.199 23.8085 239.046 23.8269 238.93 23.8423C238.646 23.8793 238.382 23.9393 238.135 24.0224C237.892 24.1024 237.695 24.224 237.545 24.3871C237.397 24.5471 237.323 24.7656 237.323 25.0426C237.323 25.4212 237.463 25.7074 237.743 25.9013C238.026 26.0921 238.385 26.1875 238.819 26.1875ZM246.936 19.9091V20.8324H243.261V19.9091H246.936ZM244.332 18.2102H245.422V24.9688C245.422 25.2765 245.466 25.5073 245.556 25.6612C245.648 25.812 245.765 25.9136 245.906 25.9659C246.051 26.0152 246.203 26.0398 246.363 26.0398C246.483 26.0398 246.582 26.0336 246.659 26.0213C246.736 26.0059 246.797 25.9936 246.844 25.9844L247.065 26.9631C246.991 26.9908 246.888 27.0185 246.756 27.0462C246.624 27.0769 246.456 27.0923 246.253 27.0923C245.945 27.0923 245.643 27.0262 245.348 26.8938C245.055 26.7615 244.812 26.5599 244.618 26.2891C244.428 26.0182 244.332 25.6766 244.332 25.2642V18.2102ZM250.622 27.1662C250.173 27.1662 249.765 27.0816 249.399 26.9123C249.033 26.7399 248.742 26.4922 248.527 26.169C248.311 25.8428 248.203 25.4489 248.203 24.9872C248.203 24.581 248.283 24.2517 248.443 23.9993C248.603 23.7438 248.817 23.5438 249.085 23.3991C249.353 23.2545 249.648 23.1468 249.971 23.076C250.298 23.0021 250.625 22.9437 250.955 22.9006C251.386 22.8452 251.735 22.8036 252.003 22.7759C252.274 22.7451 252.471 22.6944 252.594 22.6236C252.72 22.5528 252.783 22.4297 252.783 22.2543V22.2173C252.783 21.7618 252.658 21.4079 252.409 21.1555C252.163 20.9032 251.789 20.777 251.287 20.777C250.767 20.777 250.359 20.8909 250.064 21.1186C249.768 21.3464 249.561 21.5895 249.441 21.848L248.406 21.4787C248.591 21.0478 248.837 20.7124 249.145 20.4723C249.456 20.2292 249.794 20.0599 250.161 19.9645C250.53 19.866 250.893 19.8168 251.25 19.8168C251.478 19.8168 251.74 19.8445 252.035 19.8999C252.334 19.9522 252.621 20.0614 252.898 20.2276C253.178 20.3938 253.411 20.6446 253.595 20.9801C253.78 21.3156 253.872 21.7649 253.872 22.3281V27H252.783V26.0398H252.727C252.654 26.1937 252.531 26.3583 252.358 26.5337C252.186 26.7092 251.957 26.8584 251.67 26.9815C251.384 27.1046 251.035 27.1662 250.622 27.1662ZM250.789 26.1875C251.219 26.1875 251.583 26.1029 251.878 25.9336C252.177 25.7643 252.401 25.5458 252.552 25.2781C252.706 25.0103 252.783 24.7287 252.783 24.4332V23.4361C252.737 23.4915 252.635 23.5423 252.478 23.5884C252.324 23.6315 252.146 23.67 251.943 23.7038C251.743 23.7346 251.547 23.7623 251.356 23.7869C251.169 23.8085 251.016 23.8269 250.899 23.8423C250.616 23.8793 250.352 23.9393 250.105 24.0224C249.862 24.1024 249.665 24.224 249.514 24.3871C249.367 24.5471 249.293 24.7656 249.293 25.0426C249.293 25.4212 249.433 25.7074 249.713 25.9013C249.996 26.0921 250.355 26.1875 250.789 26.1875ZM262.32 27.1477C261.655 27.1477 261.083 26.9908 260.603 26.6768C260.122 26.3629 259.753 25.9305 259.495 25.3796C259.236 24.8287 259.107 24.1993 259.107 23.4915C259.107 22.7713 259.239 22.1358 259.504 21.5849C259.772 21.0309 260.144 20.5985 260.621 20.2876C261.101 19.9737 261.661 19.8168 262.301 19.8168C262.8 19.8168 263.249 19.9091 263.649 20.0938C264.049 20.2784 264.377 20.5369 264.633 20.8693C264.888 21.2017 265.047 21.5895 265.108 22.0327H264.019C263.936 21.7095 263.751 21.4233 263.465 21.174C263.182 20.9216 262.8 20.7955 262.32 20.7955C261.895 20.7955 261.523 20.9062 261.203 21.1278C260.886 21.3464 260.638 21.6557 260.459 22.0558C260.284 22.4528 260.196 22.919 260.196 23.4545C260.196 24.0024 260.282 24.4794 260.455 24.8857C260.63 25.2919 260.876 25.6074 261.193 25.832C261.514 26.0567 261.889 26.169 262.32 26.169C262.603 26.169 262.86 26.1198 263.091 26.0213C263.322 25.9228 263.517 25.7812 263.677 25.5966C263.837 25.4119 263.951 25.1903 264.019 24.9318H265.108C265.047 25.3504 264.894 25.7274 264.651 26.0629C264.411 26.3952 264.093 26.6599 263.696 26.8569C263.302 27.0508 262.843 27.1477 262.32 27.1477ZM269.543 27.1477C268.902 27.1477 268.341 26.9954 267.858 26.6907C267.377 26.386 267.002 25.9598 266.731 25.4119C266.463 24.8641 266.33 24.224 266.33 23.4915C266.33 22.7528 266.463 22.1081 266.731 21.5572C267.002 21.0063 267.377 20.5785 267.858 20.2738C268.341 19.9691 268.902 19.8168 269.543 19.8168C270.183 19.8168 270.743 19.9691 271.223 20.2738C271.706 20.5785 272.082 21.0063 272.349 21.5572C272.62 22.1081 272.756 22.7528 272.756 23.4915C272.756 24.224 272.62 24.8641 272.349 25.4119C272.082 25.9598 271.706 26.386 271.223 26.6907C270.743 26.9954 270.183 27.1477 269.543 27.1477ZM269.543 26.169C270.029 26.169 270.429 26.0444 270.743 25.7951C271.057 25.5458 271.289 25.218 271.44 24.8118C271.591 24.4055 271.666 23.9654 271.666 23.4915C271.666 23.0175 271.591 22.5759 271.44 22.1665C271.289 21.7572 271.057 21.4264 270.743 21.174C270.429 20.9216 270.029 20.7955 269.543 20.7955C269.056 20.7955 268.656 20.9216 268.342 21.174C268.028 21.4264 267.796 21.7572 267.645 22.1665C267.494 22.5759 267.419 23.0175 267.419 23.4915C267.419 23.9654 267.494 24.4055 267.645 24.8118C267.796 25.218 268.028 25.5458 268.342 25.7951C268.656 26.0444 269.056 26.169 269.543 26.169ZM274.38 27V19.9091H275.432V20.9801H275.506C275.635 20.6293 275.869 20.3446 276.208 20.1261C276.546 19.9076 276.928 19.7983 277.353 19.7983C277.433 19.7983 277.533 19.7998 277.653 19.8029C277.773 19.806 277.864 19.8106 277.925 19.8168V20.9247C277.888 20.9155 277.804 20.9016 277.671 20.8832C277.542 20.8616 277.405 20.8509 277.26 20.8509C276.916 20.8509 276.608 20.9232 276.337 21.0678C276.069 21.2094 275.857 21.4064 275.7 21.6587C275.546 21.908 275.469 22.1927 275.469 22.5128V27H274.38ZM279.178 29.6591V19.9091H280.23V21.0355H280.359C280.439 20.9124 280.55 20.7554 280.692 20.5646C280.836 20.3707 281.043 20.1984 281.31 20.0476C281.581 19.8937 281.948 19.8168 282.409 19.8168C283.006 19.8168 283.533 19.966 283.988 20.2646C284.444 20.5631 284.799 20.9863 285.054 21.5341C285.31 22.0819 285.438 22.7282 285.438 23.473C285.438 24.224 285.31 24.8749 285.054 25.4258C284.799 25.9736 284.445 26.3983 283.993 26.6999C283.54 26.9985 283.019 27.1477 282.428 27.1477C281.972 27.1477 281.607 27.0723 281.334 26.9215C281.06 26.7676 280.849 26.5937 280.701 26.3999C280.553 26.2029 280.439 26.0398 280.359 25.9105H280.267V29.6591H279.178ZM280.249 23.4545C280.249 23.9901 280.327 24.4625 280.484 24.8718C280.641 25.2781 280.87 25.5966 281.172 25.8274C281.474 26.0552 281.843 26.169 282.28 26.169C282.735 26.169 283.115 26.049 283.42 25.8089C283.728 25.5658 283.959 25.2396 284.113 24.8303C284.27 24.4179 284.348 23.9593 284.348 23.4545C284.348 22.956 284.271 22.5066 284.117 22.1065C283.966 21.7034 283.737 21.3848 283.429 21.1509C283.125 20.9139 282.742 20.7955 282.28 20.7955C281.837 20.7955 281.464 20.9078 281.163 21.1325C280.861 21.354 280.633 21.6649 280.479 22.065C280.326 22.462 280.249 22.9252 280.249 23.4545ZM291.529 24.1009V19.9091H292.619V27H291.529V25.7997H291.455C291.289 26.1598 291.031 26.466 290.68 26.7184C290.329 26.9677 289.886 27.0923 289.35 27.0923C288.907 27.0923 288.513 26.9954 288.168 26.8015C287.824 26.6045 287.553 26.3091 287.356 25.9151C287.159 25.5181 287.061 25.018 287.061 24.4148V19.9091H288.15V24.3409C288.15 24.858 288.295 25.2704 288.584 25.5781C288.876 25.8859 289.249 26.0398 289.701 26.0398C289.972 26.0398 290.247 25.9705 290.527 25.832C290.811 25.6935 291.048 25.4812 291.238 25.195C291.432 24.9087 291.529 24.544 291.529 24.1009ZM299.598 21.4972L298.619 21.7741C298.558 21.611 298.467 21.4525 298.347 21.2987C298.23 21.1417 298.07 21.0124 297.867 20.9109C297.664 20.8093 297.404 20.7585 297.087 20.7585C296.653 20.7585 296.291 20.8585 296.002 21.0586C295.715 21.2556 295.572 21.5064 295.572 21.8111C295.572 22.0819 295.671 22.2958 295.868 22.4528C296.065 22.6097 296.373 22.7405 296.791 22.8452L297.844 23.1037C298.478 23.2576 298.95 23.493 299.261 23.81C299.572 24.1239 299.727 24.5286 299.727 25.0241C299.727 25.4304 299.61 25.7936 299.376 26.1136C299.146 26.4337 298.822 26.6861 298.407 26.8707C297.991 27.0554 297.508 27.1477 296.957 27.1477C296.234 27.1477 295.635 26.9908 295.162 26.6768C294.688 26.3629 294.387 25.9044 294.261 25.3011L295.295 25.0426C295.394 25.4242 295.58 25.7105 295.854 25.9013C296.131 26.0921 296.493 26.1875 296.939 26.1875C297.447 26.1875 297.85 26.0798 298.148 25.8643C298.45 25.6458 298.601 25.3842 298.601 25.0795C298.601 24.8333 298.515 24.6271 298.342 24.4609C298.17 24.2917 297.905 24.1655 297.548 24.0824L296.366 23.8054C295.717 23.6515 295.24 23.413 294.935 23.0898C294.634 22.7636 294.483 22.3558 294.483 21.8665C294.483 21.4664 294.595 21.1125 294.82 20.8047C295.048 20.4969 295.357 20.2553 295.748 20.0799C296.142 19.9045 296.588 19.8168 297.087 19.8168C297.788 19.8168 298.339 19.9706 298.739 20.2784C299.142 20.5862 299.429 20.9924 299.598 21.4972ZM308.029 19.9091V20.8324H304.355V19.9091H308.029ZM305.426 18.2102H306.515V24.9688C306.515 25.2765 306.56 25.5073 306.649 25.6612C306.741 25.812 306.858 25.9136 307 25.9659C307.144 26.0152 307.297 26.0398 307.457 26.0398C307.577 26.0398 307.675 26.0336 307.752 26.0213C307.829 26.0059 307.891 25.9936 307.937 25.9844L308.159 26.9631C308.085 26.9908 307.982 27.0185 307.849 27.0462C307.717 27.0769 307.549 27.0923 307.346 27.0923C307.038 27.0923 306.737 27.0262 306.441 26.8938C306.149 26.7615 305.906 26.5599 305.712 26.2891C305.521 26.0182 305.426 25.6766 305.426 25.2642V18.2102ZM312.434 27.1477C311.793 27.1477 311.232 26.9954 310.749 26.6907C310.268 26.386 309.893 25.9598 309.622 25.4119C309.354 24.8641 309.221 24.224 309.221 23.4915C309.221 22.7528 309.354 22.1081 309.622 21.5572C309.893 21.0063 310.268 20.5785 310.749 20.2738C311.232 19.9691 311.793 19.8168 312.434 19.8168C313.074 19.8168 313.634 19.9691 314.114 20.2738C314.597 20.5785 314.973 21.0063 315.24 21.5572C315.511 22.1081 315.647 22.7528 315.647 23.4915C315.647 24.224 315.511 24.8641 315.24 25.4119C314.973 25.9598 314.597 26.386 314.114 26.6907C313.634 26.9954 313.074 27.1477 312.434 27.1477ZM312.434 26.169C312.92 26.169 313.32 26.0444 313.634 25.7951C313.948 25.5458 314.18 25.218 314.331 24.8118C314.482 24.4055 314.557 23.9654 314.557 23.4915C314.557 23.0175 314.482 22.5759 314.331 22.1665C314.18 21.7572 313.948 21.4264 313.634 21.174C313.32 20.9216 312.92 20.7955 312.434 20.7955C311.947 20.7955 311.547 20.9216 311.233 21.174C310.919 21.4264 310.687 21.7572 310.536 22.1665C310.385 22.5759 310.31 23.0175 310.31 23.4915C310.31 23.9654 310.385 24.4055 310.536 24.8118C310.687 25.218 310.919 25.5458 311.233 25.7951C311.547 26.0444 311.947 26.169 312.434 26.169ZM320.888 27V19.9091H321.978V27H320.888ZM321.442 18.7273C321.23 18.7273 321.047 18.6549 320.893 18.5103C320.742 18.3656 320.666 18.1918 320.666 17.9886C320.666 17.7855 320.742 17.6116 320.893 17.467C321.047 17.3223 321.23 17.25 321.442 17.25C321.654 17.25 321.836 17.3223 321.987 17.467C322.141 17.6116 322.218 17.7855 322.218 17.9886C322.218 18.1918 322.141 18.3656 321.987 18.5103C321.836 18.6549 321.654 18.7273 321.442 18.7273ZM323.934 27V19.9091H324.987V21.017H325.079C325.227 20.6385 325.465 20.3446 325.794 20.1353C326.124 19.9229 326.519 19.8168 326.981 19.8168C327.449 19.8168 327.838 19.9229 328.149 20.1353C328.463 20.3446 328.707 20.6385 328.883 21.017H328.957C329.138 20.6508 329.411 20.36 329.774 20.1445C330.137 19.926 330.572 19.8168 331.08 19.8168C331.714 19.8168 332.233 20.0153 332.636 20.4123C333.039 20.8062 333.241 21.4202 333.241 22.2543V27H332.151V22.2543C332.151 21.7311 332.008 21.3571 331.722 21.1325C331.436 20.9078 331.099 20.7955 330.711 20.7955C330.212 20.7955 329.826 20.9463 329.552 21.2479C329.278 21.5464 329.141 21.925 329.141 22.3835V27H328.033V22.1435C328.033 21.7403 327.903 21.4156 327.641 21.1694C327.379 20.9201 327.042 20.7955 326.63 20.7955C326.347 20.7955 326.082 20.8709 325.836 21.0217C325.593 21.1725 325.396 21.3817 325.245 21.6495C325.097 21.9142 325.023 22.2204 325.023 22.5682V27H323.934ZM335.194 29.6591V19.9091H336.246V21.0355H336.376C336.456 20.9124 336.566 20.7554 336.708 20.5646C336.853 20.3707 337.059 20.1984 337.327 20.0476C337.597 19.8937 337.964 19.8168 338.425 19.8168C339.022 19.8168 339.549 19.966 340.004 20.2646C340.46 20.5631 340.815 20.9863 341.071 21.5341C341.326 22.0819 341.454 22.7282 341.454 23.473C341.454 24.224 341.326 24.8749 341.071 25.4258C340.815 25.9736 340.461 26.3983 340.009 26.6999C339.556 26.9985 339.035 27.1477 338.444 27.1477C337.988 27.1477 337.624 27.0723 337.35 26.9215C337.076 26.7676 336.865 26.5937 336.717 26.3999C336.57 26.2029 336.456 26.0398 336.376 25.9105H336.283V29.6591H335.194ZM336.265 23.4545C336.265 23.9901 336.343 24.4625 336.5 24.8718C336.657 25.2781 336.887 25.5966 337.188 25.8274C337.49 26.0552 337.859 26.169 338.296 26.169C338.752 26.169 339.132 26.049 339.436 25.8089C339.744 25.5658 339.975 25.2396 340.129 24.8303C340.286 24.4179 340.364 23.9593 340.364 23.4545C340.364 22.956 340.287 22.5066 340.133 22.1065C339.983 21.7034 339.753 21.3848 339.446 21.1509C339.141 20.9139 338.758 20.7955 338.296 20.7955C337.853 20.7955 337.481 20.9078 337.179 21.1325C336.877 21.354 336.65 21.6649 336.496 22.065C336.342 22.462 336.265 22.9252 336.265 23.4545ZM343.077 27V19.9091H344.129V20.9801H344.203C344.332 20.6293 344.566 20.3446 344.905 20.1261C345.243 19.9076 345.625 19.7983 346.05 19.7983C346.13 19.7983 346.23 19.7998 346.35 19.8029C346.47 19.806 346.561 19.8106 346.622 19.8168V20.9247C346.585 20.9155 346.501 20.9016 346.368 20.8832C346.239 20.8616 346.102 20.8509 345.957 20.8509C345.613 20.8509 345.305 20.9232 345.034 21.0678C344.766 21.2094 344.554 21.4064 344.397 21.6587C344.243 21.908 344.166 22.1927 344.166 22.5128V27H343.077ZM350.539 27.1477C349.899 27.1477 349.338 26.9954 348.854 26.6907C348.374 26.386 347.999 25.9598 347.728 25.4119C347.46 24.8641 347.326 24.224 347.326 23.4915C347.326 22.7528 347.46 22.1081 347.728 21.5572C347.999 21.0063 348.374 20.5785 348.854 20.2738C349.338 19.9691 349.899 19.8168 350.539 19.8168C351.18 19.8168 351.74 19.9691 352.22 20.2738C352.703 20.5785 353.079 21.0063 353.346 21.5572C353.617 22.1081 353.753 22.7528 353.753 23.4915C353.753 24.224 353.617 24.8641 353.346 25.4119C353.079 25.9598 352.703 26.386 352.22 26.6907C351.74 26.9954 351.18 27.1477 350.539 27.1477ZM350.539 26.169C351.026 26.169 351.426 26.0444 351.74 25.7951C352.054 25.5458 352.286 25.218 352.437 24.8118C352.588 24.4055 352.663 23.9654 352.663 23.4915C352.663 23.0175 352.588 22.5759 352.437 22.1665C352.286 21.7572 352.054 21.4264 351.74 21.174C351.426 20.9216 351.026 20.7955 350.539 20.7955C350.053 20.7955 349.653 20.9216 349.339 21.174C349.025 21.4264 348.793 21.7572 348.642 22.1665C348.491 22.5759 348.416 23.0175 348.416 23.4915C348.416 23.9654 348.491 24.4055 348.642 24.8118C348.793 25.218 349.025 25.5458 349.339 25.7951C349.653 26.0444 350.053 26.169 350.539 26.169ZM360.921 19.9091L358.299 27H357.191L354.569 19.9091H355.751L357.708 25.5597H357.782L359.739 19.9091H360.921ZM365.039 27.1477C364.356 27.1477 363.767 26.9969 363.271 26.6953C362.779 26.3906 362.399 25.9659 362.131 25.4212C361.866 24.8733 361.734 24.2363 361.734 23.5099C361.734 22.7836 361.866 22.1435 362.131 21.5895C362.399 21.0324 362.771 20.5985 363.248 20.2876C363.728 19.9737 364.288 19.8168 364.928 19.8168C365.298 19.8168 365.662 19.8783 366.022 20.0014C366.383 20.1245 366.71 20.3246 367.006 20.6016C367.301 20.8755 367.537 21.2386 367.712 21.6911C367.888 22.1435 367.975 22.7005 367.975 23.3622V23.8239H362.509V22.8821H366.867C366.867 22.482 366.787 22.125 366.627 21.8111C366.47 21.4972 366.246 21.2494 365.953 21.0678C365.664 20.8862 365.322 20.7955 364.928 20.7955C364.494 20.7955 364.119 20.9032 363.802 21.1186C363.488 21.331 363.246 21.608 363.077 21.9496C362.908 22.2912 362.823 22.6574 362.823 23.0483V23.6761C362.823 24.2116 362.916 24.6656 363.1 25.038C363.288 25.4073 363.548 25.6889 363.88 25.8828C364.213 26.0736 364.599 26.169 365.039 26.169C365.325 26.169 365.584 26.129 365.815 26.049C366.049 25.9659 366.25 25.8428 366.419 25.6797C366.589 25.5135 366.72 25.3073 366.812 25.0611L367.864 25.3565C367.754 25.7135 367.567 26.0275 367.306 26.2983C367.044 26.5661 366.721 26.7753 366.336 26.9261C365.952 27.0739 365.519 27.1477 365.039 27.1477ZM373.211 27V19.9091H374.263V20.9801H374.337C374.466 20.6293 374.7 20.3446 375.039 20.1261C375.377 19.9076 375.759 19.7983 376.184 19.7983C376.264 19.7983 376.364 19.7998 376.484 19.8029C376.604 19.806 376.695 19.8106 376.756 19.8168V20.9247C376.719 20.9155 376.635 20.9016 376.502 20.8832C376.373 20.8616 376.236 20.8509 376.091 20.8509C375.747 20.8509 375.439 20.9232 375.168 21.0678C374.9 21.2094 374.688 21.4064 374.531 21.6587C374.377 21.908 374.3 22.1927 374.3 22.5128V27H373.211ZM380.766 27.1477C380.083 27.1477 379.493 26.9969 378.998 26.6953C378.505 26.3906 378.125 25.9659 377.858 25.4212C377.593 24.8733 377.461 24.2363 377.461 23.5099C377.461 22.7836 377.593 22.1435 377.858 21.5895C378.125 21.0324 378.498 20.5985 378.975 20.2876C379.455 19.9737 380.015 19.8168 380.655 19.8168C381.024 19.8168 381.389 19.8783 381.749 20.0014C382.109 20.1245 382.437 20.3246 382.733 20.6016C383.028 20.8755 383.263 21.2386 383.439 21.6911C383.614 22.1435 383.702 22.7005 383.702 23.3622V23.8239H378.236V22.8821H382.594C382.594 22.482 382.514 22.125 382.354 21.8111C382.197 21.4972 381.972 21.2494 381.68 21.0678C381.391 20.8862 381.049 20.7955 380.655 20.7955C380.221 20.7955 379.846 20.9032 379.529 21.1186C379.215 21.331 378.973 21.608 378.804 21.9496C378.635 22.2912 378.55 22.6574 378.55 23.0483V23.6761C378.55 24.2116 378.642 24.6656 378.827 25.038C379.015 25.4073 379.275 25.6889 379.607 25.8828C379.94 26.0736 380.326 26.169 380.766 26.169C381.052 26.169 381.311 26.129 381.541 26.049C381.775 25.9659 381.977 25.8428 382.146 25.6797C382.316 25.5135 382.446 25.3073 382.539 25.0611L383.591 25.3565C383.48 25.7135 383.294 26.0275 383.033 26.2983C382.771 26.5661 382.448 26.7753 382.063 26.9261C381.678 27.0739 381.246 27.1477 380.766 27.1477ZM390.343 21.4972L389.364 21.7741C389.303 21.611 389.212 21.4525 389.092 21.2987C388.975 21.1417 388.815 21.0124 388.612 20.9109C388.409 20.8093 388.149 20.7585 387.832 20.7585C387.398 20.7585 387.036 20.8585 386.747 21.0586C386.461 21.2556 386.317 21.5064 386.317 21.8111C386.317 22.0819 386.416 22.2958 386.613 22.4528C386.81 22.6097 387.118 22.7405 387.536 22.8452L388.589 23.1037C389.223 23.2576 389.695 23.493 390.006 23.81C390.317 24.1239 390.472 24.5286 390.472 25.0241C390.472 25.4304 390.355 25.7936 390.121 26.1136C389.891 26.4337 389.567 26.6861 389.152 26.8707C388.737 27.0554 388.253 27.1477 387.702 27.1477C386.979 27.1477 386.381 26.9908 385.907 26.6768C385.433 26.3629 385.133 25.9044 385.006 25.3011L386.04 25.0426C386.139 25.4242 386.325 25.7105 386.599 25.9013C386.876 26.0921 387.238 26.1875 387.684 26.1875C388.192 26.1875 388.595 26.0798 388.893 25.8643C389.195 25.6458 389.346 25.3842 389.346 25.0795C389.346 24.8333 389.26 24.6271 389.087 24.4609C388.915 24.2917 388.65 24.1655 388.293 24.0824L387.112 23.8054C386.462 23.6515 385.985 23.413 385.68 23.0898C385.379 22.7636 385.228 22.3558 385.228 21.8665C385.228 21.4664 385.34 21.1125 385.565 20.8047C385.793 20.4969 386.102 20.2553 386.493 20.0799C386.887 19.9045 387.333 19.8168 387.832 19.8168C388.533 19.8168 389.084 19.9706 389.484 20.2784C389.888 20.5862 390.174 20.9924 390.343 21.4972ZM396.542 24.1009V19.9091H397.632V27H396.542V25.7997H396.468C396.302 26.1598 396.043 26.466 395.693 26.7184C395.342 26.9677 394.899 27.0923 394.363 27.0923C393.92 27.0923 393.526 26.9954 393.181 26.8015C392.837 26.6045 392.566 26.3091 392.369 25.9151C392.172 25.5181 392.073 25.018 392.073 24.4148V19.9091H393.163V24.3409C393.163 24.858 393.307 25.2704 393.597 25.5781C393.889 25.8859 394.262 26.0398 394.714 26.0398C394.985 26.0398 395.26 25.9705 395.54 25.832C395.823 25.6935 396.06 25.4812 396.251 25.195C396.445 24.9087 396.542 24.544 396.542 24.1009ZM400.677 17.5455V27H399.588V17.5455H400.677ZM405.718 19.9091V20.8324H402.043V19.9091H405.718ZM403.114 18.2102H404.204V24.9688C404.204 25.2765 404.248 25.5073 404.337 25.6612C404.43 25.812 404.547 25.9136 404.688 25.9659C404.833 26.0152 404.985 26.0398 405.145 26.0398C405.265 26.0398 405.364 26.0336 405.441 26.0213C405.518 26.0059 405.579 25.9936 405.625 25.9844L405.847 26.9631C405.773 26.9908 405.67 27.0185 405.538 27.0462C405.405 27.0769 405.238 27.0923 405.035 27.0923C404.727 27.0923 404.425 27.0262 404.13 26.8938C403.837 26.7615 403.594 26.5599 403.4 26.2891C403.209 26.0182 403.114 25.6766 403.114 25.2642V18.2102ZM412.213 21.4972L411.235 21.7741C411.173 21.611 411.082 21.4525 410.962 21.2987C410.845 21.1417 410.685 21.0124 410.482 20.9109C410.279 20.8093 410.019 20.7585 409.702 20.7585C409.268 20.7585 408.906 20.8585 408.617 21.0586C408.331 21.2556 408.188 21.5064 408.188 21.8111C408.188 22.0819 408.286 22.2958 408.483 22.4528C408.68 22.6097 408.988 22.7405 409.407 22.8452L410.459 23.1037C411.093 23.2576 411.566 23.493 411.876 23.81C412.187 24.1239 412.343 24.5286 412.343 25.0241C412.343 25.4304 412.226 25.7936 411.992 26.1136C411.761 26.4337 411.438 26.6861 411.022 26.8707C410.607 27.0554 410.124 27.1477 409.573 27.1477C408.85 27.1477 408.251 26.9908 407.777 26.6768C407.303 26.3629 407.003 25.9044 406.877 25.3011L407.911 25.0426C408.009 25.4242 408.196 25.7105 408.469 25.9013C408.746 26.0921 409.108 26.1875 409.554 26.1875C410.062 26.1875 410.465 26.0798 410.764 25.8643C411.065 25.6458 411.216 25.3842 411.216 25.0795C411.216 24.8333 411.13 24.6271 410.958 24.4609C410.785 24.2917 410.521 24.1655 410.164 24.0824L408.982 23.8054C408.332 23.6515 407.855 23.413 407.551 23.0898C407.249 22.7636 407.098 22.3558 407.098 21.8665C407.098 21.4664 407.211 21.1125 407.435 20.8047C407.663 20.4969 407.972 20.2553 408.363 20.0799C408.757 19.9045 409.203 19.8168 409.702 19.8168C410.404 19.8168 410.955 19.9706 411.355 20.2784C411.758 20.5862 412.044 20.9924 412.213 21.4972ZM414.738 27.0739C414.51 27.0739 414.315 26.9923 414.151 26.8292C413.988 26.6661 413.907 26.4706 413.907 26.2429C413.907 26.0152 413.988 25.8197 414.151 25.6566C414.315 25.4935 414.51 25.4119 414.738 25.4119C414.965 25.4119 415.161 25.4935 415.324 25.6566C415.487 25.8197 415.569 26.0152 415.569 26.2429C415.569 26.3937 415.53 26.5322 415.453 26.6584C415.379 26.7846 415.279 26.8861 415.153 26.9631C415.03 27.0369 414.892 27.0739 414.738 27.0739ZM424.167 27H421.25V17.5455H424.297C425.214 17.5455 425.999 17.7347 426.651 18.1133C427.303 18.4888 427.804 19.0289 428.151 19.7337C428.499 20.4354 428.673 21.2756 428.673 22.2543C428.673 23.2391 428.498 24.087 428.147 24.7979C427.796 25.5058 427.285 26.0505 426.614 26.4322C425.943 26.8107 425.128 27 424.167 27ZM422.395 25.9844H424.093C424.875 25.9844 425.523 25.8336 426.037 25.532C426.551 25.2304 426.934 24.801 427.187 24.244C427.439 23.6869 427.565 23.0237 427.565 22.2543C427.565 21.491 427.44 20.8339 427.191 20.283C426.942 19.729 426.569 19.3043 426.074 19.0089C425.578 18.7103 424.961 18.5611 424.223 18.5611H422.395V25.9844ZM433.287 27.1477C432.647 27.1477 432.086 26.9954 431.602 26.6907C431.122 26.386 430.747 25.9598 430.476 25.4119C430.208 24.8641 430.074 24.224 430.074 23.4915C430.074 22.7528 430.208 22.1081 430.476 21.5572C430.747 21.0063 431.122 20.5785 431.602 20.2738C432.086 19.9691 432.647 19.8168 433.287 19.8168C433.928 19.8168 434.488 19.9691 434.968 20.2738C435.451 20.5785 435.826 21.0063 436.094 21.5572C436.365 22.1081 436.5 22.7528 436.5 23.4915C436.5 24.224 436.365 24.8641 436.094 25.4119C435.826 25.9598 435.451 26.386 434.968 26.6907C434.488 26.9954 433.928 27.1477 433.287 27.1477ZM433.287 26.169C433.774 26.169 434.174 26.0444 434.488 25.7951C434.802 25.5458 435.034 25.218 435.185 24.8118C435.336 24.4055 435.411 23.9654 435.411 23.4915C435.411 23.0175 435.336 22.5759 435.185 22.1665C435.034 21.7572 434.802 21.4264 434.488 21.174C434.174 20.9216 433.774 20.7955 433.287 20.7955C432.801 20.7955 432.401 20.9216 432.087 21.174C431.773 21.4264 431.541 21.7572 431.39 22.1665C431.239 22.5759 431.164 23.0175 431.164 23.4915C431.164 23.9654 431.239 24.4055 431.39 24.8118C431.541 25.218 431.773 25.5458 432.087 25.7951C432.401 26.0444 432.801 26.169 433.287 26.169ZM439.49 27L437.329 19.9091H438.474L440.007 25.3381H440.081L441.595 19.9091H442.758L444.254 25.3196H444.328L445.861 19.9091H447.006L444.845 27H443.774L442.223 21.5526H442.112L440.561 27H439.49ZM449.496 22.7344V27H448.407V19.9091H449.459V21.017H449.552C449.718 20.657 449.97 20.3677 450.309 20.1491C450.647 19.9276 451.084 19.8168 451.62 19.8168C452.1 19.8168 452.52 19.9152 452.88 20.1122C453.24 20.3061 453.52 20.6016 453.72 20.9986C453.92 21.3925 454.02 21.8911 454.02 22.4943V27H452.931V22.5682C452.931 22.0111 452.786 21.5772 452.497 21.2663C452.208 20.9524 451.811 20.7955 451.306 20.7955C450.958 20.7955 450.647 20.8709 450.373 21.0217C450.103 21.1725 449.889 21.3925 449.732 21.6818C449.575 21.9711 449.496 22.322 449.496 22.7344ZM457.062 17.5455V27H455.972V17.5455H457.062ZM461.899 27.1477C461.259 27.1477 460.697 26.9954 460.214 26.6907C459.734 26.386 459.358 25.9598 459.088 25.4119C458.82 24.8641 458.686 24.224 458.686 23.4915C458.686 22.7528 458.82 22.1081 459.088 21.5572C459.358 21.0063 459.734 20.5785 460.214 20.2738C460.697 19.9691 461.259 19.8168 461.899 19.8168C462.539 19.8168 463.099 19.9691 463.579 20.2738C464.063 20.5785 464.438 21.0063 464.706 21.5572C464.977 22.1081 465.112 22.7528 465.112 23.4915C465.112 24.224 464.977 24.8641 464.706 25.4119C464.438 25.9598 464.063 26.386 463.579 26.6907C463.099 26.9954 462.539 27.1477 461.899 27.1477ZM461.899 26.169C462.385 26.169 462.785 26.0444 463.099 25.7951C463.413 25.5458 463.646 25.218 463.796 24.8118C463.947 24.4055 464.023 23.9654 464.023 23.4915C464.023 23.0175 463.947 22.5759 463.796 22.1665C463.646 21.7572 463.413 21.4264 463.099 21.174C462.785 20.9216 462.385 20.7955 461.899 20.7955C461.413 20.7955 461.013 20.9216 460.699 21.174C460.385 21.4264 460.152 21.7572 460.002 22.1665C459.851 22.5759 459.775 23.0175 459.775 23.4915C459.775 23.9654 459.851 24.4055 460.002 24.8118C460.152 25.218 460.385 25.5458 460.699 25.7951C461.013 26.0444 461.413 26.169 461.899 26.169ZM468.823 27.1662C468.373 27.1662 467.966 27.0816 467.599 26.9123C467.233 26.7399 466.942 26.4922 466.727 26.169C466.512 25.8428 466.404 25.4489 466.404 24.9872C466.404 24.581 466.484 24.2517 466.644 23.9993C466.804 23.7438 467.018 23.5438 467.286 23.3991C467.553 23.2545 467.849 23.1468 468.172 23.076C468.498 23.0021 468.826 22.9437 469.155 22.9006C469.586 22.8452 469.935 22.8036 470.203 22.7759C470.474 22.7451 470.671 22.6944 470.794 22.6236C470.92 22.5528 470.983 22.4297 470.983 22.2543V22.2173C470.983 21.7618 470.859 21.4079 470.609 21.1555C470.363 20.9032 469.989 20.777 469.488 20.777C468.967 20.777 468.56 20.8909 468.264 21.1186C467.969 21.3464 467.761 21.5895 467.641 21.848L466.607 21.4787C466.792 21.0478 467.038 20.7124 467.346 20.4723C467.656 20.2292 467.995 20.0599 468.361 19.9645C468.73 19.866 469.094 19.8168 469.451 19.8168C469.678 19.8168 469.94 19.8445 470.235 19.8999C470.534 19.9522 470.822 20.0614 471.099 20.2276C471.379 20.3938 471.611 20.6446 471.796 20.9801C471.98 21.3156 472.073 21.7649 472.073 22.3281V27H470.983V26.0398H470.928C470.854 26.1937 470.731 26.3583 470.559 26.5337C470.386 26.7092 470.157 26.8584 469.871 26.9815C469.585 27.1046 469.235 27.1662 468.823 27.1662ZM468.989 26.1875C469.42 26.1875 469.783 26.1029 470.079 25.9336C470.377 25.7643 470.602 25.5458 470.753 25.2781C470.906 25.0103 470.983 24.7287 470.983 24.4332V23.4361C470.937 23.4915 470.836 23.5423 470.679 23.5884C470.525 23.6315 470.346 23.67 470.143 23.7038C469.943 23.7346 469.748 23.7623 469.557 23.7869C469.369 23.8085 469.217 23.8269 469.1 23.8423C468.817 23.8793 468.552 23.9393 468.306 24.0224C468.063 24.1024 467.866 24.224 467.715 24.3871C467.567 24.5471 467.493 24.7656 467.493 25.0426C467.493 25.4212 467.633 25.7074 467.913 25.9013C468.197 26.0921 468.555 26.1875 468.989 26.1875ZM476.7 27.1477C476.109 27.1477 475.587 26.9985 475.135 26.6999C474.683 26.3983 474.329 25.9736 474.073 25.4258C473.818 24.8749 473.69 24.224 473.69 23.473C473.69 22.7282 473.818 22.0819 474.073 21.5341C474.329 20.9863 474.684 20.5631 475.14 20.2646C475.595 19.966 476.121 19.8168 476.718 19.8168C477.18 19.8168 477.545 19.8937 477.812 20.0476C478.083 20.1984 478.29 20.3707 478.431 20.5646C478.576 20.7554 478.688 20.9124 478.768 21.0355H478.86V17.5455H479.95V27H478.897V25.9105H478.768C478.688 26.0398 478.574 26.2029 478.426 26.3999C478.279 26.5937 478.068 26.7676 477.794 26.9215C477.52 27.0723 477.155 27.1477 476.7 27.1477ZM476.848 26.169C477.285 26.169 477.654 26.0552 477.956 25.8274C478.257 25.5966 478.487 25.2781 478.643 24.8718C478.8 24.4625 478.879 23.9901 478.879 23.4545C478.879 22.9252 478.802 22.462 478.648 22.065C478.494 21.6649 478.266 21.354 477.965 21.1325C477.663 20.9078 477.291 20.7955 476.848 20.7955C476.386 20.7955 476.001 20.9139 475.694 21.1509C475.389 21.3848 475.16 21.7034 475.006 22.1065C474.855 22.5066 474.779 22.956 474.779 23.4545C474.779 23.9593 474.856 24.4179 475.01 24.8303C475.167 25.2396 475.398 25.5658 475.703 25.8089C476.011 26.049 476.392 26.169 476.848 26.169ZM485.031 27.1477C484.347 27.1477 483.758 26.9969 483.262 26.6953C482.77 26.3906 482.39 25.9659 482.122 25.4212C481.858 24.8733 481.725 24.2363 481.725 23.5099C481.725 22.7836 481.858 22.1435 482.122 21.5895C482.39 21.0324 482.762 20.5985 483.239 20.2876C483.72 19.9737 484.28 19.8168 484.92 19.8168C485.289 19.8168 485.654 19.8783 486.014 20.0014C486.374 20.1245 486.702 20.3246 486.997 20.6016C487.293 20.8755 487.528 21.2386 487.704 21.6911C487.879 22.1435 487.967 22.7005 487.967 23.3622V23.8239H482.501V22.8821H486.859C486.859 22.482 486.779 22.125 486.619 21.8111C486.462 21.4972 486.237 21.2494 485.945 21.0678C485.655 20.8862 485.314 20.7955 484.92 20.7955C484.486 20.7955 484.11 20.9032 483.793 21.1186C483.479 21.331 483.238 21.608 483.069 21.9496C482.899 22.2912 482.815 22.6574 482.815 23.0483V23.6761C482.815 24.2116 482.907 24.6656 483.092 25.038C483.279 25.4073 483.539 25.6889 483.872 25.8828C484.204 26.0736 484.59 26.169 485.031 26.169C485.317 26.169 485.575 26.129 485.806 26.049C486.04 25.9659 486.242 25.8428 486.411 25.6797C486.58 25.5135 486.711 25.3073 486.803 25.0611L487.856 25.3565C487.745 25.7135 487.559 26.0275 487.297 26.2983C487.036 26.5661 486.713 26.7753 486.328 26.9261C485.943 27.0739 485.511 27.1477 485.031 27.1477ZM492.263 27.1477C491.672 27.1477 491.15 26.9985 490.698 26.6999C490.245 26.3983 489.891 25.9736 489.636 25.4258C489.38 24.8749 489.253 24.224 489.253 23.473C489.253 22.7282 489.38 22.0819 489.636 21.5341C489.891 20.9863 490.247 20.5631 490.702 20.2646C491.158 19.966 491.684 19.8168 492.281 19.8168C492.743 19.8168 493.107 19.8937 493.375 20.0476C493.646 20.1984 493.852 20.3707 493.994 20.5646C494.138 20.7554 494.251 20.9124 494.331 21.0355H494.423V17.5455H495.513V27H494.46V25.9105H494.331C494.251 26.0398 494.137 26.2029 493.989 26.3999C493.841 26.5937 493.631 26.7676 493.357 26.9215C493.083 27.0723 492.718 27.1477 492.263 27.1477ZM492.41 26.169C492.847 26.169 493.217 26.0552 493.518 25.8274C493.82 25.5966 494.049 25.2781 494.206 24.8718C494.363 24.4625 494.442 23.9901 494.442 23.4545C494.442 22.9252 494.365 22.462 494.211 22.065C494.057 21.6649 493.829 21.354 493.527 21.1325C493.226 20.9078 492.853 20.7955 492.41 20.7955C491.949 20.7955 491.564 20.9139 491.256 21.1509C490.951 21.3848 490.722 21.7034 490.568 22.1065C490.417 22.5066 490.342 22.956 490.342 23.4545C490.342 23.9593 490.419 24.4179 490.573 24.8303C490.73 25.2396 490.961 25.5658 491.265 25.8089C491.573 26.049 491.955 26.169 492.41 26.169ZM503.915 27.1477C503.324 27.1477 502.802 26.9985 502.35 26.6999C501.898 26.3983 501.544 25.9736 501.288 25.4258C501.033 24.8749 500.905 24.224 500.905 23.473C500.905 22.7282 501.033 22.0819 501.288 21.5341C501.544 20.9863 501.899 20.5631 502.355 20.2646C502.81 19.966 503.336 19.8168 503.933 19.8168C504.395 19.8168 504.76 19.8937 505.028 20.0476C505.298 20.1984 505.505 20.3707 505.646 20.5646C505.791 20.7554 505.903 20.9124 505.983 21.0355H506.076V17.5455H507.165V27H506.112V25.9105H505.983C505.903 26.0398 505.789 26.2029 505.642 26.3999C505.494 26.5937 505.283 26.7676 505.009 26.9215C504.735 27.0723 504.371 27.1477 503.915 27.1477ZM504.063 26.169C504.5 26.169 504.869 26.0552 505.171 25.8274C505.472 25.5966 505.702 25.2781 505.859 24.8718C506.016 24.4625 506.094 23.9901 506.094 23.4545C506.094 22.9252 506.017 22.462 505.863 22.065C505.709 21.6649 505.482 21.354 505.18 21.1325C504.878 20.9078 504.506 20.7955 504.063 20.7955C503.601 20.7955 503.216 20.9139 502.909 21.1509C502.604 21.3848 502.375 21.7034 502.221 22.1065C502.07 22.5066 501.995 22.956 501.995 23.4545C501.995 23.9593 502.071 24.4179 502.225 24.8303C502.382 25.2396 502.613 25.5658 502.918 25.8089C503.226 26.049 503.607 26.169 504.063 26.169ZM511.359 27.1662C510.91 27.1662 510.502 27.0816 510.136 26.9123C509.77 26.7399 509.479 26.4922 509.263 26.169C509.048 25.8428 508.94 25.4489 508.94 24.9872C508.94 24.581 509.02 24.2517 509.18 23.9993C509.34 23.7438 509.554 23.5438 509.822 23.3991C510.09 23.2545 510.385 23.1468 510.708 23.076C511.035 23.0021 511.362 22.9437 511.692 22.9006C512.123 22.8452 512.472 22.8036 512.74 22.7759C513.01 22.7451 513.207 22.6944 513.331 22.6236C513.457 22.5528 513.52 22.4297 513.52 22.2543V22.2173C513.52 21.7618 513.395 21.4079 513.146 21.1555C512.9 20.9032 512.526 20.777 512.024 20.777C511.504 20.777 511.096 20.8909 510.801 21.1186C510.505 21.3464 510.298 21.5895 510.178 21.848L509.143 21.4787C509.328 21.0478 509.574 20.7124 509.882 20.4723C510.193 20.2292 510.531 20.0599 510.898 19.9645C511.267 19.866 511.63 19.8168 511.987 19.8168C512.215 19.8168 512.477 19.8445 512.772 19.8999C513.07 19.9522 513.358 20.0614 513.635 20.2276C513.915 20.3938 514.148 20.6446 514.332 20.9801C514.517 21.3156 514.609 21.7649 514.609 22.3281V27H513.52V26.0398H513.464C513.391 26.1937 513.267 26.3583 513.095 26.5337C512.923 26.7092 512.693 26.8584 512.407 26.9815C512.121 27.1046 511.772 27.1662 511.359 27.1662ZM511.526 26.1875C511.956 26.1875 512.32 26.1029 512.615 25.9336C512.914 25.7643 513.138 25.5458 513.289 25.2781C513.443 25.0103 513.52 24.7287 513.52 24.4332V23.4361C513.474 23.4915 513.372 23.5423 513.215 23.5884C513.061 23.6315 512.883 23.67 512.68 23.7038C512.48 23.7346 512.284 23.7623 512.093 23.7869C511.906 23.8085 511.753 23.8269 511.636 23.8423C511.353 23.8793 511.088 23.9393 510.842 24.0224C510.599 24.1024 510.402 24.224 510.251 24.3871C510.104 24.5471 510.03 24.7656 510.03 25.0426C510.03 25.4212 510.17 25.7074 510.45 25.9013C510.733 26.0921 511.092 26.1875 511.526 26.1875ZM519.643 19.9091V20.8324H515.968V19.9091H519.643ZM517.039 18.2102H518.128V24.9688C518.128 25.2765 518.173 25.5073 518.262 25.6612C518.355 25.812 518.472 25.9136 518.613 25.9659C518.758 26.0152 518.91 26.0398 519.07 26.0398C519.19 26.0398 519.289 26.0336 519.366 26.0213C519.443 26.0059 519.504 25.9936 519.55 25.9844L519.772 26.9631C519.698 26.9908 519.595 27.0185 519.463 27.0462C519.33 27.0769 519.163 27.0923 518.959 27.0923C518.652 27.0923 518.35 27.0262 518.055 26.8938C517.762 26.7615 517.519 26.5599 517.325 26.2891C517.134 26.0182 517.039 25.6766 517.039 25.2642V18.2102ZM523.329 27.1662C522.88 27.1662 522.472 27.0816 522.106 26.9123C521.74 26.7399 521.449 26.4922 521.233 26.169C521.018 25.8428 520.91 25.4489 520.91 24.9872C520.91 24.581 520.99 24.2517 521.15 23.9993C521.31 23.7438 521.524 23.5438 521.792 23.3991C522.06 23.2545 522.355 23.1468 522.678 23.076C523.004 23.0021 523.332 22.9437 523.662 22.9006C524.092 22.8452 524.442 22.8036 524.709 22.7759C524.98 22.7451 525.177 22.6944 525.3 22.6236C525.427 22.5528 525.49 22.4297 525.49 22.2543V22.2173C525.49 21.7618 525.365 21.4079 525.116 21.1555C524.87 20.9032 524.496 20.777 523.994 20.777C523.474 20.777 523.066 20.8909 522.771 21.1186C522.475 21.3464 522.267 21.5895 522.147 21.848L521.113 21.4787C521.298 21.0478 521.544 20.7124 521.852 20.4723C522.163 20.2292 522.501 20.0599 522.868 19.9645C523.237 19.866 523.6 19.8168 523.957 19.8168C524.185 19.8168 524.446 19.8445 524.742 19.8999C525.04 19.9522 525.328 20.0614 525.605 20.2276C525.885 20.3938 526.118 20.6446 526.302 20.9801C526.487 21.3156 526.579 21.7649 526.579 22.3281V27H525.49V26.0398H525.434C525.36 26.1937 525.237 26.3583 525.065 26.5337C524.893 26.7092 524.663 26.8584 524.377 26.9815C524.091 27.1046 523.742 27.1662 523.329 27.1662ZM523.495 26.1875C523.926 26.1875 524.289 26.1029 524.585 25.9336C524.883 25.7643 525.108 25.5458 525.259 25.2781C525.413 25.0103 525.49 24.7287 525.49 24.4332V23.4361C525.444 23.4915 525.342 23.5423 525.185 23.5884C525.031 23.6315 524.853 23.67 524.649 23.7038C524.449 23.7346 524.254 23.7623 524.063 23.7869C523.875 23.8085 523.723 23.8269 523.606 23.8423C523.323 23.8793 523.058 23.9393 522.812 24.0224C522.569 24.1024 522.372 24.224 522.221 24.3871C522.073 24.5471 522 24.7656 522 25.0426C522 25.4212 522.14 25.7074 522.42 25.9013C522.703 26.0921 523.061 26.1875 523.495 26.1875ZM15.0838 47.1662C14.6345 47.1662 14.2267 47.0816 13.8604 46.9123C13.4942 46.7399 13.2034 46.4922 12.9879 46.169C12.7725 45.8428 12.6648 45.4489 12.6648 44.9872C12.6648 44.581 12.7448 44.2517 12.9048 43.9993C13.0649 43.7438 13.2788 43.5438 13.5465 43.3991C13.8143 43.2545 14.1097 43.1468 14.4329 43.076C14.7591 43.0021 15.0869 42.9437 15.4162 42.9006C15.8471 42.8452 16.1964 42.8036 16.4641 42.7759C16.735 42.7451 16.9319 42.6944 17.055 42.6236C17.1812 42.5528 17.2443 42.4297 17.2443 42.2543V42.2173C17.2443 41.7618 17.1197 41.4079 16.8704 41.1555C16.6242 40.9032 16.2502 40.777 15.7486 40.777C15.2285 40.777 14.8207 40.8909 14.5252 41.1186C14.2298 41.3464 14.022 41.5895 13.902 41.848L12.8679 41.4787C13.0526 41.0478 13.2988 40.7124 13.6065 40.4723C13.9174 40.2292 14.2559 40.0599 14.6222 39.9645C14.9915 39.866 15.3546 39.8168 15.7116 39.8168C15.9394 39.8168 16.201 39.8445 16.4964 39.8999C16.795 39.9522 17.0827 40.0614 17.3597 40.2276C17.6398 40.3938 17.8722 40.6446 18.0568 40.9801C18.2415 41.3156 18.3338 41.7649 18.3338 42.3281V47H17.2443V46.0398H17.1889C17.1151 46.1937 16.992 46.3583 16.8196 46.5337C16.6473 46.7092 16.418 46.8584 16.1317 46.9815C15.8455 47.1046 15.4962 47.1662 15.0838 47.1662ZM15.25 46.1875C15.6809 46.1875 16.044 46.1029 16.3395 45.9336C16.638 45.7643 16.8627 45.5458 17.0135 45.2781C17.1674 45.0103 17.2443 44.7287 17.2443 44.4332V43.4361C17.1982 43.4915 17.0966 43.5423 16.9396 43.5884C16.7857 43.6315 16.6072 43.67 16.4041 43.7038C16.2041 43.7346 16.0086 43.7623 15.8178 43.7869C15.6301 43.8085 15.4777 43.8269 15.3608 43.8423C15.0777 43.8793 14.813 43.9393 14.5668 44.0224C14.3236 44.1024 14.1267 44.224 13.9759 44.3871C13.8281 44.5471 13.7543 44.7656 13.7543 45.0426C13.7543 45.4212 13.8943 45.7074 14.1744 45.9013C14.4575 46.0921 14.8161 46.1875 15.25 46.1875ZM21.3728 42.7344V47H20.2834V39.9091H21.3359V41.017H21.4282C21.5944 40.657 21.8468 40.3677 22.1853 40.1491C22.5239 39.9276 22.9609 39.8168 23.4964 39.8168C23.9765 39.8168 24.3966 39.9152 24.7567 40.1122C25.1168 40.3061 25.3969 40.6016 25.5969 40.9986C25.797 41.3925 25.897 41.8911 25.897 42.4943V47H24.8075V42.5682C24.8075 42.0111 24.6629 41.5772 24.3736 41.2663C24.0843 40.9524 23.6872 40.7955 23.1825 40.7955C22.8347 40.7955 22.5239 40.8709 22.25 41.0217C21.9791 41.1725 21.7652 41.3925 21.6083 41.6818C21.4513 41.9711 21.3728 42.322 21.3728 42.7344ZM30.5264 47.1477C29.9355 47.1477 29.4138 46.9985 28.9614 46.6999C28.509 46.3983 28.1551 45.9736 27.8996 45.4258C27.6442 44.8749 27.5165 44.224 27.5165 43.473C27.5165 42.7282 27.6442 42.0819 27.8996 41.5341C28.1551 40.9863 28.5105 40.5631 28.966 40.2646C29.4215 39.966 29.9478 39.8168 30.5449 39.8168C31.0065 39.8168 31.3712 39.8937 31.639 40.0476C31.9098 40.1984 32.116 40.3707 32.2576 40.5646C32.4022 40.7554 32.5146 40.9124 32.5946 41.0355H32.6869V37.5455H33.7764V47H32.7238V45.9105H32.5946C32.5146 46.0398 32.4007 46.2029 32.253 46.3999C32.1052 46.5937 31.8944 46.7676 31.6205 46.9215C31.3466 47.0723 30.9819 47.1477 30.5264 47.1477ZM30.6741 46.169C31.1112 46.169 31.4805 46.0552 31.7821 45.8274C32.0837 45.5966 32.313 45.2781 32.4699 44.8718C32.6269 44.4625 32.7054 43.9901 32.7054 43.4545C32.7054 42.9252 32.6284 42.462 32.4746 42.065C32.3207 41.6649 32.0929 41.354 31.7913 41.1325C31.4897 40.9078 31.1173 40.7955 30.6741 40.7955C30.2125 40.7955 29.8278 40.9139 29.52 41.1509C29.2153 41.3848 28.986 41.7034 28.8322 42.1065C28.6814 42.5066 28.6059 42.956 28.6059 43.4545C28.6059 43.9593 28.6829 44.4179 28.8368 44.8303C28.9937 45.2396 29.2246 45.5658 29.5292 45.8089C29.837 46.049 30.2186 46.169 30.6741 46.169ZM42.659 39.9091V40.8324H38.8365V39.9091H42.659ZM39.9814 47V38.9304C39.9814 38.5241 40.0768 38.1856 40.2677 37.9148C40.4585 37.6439 40.7062 37.4408 41.0109 37.3054C41.3156 37.17 41.6372 37.1023 41.9757 37.1023C42.2435 37.1023 42.462 37.1238 42.6313 37.1669C42.8006 37.21 42.9267 37.25 43.0098 37.2869L42.6959 38.2287C42.6405 38.2102 42.5636 38.1871 42.4651 38.1594C42.3697 38.1317 42.2435 38.1179 42.0865 38.1179C41.7265 38.1179 41.4664 38.2087 41.3064 38.3903C41.1494 38.5719 41.0709 38.8381 41.0709 39.1889V47H39.9814ZM44.1469 47V39.9091H45.2364V47H44.1469ZM44.7009 38.7273C44.4885 38.7273 44.3054 38.6549 44.1515 38.5103C44.0007 38.3656 43.9253 38.1918 43.9253 37.9886C43.9253 37.7855 44.0007 37.6116 44.1515 37.467C44.3054 37.3223 44.4885 37.25 44.7009 37.25C44.9132 37.25 45.0948 37.3223 45.2456 37.467C45.3995 37.6116 45.4764 37.7855 45.4764 37.9886C45.4764 38.1918 45.3995 38.3656 45.2456 38.5103C45.0948 38.6549 44.9132 38.7273 44.7009 38.7273ZM50.0551 49.8068C49.5288 49.8068 49.0764 49.7391 48.6978 49.6037C48.3193 49.4714 48.0038 49.2959 47.7514 49.0774C47.5021 48.862 47.3036 48.6312 47.1559 48.3849L48.0238 47.7756C48.1223 47.9048 48.2469 48.0526 48.3977 48.2188C48.5486 48.388 48.7548 48.5342 49.0164 48.6573C49.281 48.7835 49.6273 48.8466 50.0551 48.8466C50.6275 48.8466 51.0999 48.7081 51.4723 48.4311C51.8447 48.1541 52.0309 47.7202 52.0309 47.1293V45.6889H51.9386C51.8586 45.8182 51.7447 45.9782 51.597 46.169C51.4523 46.3568 51.243 46.5245 50.9691 46.6722C50.6983 46.8169 50.3321 46.8892 49.8704 46.8892C49.298 46.8892 48.784 46.7538 48.3285 46.483C47.8761 46.2121 47.5175 45.8182 47.2529 45.3011C46.9913 44.7841 46.8605 44.1562 46.8605 43.4176C46.8605 42.6913 46.9882 42.0588 47.2436 41.5202C47.4991 40.9786 47.8545 40.56 48.31 40.2646C48.7655 39.966 49.2918 39.8168 49.8889 39.8168C50.3505 39.8168 50.7168 39.8937 50.9876 40.0476C51.2615 40.1984 51.4708 40.3707 51.6154 40.5646C51.7632 40.7554 51.877 40.9124 51.9571 41.0355H52.0678V39.9091H53.1204V47.2031C53.1204 47.8125 52.9819 48.308 52.7049 48.6896C52.431 49.0743 52.0617 49.3559 51.597 49.5344C51.1353 49.716 50.6213 49.8068 50.0551 49.8068ZM50.0181 45.9105C50.4552 45.9105 50.8245 45.8105 51.1261 45.6104C51.4277 45.4104 51.657 45.1226 51.8139 44.7472C51.9709 44.3717 52.0494 43.9223 52.0494 43.3991C52.0494 42.8883 51.9724 42.4374 51.8186 42.0465C51.6647 41.6557 51.4369 41.3494 51.1353 41.1278C50.8337 40.9062 50.4613 40.7955 50.0181 40.7955C49.5565 40.7955 49.1718 40.9124 48.864 41.1463C48.5593 41.3802 48.33 41.6941 48.1762 42.0881C48.0254 42.482 47.9499 42.919 47.9499 43.3991C47.9499 43.8916 48.0269 44.3271 48.1808 44.7056C48.3377 45.0811 48.5686 45.3765 48.8732 45.592C49.181 45.8043 49.5626 45.9105 50.0181 45.9105ZM59.5445 44.1009V39.9091H60.634V47H59.5445V45.7997H59.4706C59.3044 46.1598 59.0459 46.466 58.695 46.7184C58.3442 46.9677 57.901 47.0923 57.3655 47.0923C56.9223 47.0923 56.5284 46.9954 56.1837 46.8015C55.839 46.6045 55.5681 46.3091 55.3712 45.9151C55.1742 45.5181 55.0757 45.018 55.0757 44.4148V39.9091H56.1652V44.3409C56.1652 44.858 56.3099 45.2704 56.5992 45.5781C56.8915 45.8859 57.2639 46.0398 57.7163 46.0398C57.9872 46.0398 58.2626 45.9705 58.5427 45.832C58.8258 45.6935 59.0628 45.4812 59.2536 45.195C59.4475 44.9087 59.5445 44.544 59.5445 44.1009ZM62.5904 47V39.9091H63.643V40.9801H63.7169C63.8461 40.6293 64.08 40.3446 64.4186 40.1261C64.7571 39.9076 65.1387 39.7983 65.5634 39.7983C65.6435 39.7983 65.7435 39.7998 65.8635 39.8029C65.9835 39.806 66.0743 39.8106 66.1359 39.8168V40.9247C66.099 40.9155 66.0143 40.9016 65.882 40.8832C65.7527 40.8616 65.6158 40.8509 65.4711 40.8509C65.1264 40.8509 64.8187 40.9232 64.5478 41.0678C64.2801 41.2094 64.0677 41.4064 63.9107 41.6587C63.7569 41.908 63.6799 42.1927 63.6799 42.5128V47H62.5904ZM70.1455 47.1477C69.4623 47.1477 68.8729 46.9969 68.3774 46.6953C67.885 46.3906 67.5049 45.9659 67.2372 45.4212C66.9725 44.8733 66.8401 44.2363 66.8401 43.5099C66.8401 42.7836 66.9725 42.1435 67.2372 41.5895C67.5049 41.0324 67.8773 40.5985 68.3543 40.2876C68.8345 39.9737 69.3946 39.8168 70.0347 39.8168C70.4041 39.8168 70.7688 39.8783 71.1288 40.0014C71.4889 40.1245 71.8167 40.3246 72.1122 40.6016C72.4076 40.8755 72.6431 41.2386 72.8185 41.6911C72.9939 42.1435 73.0816 42.7005 73.0816 43.3622V43.8239H67.6157V42.8821H71.9737C71.9737 42.482 71.8936 42.125 71.7336 41.8111C71.5766 41.4972 71.352 41.2494 71.0596 41.0678C70.7703 40.8862 70.4287 40.7955 70.0347 40.7955C69.6008 40.7955 69.2253 40.9032 68.9083 41.1186C68.5944 41.331 68.3528 41.608 68.1835 41.9496C68.0143 42.2912 67.9296 42.6574 67.9296 43.0483V43.6761C67.9296 44.2116 68.022 44.6656 68.2066 45.038C68.3944 45.4073 68.6544 45.6889 68.9868 45.8828C69.3192 46.0736 69.7054 46.169 70.1455 46.169C70.4318 46.169 70.6903 46.129 70.9211 46.049C71.155 45.9659 71.3566 45.8428 71.5259 45.6797C71.6951 45.5135 71.8259 45.3073 71.9183 45.0611L72.9708 45.3565C72.86 45.7135 72.6738 46.0275 72.4122 46.2983C72.1506 46.5661 71.8275 46.7753 71.4428 46.9261C71.0581 47.0739 70.6256 47.1477 70.1455 47.1477ZM79.7227 41.4972L78.744 41.7741C78.6824 41.611 78.5916 41.4525 78.4716 41.2987C78.3546 41.1417 78.1946 41.0124 77.9915 40.9109C77.7884 40.8093 77.5283 40.7585 77.2113 40.7585C76.7773 40.7585 76.4157 40.8585 76.1264 41.0586C75.8402 41.2556 75.6971 41.5064 75.6971 41.8111C75.6971 42.0819 75.7956 42.2958 75.9925 42.4528C76.1895 42.6097 76.4973 42.7405 76.9158 42.8452L77.9684 43.1037C78.6024 43.2576 79.0748 43.493 79.3857 43.81C79.6965 44.1239 79.8519 44.5286 79.8519 45.0241C79.8519 45.4304 79.735 45.7936 79.5011 46.1136C79.2702 46.4337 78.9471 46.6861 78.5316 46.8707C78.1161 47.0554 77.6329 47.1477 77.082 47.1477C76.3588 47.1477 75.7602 46.9908 75.2862 46.6768C74.8123 46.3629 74.5122 45.9044 74.386 45.3011L75.4201 45.0426C75.5186 45.4242 75.7048 45.7105 75.9787 45.9013C76.2557 46.0921 76.6173 46.1875 77.0636 46.1875C77.5714 46.1875 77.9746 46.0798 78.2731 45.8643C78.5747 45.6458 78.7255 45.3842 78.7255 45.0795C78.7255 44.8333 78.6393 44.6271 78.467 44.4609C78.2946 44.2917 78.03 44.1655 77.6729 44.0824L76.4911 43.8054C75.8417 43.6515 75.3647 43.413 75.06 43.0898C74.7584 42.7636 74.6076 42.3558 74.6076 41.8665C74.6076 41.4664 74.7199 41.1125 74.9446 40.8047C75.1724 40.4969 75.4817 40.2553 75.8725 40.0799C76.2665 39.9045 76.7127 39.8168 77.2113 39.8168C77.913 39.8168 78.4639 39.9706 78.864 40.2784C79.2672 40.5862 79.5534 40.9924 79.7227 41.4972ZM85.0702 47V39.9091H86.1227V41.017H86.2151C86.3628 40.6385 86.6013 40.3446 86.9306 40.1353C87.2599 39.9229 87.6554 39.8168 88.117 39.8168C88.5849 39.8168 88.9742 39.9229 89.285 40.1353C89.5989 40.3446 89.8436 40.6385 90.019 41.017H90.0929C90.2745 40.6508 90.5469 40.36 90.91 40.1445C91.2732 39.926 91.7087 39.8168 92.2165 39.8168C92.8505 39.8168 93.3691 40.0153 93.7722 40.4123C94.1754 40.8062 94.377 41.4202 94.377 42.2543V47H93.2875V42.2543C93.2875 41.7311 93.1444 41.3571 92.8582 41.1325C92.5719 40.9078 92.2349 40.7955 91.8472 40.7955C91.3486 40.7955 90.9623 40.9463 90.6884 41.2479C90.4145 41.5464 90.2776 41.925 90.2776 42.3835V47H89.1696V42.1435C89.1696 41.7403 89.0388 41.4156 88.7772 41.1694C88.5156 40.9201 88.1786 40.7955 87.7662 40.7955C87.4831 40.7955 87.2184 40.8709 86.9722 41.0217C86.729 41.1725 86.5321 41.3817 86.3813 41.6495C86.2335 41.9142 86.1597 42.2204 86.1597 42.5682V47H85.0702ZM98.4166 47.1662C97.9673 47.1662 97.5595 47.0816 97.1933 46.9123C96.827 46.7399 96.5362 46.4922 96.3208 46.169C96.1053 45.8428 95.9976 45.4489 95.9976 44.9872C95.9976 44.581 96.0776 44.2517 96.2377 43.9993C96.3977 43.7438 96.6116 43.5438 96.8794 43.3991C97.1471 43.2545 97.4426 43.1468 97.7657 43.076C98.092 43.0021 98.4197 42.9437 98.749 42.9006C99.1799 42.8452 99.5292 42.8036 99.797 42.7759C100.068 42.7451 100.265 42.6944 100.388 42.6236C100.514 42.5528 100.577 42.4297 100.577 42.2543V42.2173C100.577 41.7618 100.453 41.4079 100.203 41.1555C99.957 40.9032 99.5831 40.777 99.0814 40.777C98.5613 40.777 98.1535 40.8909 97.8581 41.1186C97.5626 41.3464 97.3549 41.5895 97.2348 41.848L96.2007 41.4787C96.3854 41.0478 96.6316 40.7124 96.9394 40.4723C97.2502 40.2292 97.5888 40.0599 97.955 39.9645C98.3243 39.866 98.6875 39.8168 99.0445 39.8168C99.2722 39.8168 99.5338 39.8445 99.8293 39.8999C100.128 39.9522 100.416 40.0614 100.693 40.2276C100.973 40.3938 101.205 40.6446 101.39 40.9801C101.574 41.3156 101.667 41.7649 101.667 42.3281V47H100.577V46.0398H100.522C100.448 46.1937 100.325 46.3583 100.152 46.5337C99.9801 46.7092 99.7508 46.8584 99.4646 46.9815C99.1784 47.1046 98.8291 47.1662 98.4166 47.1662ZM98.5828 46.1875C99.0137 46.1875 99.3769 46.1029 99.6723 45.9336C99.9709 45.7643 100.196 45.5458 100.346 45.2781C100.5 45.0103 100.577 44.7287 100.577 44.4332V43.4361C100.531 43.4915 100.429 43.5423 100.272 43.5884C100.119 43.6315 99.9401 43.67 99.737 43.7038C99.5369 43.7346 99.3415 43.7623 99.1507 43.7869C98.9629 43.8085 98.8106 43.8269 98.6936 43.8423C98.4105 43.8793 98.1458 43.9393 97.8996 44.0224C97.6565 44.1024 97.4595 44.224 97.3087 44.3871C97.161 44.5471 97.0871 44.7656 97.0871 45.0426C97.0871 45.4212 97.2271 45.7074 97.5072 45.9013C97.7903 46.0921 98.1489 46.1875 98.5828 46.1875ZM104.064 49.6591C103.879 49.6591 103.715 49.6437 103.57 49.6129C103.425 49.5852 103.325 49.5575 103.27 49.5298L103.547 48.5696C103.812 48.6373 104.046 48.6619 104.249 48.6435C104.452 48.625 104.632 48.5342 104.789 48.3711C104.949 48.2111 105.095 47.951 105.227 47.5909L105.43 47.0369L102.808 39.9091H103.99L105.948 45.5597H106.021L107.979 39.9091H109.161L106.151 48.0341C106.015 48.4003 105.847 48.7035 105.647 48.9435C105.447 49.1867 105.215 49.3667 104.95 49.4837C104.689 49.6006 104.393 49.6591 104.064 49.6591ZM116.854 47.1477C116.264 47.1477 115.742 46.9985 115.289 46.6999C114.837 46.3983 114.483 45.9736 114.228 45.4258C113.972 44.8749 113.844 44.224 113.844 43.473C113.844 42.7282 113.972 42.0819 114.228 41.5341C114.483 40.9863 114.839 40.5631 115.294 40.2646C115.75 39.966 116.276 39.8168 116.873 39.8168C117.335 39.8168 117.699 39.8937 117.967 40.0476C118.238 40.1984 118.444 40.3707 118.586 40.5646C118.73 40.7554 118.843 40.9124 118.923 41.0355H119.015V37.5455H120.104V47H119.052V45.9105H118.923C118.843 46.0398 118.729 46.2029 118.581 46.3999C118.433 46.5937 118.222 46.7676 117.949 46.9215C117.675 47.0723 117.31 47.1477 116.854 47.1477ZM117.002 46.169C117.439 46.169 117.808 46.0552 118.11 45.8274C118.412 45.5966 118.641 45.2781 118.798 44.8718C118.955 44.4625 119.033 43.9901 119.033 43.4545C119.033 42.9252 118.956 42.462 118.803 42.065C118.649 41.6649 118.421 41.354 118.119 41.1325C117.818 40.9078 117.445 40.7955 117.002 40.7955C116.541 40.7955 116.156 40.9139 115.848 41.1509C115.543 41.3848 115.314 41.7034 115.16 42.1065C115.009 42.5066 114.934 42.956 114.934 43.4545C114.934 43.9593 115.011 44.4179 115.165 44.8303C115.322 45.2396 115.553 45.5658 115.857 45.8089C116.165 46.049 116.547 46.169 117.002 46.169ZM122.212 47V39.9091H123.302V47H122.212ZM122.766 38.7273C122.554 38.7273 122.371 38.6549 122.217 38.5103C122.066 38.3656 121.99 38.1918 121.99 37.9886C121.99 37.7855 122.066 37.6116 122.217 37.467C122.371 37.3223 122.554 37.25 122.766 37.25C122.978 37.25 123.16 37.3223 123.311 37.467C123.465 37.6116 123.542 37.7855 123.542 37.9886C123.542 38.1918 123.465 38.3656 123.311 38.5103C123.16 38.6549 122.978 38.7273 122.766 38.7273ZM128.416 39.9091V40.8324H124.593V39.9091H128.416ZM125.738 47V38.9304C125.738 38.5241 125.834 38.1856 126.024 37.9148C126.215 37.6439 126.463 37.4408 126.768 37.3054C127.072 37.17 127.394 37.1023 127.732 37.1023C128 37.1023 128.219 37.1238 128.388 37.1669C128.557 37.21 128.683 37.25 128.767 37.2869L128.453 38.2287C128.397 38.2102 128.32 38.1871 128.222 38.1594C128.126 38.1317 128 38.1179 127.843 38.1179C127.483 38.1179 127.223 38.2087 127.063 38.3903C126.906 38.5719 126.828 38.8381 126.828 39.1889V47H125.738ZM133.061 39.9091V40.8324H129.239V39.9091H133.061ZM130.384 47V38.9304C130.384 38.5241 130.479 38.1856 130.67 37.9148C130.861 37.6439 131.109 37.4408 131.413 37.3054C131.718 37.17 132.04 37.1023 132.378 37.1023C132.646 37.1023 132.864 37.1238 133.034 37.1669C133.203 37.21 133.329 37.25 133.412 37.2869L133.098 38.2287C133.043 38.2102 132.966 38.1871 132.867 38.1594C132.772 38.1317 132.646 38.1179 132.489 38.1179C132.129 38.1179 131.869 38.2087 131.709 38.3903C131.552 38.5719 131.473 38.8381 131.473 39.1889V47H130.384ZM137.23 47.1477C136.547 47.1477 135.958 46.9969 135.462 46.6953C134.97 46.3906 134.59 45.9659 134.322 45.4212C134.057 44.8733 133.925 44.2363 133.925 43.5099C133.925 42.7836 134.057 42.1435 134.322 41.5895C134.59 41.0324 134.962 40.5985 135.439 40.2876C135.919 39.9737 136.479 39.8168 137.119 39.8168C137.489 39.8168 137.853 39.8783 138.214 40.0014C138.574 40.1245 138.901 40.3246 139.197 40.6016C139.492 40.8755 139.728 41.2386 139.903 41.6911C140.079 42.1435 140.166 42.7005 140.166 43.3622V43.8239H134.7V42.8821H139.058C139.058 42.482 138.978 42.125 138.818 41.8111C138.661 41.4972 138.437 41.2494 138.144 41.0678C137.855 40.8862 137.513 40.7955 137.119 40.7955C136.685 40.7955 136.31 40.9032 135.993 41.1186C135.679 41.331 135.437 41.608 135.268 41.9496C135.099 42.2912 135.014 42.6574 135.014 43.0483V43.6761C135.014 44.2116 135.107 44.6656 135.291 45.038C135.479 45.4073 135.739 45.6889 136.071 45.8828C136.404 46.0736 136.79 46.169 137.23 46.169C137.516 46.169 137.775 46.129 138.006 46.049C138.24 45.9659 138.441 45.8428 138.611 45.6797C138.78 45.5135 138.911 45.3073 139.003 45.0611L140.055 45.3565C139.945 45.7135 139.759 46.0275 139.497 46.2983C139.235 46.5661 138.912 46.7753 138.527 46.9261C138.143 47.0739 137.71 47.1477 137.23 47.1477ZM141.785 47V39.9091H142.837V40.9801H142.911C143.04 40.6293 143.274 40.3446 143.613 40.1261C143.951 39.9076 144.333 39.7983 144.758 39.7983C144.838 39.7983 144.938 39.7998 145.058 39.8029C145.178 39.806 145.269 39.8106 145.33 39.8168V40.9247C145.293 40.9155 145.208 40.9016 145.076 40.8832C144.947 40.8616 144.81 40.8509 144.665 40.8509C144.321 40.8509 144.013 40.9232 143.742 41.0678C143.474 41.2094 143.262 41.4064 143.105 41.6587C142.951 41.908 142.874 42.1927 142.874 42.5128V47H141.785ZM150.2 47V39.9091H151.289V47H150.2ZM150.754 38.7273C150.541 38.7273 150.358 38.6549 150.204 38.5103C150.054 38.3656 149.978 38.1918 149.978 37.9886C149.978 37.7855 150.054 37.6116 150.204 37.467C150.358 37.3223 150.541 37.25 150.754 37.25C150.966 37.25 151.148 37.3223 151.298 37.467C151.452 37.6116 151.529 37.7855 151.529 37.9886C151.529 38.1918 151.452 38.3656 151.298 38.5103C151.148 38.6549 150.966 38.7273 150.754 38.7273ZM154.335 42.7344V47H153.246V39.9091H154.298V41.017H154.391C154.557 40.657 154.809 40.3677 155.148 40.1491C155.486 39.9276 155.923 39.8168 156.459 39.8168C156.939 39.8168 157.359 39.9152 157.719 40.1122C158.079 40.3061 158.359 40.6016 158.559 40.9986C158.759 41.3925 158.859 41.8911 158.859 42.4943V47H157.77V42.5682C157.77 42.0111 157.625 41.5772 157.336 41.2663C157.047 40.9524 156.65 40.7955 156.145 40.7955C155.797 40.7955 155.486 40.8709 155.212 41.0217C154.942 41.1725 154.728 41.3925 154.571 41.6818C154.414 41.9711 154.335 42.322 154.335 42.7344ZM167.512 39.9091V40.8324H163.838V39.9091H167.512ZM164.909 38.2102H165.998V44.9688C165.998 45.2765 166.043 45.5073 166.132 45.6612C166.224 45.812 166.341 45.9136 166.483 45.9659C166.627 46.0152 166.78 46.0398 166.94 46.0398C167.06 46.0398 167.158 46.0336 167.235 46.0213C167.312 46.0059 167.374 45.9936 167.42 45.9844L167.642 46.9631C167.568 46.9908 167.465 47.0185 167.332 47.0462C167.2 47.0769 167.032 47.0923 166.829 47.0923C166.521 47.0923 166.22 47.0262 165.924 46.8938C165.632 46.7615 165.389 46.5599 165.195 46.2891C165.004 46.0182 164.909 45.6766 164.909 45.2642V38.2102ZM170.354 42.7344V47H169.264V37.5455H170.354V41.017H170.446C170.612 40.6508 170.862 40.36 171.194 40.1445C171.53 39.926 171.976 39.8168 172.533 39.8168C173.016 39.8168 173.439 39.9137 173.802 40.1076C174.166 40.2984 174.447 40.5923 174.647 40.9893C174.85 41.3833 174.952 41.8849 174.952 42.4943V47H173.862V42.5682C173.862 42.005 173.716 41.5695 173.424 41.2617C173.135 40.9509 172.733 40.7955 172.219 40.7955C171.862 40.7955 171.542 40.8709 171.259 41.0217C170.979 41.1725 170.757 41.3925 170.594 41.6818C170.434 41.9711 170.354 42.322 170.354 42.7344ZM179.879 47.1477C179.196 47.1477 178.607 46.9969 178.111 46.6953C177.619 46.3906 177.239 45.9659 176.971 45.4212C176.706 44.8733 176.574 44.2363 176.574 43.5099C176.574 42.7836 176.706 42.1435 176.971 41.5895C177.239 41.0324 177.611 40.5985 178.088 40.2876C178.568 39.9737 179.128 39.8168 179.768 39.8168C180.138 39.8168 180.502 39.8783 180.862 40.0014C181.223 40.1245 181.55 40.3246 181.846 40.6016C182.141 40.8755 182.377 41.2386 182.552 41.6911C182.728 42.1435 182.815 42.7005 182.815 43.3622V43.8239H177.349V42.8821H181.707C181.707 42.482 181.627 42.125 181.467 41.8111C181.31 41.4972 181.086 41.2494 180.793 41.0678C180.504 40.8862 180.162 40.7955 179.768 40.7955C179.334 40.7955 178.959 40.9032 178.642 41.1186C178.328 41.331 178.086 41.608 177.917 41.9496C177.748 42.2912 177.663 42.6574 177.663 43.0483V43.6761C177.663 44.2116 177.756 44.6656 177.94 45.038C178.128 45.4073 178.388 45.6889 178.72 45.8828C179.053 46.0736 179.439 46.169 179.879 46.169C180.165 46.169 180.424 46.129 180.655 46.049C180.889 45.9659 181.09 45.8428 181.259 45.6797C181.429 45.5135 181.56 45.3073 181.652 45.0611L182.704 45.3565C182.594 45.7135 182.407 46.0275 182.146 46.2983C181.884 46.5661 181.561 46.7753 181.176 46.9261C180.792 47.0739 180.359 47.1477 179.879 47.1477ZM191.208 39.9091V40.8324H187.386V39.9091H191.208ZM188.531 47V38.9304C188.531 38.5241 188.626 38.1856 188.817 37.9148C189.008 37.6439 189.256 37.4408 189.56 37.3054C189.865 37.17 190.187 37.1023 190.525 37.1023C190.793 37.1023 191.011 37.1238 191.181 37.1669C191.35 37.21 191.476 37.25 191.559 37.2869L191.245 38.2287C191.19 38.2102 191.113 38.1871 191.015 38.1594C190.919 38.1317 190.793 38.1179 190.636 38.1179C190.276 38.1179 190.016 38.2087 189.856 38.3903C189.699 38.5719 189.62 38.8381 189.62 39.1889V47H188.531ZM197.165 44.1009V39.9091H198.255V47H197.165V45.7997H197.091C196.925 46.1598 196.667 46.466 196.316 46.7184C195.965 46.9677 195.522 47.0923 194.986 47.0923C194.543 47.0923 194.149 46.9954 193.804 46.8015C193.46 46.6045 193.189 46.3091 192.992 45.9151C192.795 45.5181 192.696 45.018 192.696 44.4148V39.9091H193.786V44.3409C193.786 44.858 193.931 45.2704 194.22 45.5781C194.512 45.8859 194.885 46.0398 195.337 46.0398C195.608 46.0398 195.883 45.9705 196.163 45.832C196.446 45.6935 196.683 45.4812 196.874 45.195C197.068 44.9087 197.165 44.544 197.165 44.1009ZM203.295 39.9091V40.8324H199.62V39.9091H203.295ZM200.691 38.2102H201.781V44.9688C201.781 45.2765 201.825 45.5073 201.915 45.6612C202.007 45.812 202.124 45.9136 202.265 45.9659C202.41 46.0152 202.562 46.0398 202.722 46.0398C202.842 46.0398 202.941 46.0336 203.018 46.0213C203.095 46.0059 203.156 45.9936 203.203 45.9844L203.424 46.9631C203.35 46.9908 203.247 47.0185 203.115 47.0462C202.982 47.0769 202.815 47.0923 202.612 47.0923C202.304 47.0923 202.002 47.0262 201.707 46.8938C201.414 46.7615 201.171 46.5599 200.977 46.2891C200.787 46.0182 200.691 45.6766 200.691 45.2642V38.2102ZM209.363 44.1009V39.9091H210.453V47H209.363V45.7997H209.29C209.123 46.1598 208.865 46.466 208.514 46.7184C208.163 46.9677 207.72 47.0923 207.184 47.0923C206.741 47.0923 206.347 46.9954 206.003 46.8015C205.658 46.6045 205.387 46.3091 205.19 45.9151C204.993 45.5181 204.895 45.018 204.895 44.4148V39.9091H205.984V44.3409C205.984 44.858 206.129 45.2704 206.418 45.5781C206.711 45.8859 207.083 46.0398 207.535 46.0398C207.806 46.0398 208.082 45.9705 208.362 45.832C208.645 45.6935 208.882 45.4812 209.073 45.195C209.267 44.9087 209.363 44.544 209.363 44.1009ZM212.409 47V39.9091H213.462V40.9801H213.536C213.665 40.6293 213.899 40.3446 214.238 40.1261C214.576 39.9076 214.958 39.7983 215.382 39.7983C215.462 39.7983 215.562 39.7998 215.683 39.8029C215.803 39.806 215.893 39.8106 215.955 39.8168V40.9247C215.918 40.9155 215.833 40.9016 215.701 40.8832C215.572 40.8616 215.435 40.8509 215.29 40.8509C214.945 40.8509 214.638 40.9232 214.367 41.0678C214.099 41.2094 213.887 41.4064 213.73 41.6587C213.576 41.908 213.499 42.1927 213.499 42.5128V47H212.409ZM219.965 47.1477C219.281 47.1477 218.692 46.9969 218.196 46.6953C217.704 46.3906 217.324 45.9659 217.056 45.4212C216.791 44.8733 216.659 44.2363 216.659 43.5099C216.659 42.7836 216.791 42.1435 217.056 41.5895C217.324 41.0324 217.696 40.5985 218.173 40.2876C218.653 39.9737 219.214 39.8168 219.854 39.8168C220.223 39.8168 220.588 39.8783 220.948 40.0014C221.308 40.1245 221.636 40.3246 221.931 40.6016C222.227 40.8755 222.462 41.2386 222.637 41.6911C222.813 42.1435 222.901 42.7005 222.901 43.3622V43.8239H217.435V42.8821H221.793C221.793 42.482 221.713 42.125 221.553 41.8111C221.396 41.4972 221.171 41.2494 220.879 41.0678C220.589 40.8862 220.248 40.7955 219.854 40.7955C219.42 40.7955 219.044 40.9032 218.727 41.1186C218.413 41.331 218.172 41.608 218.003 41.9496C217.833 42.2912 217.749 42.6574 217.749 43.0483V43.6761C217.749 44.2116 217.841 44.6656 218.026 45.038C218.213 45.4073 218.473 45.6889 218.806 45.8828C219.138 46.0736 219.524 46.169 219.965 46.169C220.251 46.169 220.509 46.129 220.74 46.049C220.974 45.9659 221.176 45.8428 221.345 45.6797C221.514 45.5135 221.645 45.3073 221.737 45.0611L222.79 45.3565C222.679 45.7135 222.493 46.0275 222.231 46.2983C221.97 46.5661 221.646 46.7753 221.262 46.9261C220.877 47.0739 220.445 47.1477 219.965 47.1477ZM225.313 47.0739C225.085 47.0739 224.89 46.9923 224.727 46.8292C224.564 46.6661 224.482 46.4706 224.482 46.2429C224.482 46.0152 224.564 45.8197 224.727 45.6566C224.89 45.4935 225.085 45.4119 225.313 45.4119C225.541 45.4119 225.736 45.4935 225.899 45.6566C226.062 45.8197 226.144 46.0152 226.144 46.2429C226.144 46.3937 226.105 46.5322 226.029 46.6584C225.955 46.7846 225.855 46.8861 225.728 46.9631C225.605 47.0369 225.467 47.0739 225.313 47.0739Z";
