--- frontend/src/views/WheresMyGene/components/GeneSearchBar/components/SourceDataButton/style.ts
@@ -1,28 +1,0 @@
-import styled from "@emotion/styled";
-import { ButtonIcon, fontBodyXxs, getColors } from "@czi-sds/components";
-
-export const StyledButtonDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 0 10px;
-`;
-
-export const StyledLabel = styled.div`
-  ${fontBodyXxs}
-
-  white-space: nowrap;
-
-  ${(props) => {
-    const colors = getColors(props);
-
-    return `
-      color: ${colors?.gray[500]}
-    `;
-  }}
-`;
-
-export const StyledButtonIcon = styled(ButtonIcon)`
-  width: 30px;
-  height: 30px;
-`;
