+++ frontend/src/views/WheresMyGene/components/Loader/index.tsx
@@ -1,6 +1,6 @@
 import { LoadingIndicator } from "@czi-sds/components";
-import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "../GeneSearchBar/components/SaveExport";
 import { Wrapper } from "./style";
+import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "src/views/WheresMyGeneV2/components/GeneSearchBar/components/SaveExport";
 
 export default function Loader(): JSX.Element {
   return (
