+++ frontend/src/views/WheresMyGene/components/ScreenTint/index.tsx
@@ -1,4 +1,4 @@
-import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "../GeneSearchBar/components/SaveExport";
+import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "src/views/WheresMyGeneV2/components/GeneSearchBar/components/SaveExport";
 import Loader from "../Loader";
 import { StyledDiv } from "./style";
