+++ frontend/src/views/WheresMyGeneV2/components/GeneSearchBar/components/ShareButton/index.tsx
@@ -5,10 +5,10 @@ import { EVENTS } from "src/common/analytics/events";
 import { EMPTY_OBJECT, noop } from "src/common/constants/utils";
 import { isSSR } from "src/common/utils/isSSR";
 
-import { StyledButtonIcon } from "../QuickSelect/style";
 import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "../SaveExport";
 import {
   StyledButtonDiv,
+  StyledButtonIcon,
   StyledIcon,
   StyledLabel,
   StyledNotificationDetails,
