+++ frontend/src/views/WheresMyGeneV2/components/GeneSearchBar/components/SourceDataButton/index.tsx
@@ -1,9 +1,8 @@
 import { MouseEventHandler } from "react";
 import { track } from "src/common/analytics";
 import { EVENTS } from "src/common/analytics/events";
-import { StyledButtonIcon } from "../QuickSelect/style";
 import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "../SaveExport";
-import { StyledButtonDiv, StyledLabel } from "./style";
+import { StyledButtonDiv, StyledButtonIcon, StyledLabel } from "./style";
 
 export default function SourceDataButton({
   handleRightSidebarButtonClick,
