+++ frontend/src/views/WheresMyGeneV2/components/GeneSearchBar/style.ts
@@ -1,9 +1,20 @@
 import styled from "@emotion/styled";
 import { fontBodyXxs } from "@czi-sds/components";
 import { gray500 } from "src/common/theme";
+import { autocompleteClasses, menuItemClasses, Popper } from "@mui/material";
+import { Button, ButtonIcon, MenuItem } from "@czi-sds/components";
+import { OFF_WHITE } from "src/common/theme";
+
+export const GENE_SEARCH_BAR_HEIGHT_PX = 32;
 
 export const Container = styled.div`
-  width: 80vw;
+  height: ${GENE_SEARCH_BAR_HEIGHT_PX}px;
+  width: fit-content;
+  margin-bottom: 8px;
+`;
+
+export const AutocompleteWrapper = styled.div`
+  width: 240px;
 `;
 
 export const ActionWrapper = styled.div`
@@ -20,4 +31,13 @@ export const Label = styled.label`
 export const LoadingIndicatorWrapper = styled.div`
   display: flex;
   align-items: center;
+`;
+
+export const StyledButtonWrapper = styled.div`
+  align-self: center;
+`;
+
+export const StyledClearButton = styled(Button)`
+  white-space: nowrap;
+  font-weight: 500;
 `;
