+++ frontend/ts_node_output.txt
@@ -1,0 +1,3 @@
+Need to install the following packages:
+  ts-node
+Ok to proceed? (y)
